% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateGPLlist.R
\name{addAnno}
\alias{addAnno}
\title{Add new annotation information of GPL to GPLlist}
\usage{
addAnno(GPL, probeID, symbolID)
}
\arguments{
\item{GPL}{A character, the number of GSE.}

\item{probeID}{A character, the header of probe from specific GPL platform.}

\item{symbolID}{A character, the header of gene symbol from specific GPL platform.}
}
\value{

}
\description{
Given GPL, probeID, and symbolID from GPL data, update the GPLlist data.
}
\details{
nothing
}
\examples{
\dontrun{
GPL="GPL10558"
gplDat = Table(getGEO(GPL,getGPL=TRUE));head(gplDat)
addAnno(GPL,probeID="ID",symbolID="ILMN_Gene")
}
}
