% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEtoExpr.R
\name{annoProbe}
\alias{annoProbe}
\title{Get annotation information of probes}
\usage{
annoProbe(GPL = "GPL9061", GPLdata = NA)
}
\arguments{
\item{GPL}{A character, the id number of GPL.}

\item{GPLdata}{A data.frame, the GPL annotation data from NCBI.}
}
\value{
A data.frame.
}
\description{
Given a GPL number or GPL data, return the data with first column of probes id and second column of gene symbol.
}
\details{
nothing
}
