% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometa_mapping.R
\name{as_ISOMetadata}
\alias{as_ISOMetadata}
\alias{as_ISOMetadata.emld}
\alias{as_ISOMetadata.ncdf4}
\title{Convert foreign metadata object to an \code{\link{ISOMetadata}} object}
\usage{
as_ISOMetadata(x, ...)

\method{as_ISOMetadata}{emld}(from)

\method{as_ISOMetadata}{ncdf4}(from)
}
\arguments{
\item{x}{object to convert}

\item{...}{further arguments}
}
\description{
Convert foreign metadata object to an \code{\link{ISOMetadata}} object
}
