% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOAbstractObject.R
\name{cacheISOClasses}
\alias{cacheISOClasses}
\title{cacheISOClasses}
\usage{
cacheISOClasses()
}
\description{
\code{\link{cacheISOClasses}} allows to cache the list of
\pkg{geometa} classes or extended. This is especially required to fasten
the decoding of metadata elements from an XML file. It is called internally
by \pkg{geometa} the first function \code{\link{getISOClasses()}} is called 
and each time the function \code{\link{readISO19139}} function is called to 
integrate eventually new classes added by user to extend \pkg{geometa} model 
(case of ISO profiles).
}
\examples{
            
  cacheISOClasses()

}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
