% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_codelists.R
\docType{data}
\name{codelists}
\alias{codelists}
\title{A named list of ISO codelists by standard (ISO 19139/ISO 19115-3)}
\format{
a named list by standard (19139 / 19115-3). Each list is then a named list
of all ISO Codelists for this standard.
}
\usage{
codelists
}
\description{
A named list of ISO codelists by standard (ISO 19139/ISO 19115-3)
}
\keyword{datasets}
