% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineExprs.R
\name{combine_exprs}
\alias{combine_exprs}
\title{Download multiple GSE expression matrix and convert simultaneously probe to gene symbol}
\usage{
combine_exprs(destdir = "tmp", GSEs = NULL, is_duplicated = TRUE)
}
\arguments{
\item{destdir}{A character, the path to save expression data, where file name of expression data has pattern ^GSE.*-GPL.*-matrix.txt$ .}

\item{GSEs}{A vector of GSE character. default to combine all expression data in destdir filefolder.}

\item{is_duplicated}{A logical. default to duplicated the same features}
}
\value{
Save the files of expression matrix and phenotype information.
}
\description{
Given a vector of GSE number, download related GSE data and convert probe to gene symbol, and save the gene matrix with gene symbol.
}
\details{
combine_exprs would serach expression matrix file with pattern of ^GSE.*-GPL.*-matrix.txt$ based on the GSE ids in phe argument.
}
