% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineExprs.R
\name{combineExprs}
\alias{combineExprs}
\title{Download multiple GSE expression matrix and convert simultaneously probe to gene symbol}
\usage{
combineExprs(destdir = "tmp", GSEs = FALSE)
}
\arguments{
\item{destdir}{A character, the path to save expression data, where file name of expression data has pattern ^GSE.*-GPL.*-matrix.txt$ .}

\item{GSEs}{Logical or a vector of character.}
}
\value{
Save the files of expression matrix and phenotype information.
}
\description{
Given a vector of GSE number, download related GSE data and convert probe to gene symbol, and save the gene matrix with gene symbol.
}
\details{
combineExprs would serach expression matrix file with pattern of ^GSE.*-GPL.*-matrix.txt$ based on the GSE ids in phe argument.
}
