% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEtoExpr.R
\name{convert_probe_se}
\alias{convert_probe_se}
\title{Convert ExpressionSet with probe to SummarizedExperiment with gene symbol}
\usage{
convert_probe_se(eSet)
}
\arguments{
\item{eSet}{A ExpressionSet class}
}
\value{
A SummarizedExperiment class. The pattern of filename are GSE-GPL-phe.txt for phenotype and GSE-GPL-matrix.txt for genotype splilted by tab.
}
\description{
After download GSE data using getGEO, we use this function to convert an ExpressionSet with probe to SummarizedExperiment with gene symbol
}
