% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geometa.R
\docType{package}
\name{geometa}
\alias{geometa}
\alias{geometa-package}
\title{Facilities for handling geographic metadata}
\description{
Facilities to handle OGC/ISO geographic standard metadata
}
\details{
\tabular{ll}{
   Package: \tab geometa\cr
   Type: \tab Package\cr
   Version: \tab 0.1-0\cr
   Date: \tab 2017-03-15\cr
   License: \tab MIT\cr
   LazyLoad: \tab yes\cr
 }
}
\author{
Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}
}

