% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometa_coverage.R
\name{geometa_coverage}
\alias{geometa_coverage}
\alias{geometa_coverae}
\title{geometa_coverage}
\usage{
geometa_coverage()
}
\value{
an object of class \code{data.frame}
}
\description{
\code{geometa_coverage} is a function to report coverage of ISO/OGC standard classes
in package \pkg{geometa}. The function will inspect all classes of the ISO/OGC standards and will
scan if \pkg{geometa} supports it.
}
\note{
This function is used as Quality Assurance indicator to assess the percentage of completeness
of ISO/OGC standards in \pkg{geometa}.
}
\examples{
\donttest{
  cov <- geometa_coverage()
}

}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
