% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gene_expr_GSEs.R
\name{get_gene_expr_GSEs}
\alias{get_gene_expr_GSEs}
\title{Regional Plot Association Results of Pig}
\usage{
get_gene_expr_GSEs(geneName, GSE_GPL)
}
\arguments{
\item{geneName}{a character of gene SYMBOL name}

\item{GSE_GPL}{a dataframe of GSE and GPL number}
}
\value{
download GSE expression matrix and GPL annotation,
        and return a matrix with colanmes of "GSE" and "GPL" and row names of GSE_GPL

a dataframe of sepecific gene
}
\description{
This is a plotting tool to simultaneously display the association p-value,
             the LD, recombination rate and annotated gene in region of interesting in pig.
             The interpretation and visualization tool of association results will
             contribute to further identification of causal mutations.
}
\details{
nothing
}
