% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOAbstractObject.R
\name{getClassesInheriting}
\alias{getClassesInheriting}
\title{getClassesInheriting}
\usage{
getClassesInheriting(classname, extended, pretty)
}
\arguments{
\item{classname}{the name of the superclass for which inheriting sub-classes have to be listed}

\item{extended}{whether we want to look at user namespace for third-party sub-classes}

\item{pretty}{prettify the output as \code{data.frame}}
}
\description{
get the list of classes inheriting a given super class provided by its name
}
\examples{
  getClassesInheriting("ISAbstractObject")
}
