% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExpr.R
\name{getExpr}
\alias{getExpr}
\title{Get expression matrix without annotation}
\usage{
getExpr(GSE, destdir = "tmp")
}
\arguments{
\item{GSE}{A character, the number of GSE.}

\item{destdir}{A character, the path to download GSE related files.}
}
\value{
a character of probe
}
\description{
nothing
}
\details{
nothing
}
