% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometa_options.R
\name{getGeometaOption}
\alias{getGeometaOption}
\title{getGeometaOption}
\usage{
getGeometaOption(option)
}
\arguments{
\item{option}{the name of the option}
}
\value{
the option
}
\description{
\code{getGeometaOption} allows to get an option from \pkg{geometa}
}
\examples{
            
  getGeometaOption("schemaBaseUrl")

}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
