% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometa_options.R
\name{getGeometaOptions}
\alias{getGeometaOptions}
\title{getGeometaOptions}
\usage{
getGeometaOptions()
}
\description{
\code{getGeometaOptions} allows to get options from \pkg{geometa}
}
\examples{
            
  getGeometaOptions()

}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
