% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOCodelist.R
\name{getISOCodelist}
\alias{getISOCodelist}
\title{getISOCodelist}
\usage{
getISOCodelist(id)
}
\arguments{
\item{id}{identifier of the codelist}
}
\description{
\code{getISOCodelist} allows to get a registered ISO codelist by id
registered in \pkg{geometa}
}
\examples{
            
  getISOCodelist(id = "LanguageCode")

}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
