% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOCodelist.R
\name{getISOInternalCodelists}
\alias{getISOInternalCodelists}
\title{getISOInternalCodelists}
\usage{
getISOInternalCodelists()
}
\description{
\code{getISOInternalCodelists} allows to get the list of ISO codelists
registered in \pkg{geometa}
}
\examples{
            
  getISOInternalCodelists()

}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
