% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOMetadataNamespace.R
\name{getISOMetadataNamespace}
\alias{getISOMetadataNamespace}
\title{getISOMetadataNamespace}
\usage{
getISOMetadataNamespace(id)
}
\arguments{
\item{id}{namespace prefix}
}
\description{
\code{getISOMetadataNamespace} gets a namespace given its id
}
\examples{
            
  getISOMetadataNamespace("GMD")

}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
