% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOMetadataNamespace.R
\name{getISOMetadataNamespaces}
\alias{getISOMetadataNamespaces}
\title{getISOMetadataNamespaces}
\usage{
getISOMetadataNamespaces()
}
\description{
\code{getISOMetadataNamespaces} gets the list of namespaces registered
}
\examples{
            
  getISOMetadataNamespaces()

}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
