% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOMetadataSchemas.R
\name{getISOMetadataSchemaFile}
\alias{getISOMetadataSchemaFile}
\title{getISOMetadataSchemaFile}
\usage{
getISOMetadataSchemaFile(version)
}
\arguments{
\item{version}{the schema version}
}
\value{
the internal path to the schema file
}
\description{
\code{getISOMetadataSchemaFile} allows to get the schema file
in \pkg{geometa}
}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
