% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOMetadataSchemas.R
\name{getISOMetadataSchemas}
\alias{getISOMetadataSchemas}
\title{getISOMetadataSchemas}
\usage{
getISOMetadataSchemas()
}
\description{
\code{getISOMetadataSchemas} gets the schemas registered in \pkg{geometa}
}
\examples{
            
  getISOMetadataSchemas()

}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
