% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometa_mapping.R
\name{getMappingFormats}
\alias{getMappingFormats}
\title{getMappingFormats}
\usage{
getMappingFormats(pretty)
}
\arguments{
\item{pretty}{by default \code{TRUE} to return the list of formats as \code{data.frame}. Set
to \code{FALSE} to return a list of \code{pivot_format} objects}
}
\description{
\code{getMappingFormats} gets the mapping formats registered in \pkg{geometa}
}
\examples{
            
  getMappingFormats()

}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
