% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometa_mapping.R
\name{getMappings}
\alias{getMappings}
\title{getMappings}
\usage{
getMappings()
}
\value{
a \code{data.frame} containing the metadata mapping rules
}
\description{
List the mappings rules to convert from/to other metadata formats (currently 
EML/emld objects and NetCDF-CF/ncdf4 objects)
}
