% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOMetadataSchemas.R
\name{getMetadataStandard}
\alias{getMetadataStandard}
\title{getMetadataStandard}
\usage{
getMetadataStandard()
}
\description{
\code{getMetadataStandard} allows to set the standard to use for encoding/decoding in \pkg{geometa}.
}
\examples{
            
  getMetadataStandard()

}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
