% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExpr.R
\name{getPhe}
\alias{getPhe}
\title{Get phenotype data}
\usage{
getPhe(GSE, destdir = "tmp")
}
\arguments{
\item{GSE}{A character, the number of GSE.}

\item{destdir}{A character, the path to download GSE related files.}
}
\value{
a character of probe
}
\description{
nothing
}
\details{
nothing
}
