% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMLAbstractCoverage.R
\docType{class}
\name{GMLAbstractCoverage}
\alias{GMLAbstractCoverage}
\title{GMLAbstractCoverage}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an GML abstract coverage
}
\description{
GMLAbstractCoverage
}
\note{
Class used internally by geometa
}
\section{Methods}{

\describe{
 \item{\code{new(xml, element, attrs, defaults)}}{
   This method is used to instantiate a GML abstract coverage
 }
 \item{\code{setDomainSet(domainSet)}}{
   Set a domain set. The object should be a GML Geometric or Time object
 }
 \item{\code{setRangeSet(rangeSet)}}{
   Set a range set. Currently not implemented in \pkg{geometa}
 }
}
}

\references{
ISO 19136:2007 Geographic Information -- Geographic Markup Language.
  http://www.iso.org/iso/iso_catalogue/catalogue_tc/catalogue_detail.htm?csnumber=32554 
  OGC Geography Markup Language. http://www.opengeospatial.org/standards/gml
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{GML}
\keyword{ISO}
\keyword{abstract}
\keyword{coverage}
