% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GMLAbstractCRS.R
\docType{class}
\name{GMLAbstractCRS}
\alias{GMLAbstractCRS}
\title{GMLAbstractCRS}
\format{\code{\link{R6Class}} object.}
\usage{
GMLAbstractCRS
}
\value{
Object of \code{\link{R6Class}} for modelling an GMLAbstractCRS
}
\description{
GMLAbstractCRS
}
\section{Fields}{

\describe{
\item{\code{scope}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(xml, defaults, id)}}{
   This method is used to instantiate a GML Abstract CRS
 }
 \item{\code{addScope(scope)}}{
   Adds a scope
 }
 \item{\code{delScope(scope)}}{
   Deletes a scope
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\references{
ISO 19136:2007 Geographic Information -- Geographic Markup Language.
  http://www.iso.org/iso/iso_catalogue/catalogue_tc/catalogue_detail.htm?csnumber=32554

  OGC Geography Markup Language. http://www.opengeospatial.org/standards/gml
}
\keyword{CRS}
\keyword{GML}
\keyword{ISO}
\keyword{abstract}

