% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMLAbstractCurve.R
\docType{class}
\name{GMLAbstractCurve}
\alias{GMLAbstractCurve}
\title{GMLAbstractCurve}
\format{\code{\link{R6Class}} object.}
\usage{
GMLAbstractCurve
}
\value{
Object of \code{\link{R6Class}} for modelling an GML abstract curve
}
\description{
GMLAbstractCurve
}
\note{
Experimental
}
\section{Methods}{

\describe{
 \item{\code{new(xml, sfg)}}{
   This method is used to instantiate a GML abstract curve
 }
}
}

\references{
ISO 19136:2007 Geographic Information -- Geographic Markup Language.
  http://www.iso.org/iso/iso_catalogue/catalogue_tc/catalogue_detail.htm?csnumber=32554 
  
  OGC Geography Markup Language. http://www.opengeospatial.org/standards/gml
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{Curve}
\keyword{GML}
\keyword{ISO}
