% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMLAbstractGeneralConversion.R
\docType{class}
\name{GMLAbstractGeneralConversion}
\alias{GMLAbstractGeneralConversion}
\title{GMLAbstractGeneralConversion}
\format{\code{\link{R6Class}} object.}
\usage{
GMLAbstractGeneralConversion
}
\value{
Object of \code{\link{R6Class}} for modelling an GMLAbstractGeneralConversion
}
\description{
GMLAbstractGeneralConversion
}
\section{Inherited methods}{

\describe{
  from \code{GMLAbstractCoordinateOperation}
}
}

\section{Methods}{

\describe{
 \item{\code{new(xml, defaults, id)}}{
   This method is used to instantiate a GML Abstract CRS
 }
}
}

\references{
ISO 19136:2007 Geographic Information -- Geographic Markup Language.
  http://www.iso.org/iso/iso_catalogue/catalogue_tc/catalogue_detail.htm?csnumber=32554 
  
  OGC Geography Markup Language. http://www.opengeospatial.org/standards/gml
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{GML}
\keyword{ISO}
\keyword{abstract}
\keyword{conversion}
\keyword{general}
