% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GMLAbstractGeneralDerivedCRS.R
\docType{class}
\name{GMLAbstractGeneralDerivedCRS}
\alias{GMLAbstractGeneralDerivedCRS}
\title{GMLAbstractGeneralDerivedCRS}
\format{\code{\link{R6Class}} object.}
\usage{
GMLAbstractGeneralDerivedCRS
}
\value{
Object of \code{\link{R6Class}} for modelling an GMLAbstractGeneralDerivedCRS
}
\description{
GMLAbstractGeneralDerivedCRS
}
\section{Fields}{

\describe{
\item{\code{conversion}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(xml, defaults, id)}}{
   This method is used to instantiate a GML abstract general derived CRS
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\references{
ISO 19136:2007 Geographic Information -- Geographic Markup Language.
  http://www.iso.org/iso/iso_catalogue/catalogue_tc/catalogue_detail.htm?csnumber=32554

  OGC Geography Markup Language. http://www.opengeospatial.org/standards/gml
}
\keyword{GML}
\keyword{ISO}
\keyword{abstract}
\keyword{crs}
\keyword{single}

