% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GMLAbstractGeneralOperationParameter.R
\docType{class}
\name{GMLAbstractGeneralOperationParameter}
\alias{GMLAbstractGeneralOperationParameter}
\title{GMLAbstractGeneralOperationParameter}
\format{\code{\link{R6Class}} object.}
\usage{
GMLAbstractGeneralOperationParameter
}
\value{
Object of \code{\link{R6Class}} for modelling an GMLAbstractGeneralOperationParameter
}
\description{
GMLAbstractGeneralOperationParameter
}
\section{Fields}{

\describe{
\item{\code{minimumOccurs}}{}
}}
\section{Inherited methods}{

\describe{
  from \code{GMLDefinition}
}
}

\section{Methods}{

\describe{
 \item{\code{new(xml, defaults, id)}}{
   This method is used to instantiate a GML AbstractGeneralOperationParameter
 }
 \item{\code{setMinimumOccurs(minimumOccurs)}}{
   Sets the minimum occurs, object of class \code{integer}
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\references{
ISO 19136:2007 Geographic Information -- Geographic Markup Language.
  http://www.iso.org/iso/iso_catalogue/catalogue_tc/catalogue_detail.htm?csnumber=32554

  OGC Geography Markup Language. http://www.opengeospatial.org/standards/gml
}
\keyword{GML}
\keyword{ISO}
\keyword{abstract}
\keyword{general}
\keyword{operation}
\keyword{parameter}

