% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMLAbstractGML.R
\docType{class}
\name{GMLAbstractGML}
\alias{GMLAbstractGML}
\title{GMLAbstractGML}
\format{\code{\link{R6Class}} object.}
\usage{
GMLAbstractGML
}
\value{
Object of \code{\link{R6Class}} for modelling an GML abstract GML
}
\description{
GMLAbstractGML
}
\section{Fields}{

\describe{
\item{\code{name}}{}
}}

\note{
Class used internally by geometa
}
\section{Methods}{

\describe{
 \item{\code{new(xml, element, attrs, defaults)}}{
   This method is used to instantiate a GML abstract GML
 }
 \item{\code{setDescriptionReference(descriptionReference)}}{
   Set the descriptionReference
 }
 \item{\code{setIdentifier(identifier)}}{
   Set the identifier
 }
 \item{\code{addName(name)}}{
   Adds a name
 }
 \item{\code{delName(name)}}{
   Deletes a name
 }
}
}

\references{
ISO 19136:2007 Geographic Information -- Geographic Markup Language.
  http://www.iso.org/iso/iso_catalogue/catalogue_tc/catalogue_detail.htm?csnumber=32554 
  
  OGC Geography Markup Language. http://www.opengeospatial.org/standards/gml
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{GML}
\keyword{ISO}
