% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GMLAbstractObject.R
\docType{class}
\name{GMLAbstractObject}
\alias{GMLAbstractObject}
\title{GMLAbstractObject}
\format{\code{\link{R6Class}} object.}
\usage{
GMLAbstractObject
}
\value{
Object of \code{\link{R6Class}} for modelling an GML abstract object
}
\description{
GMLAbstractObject
}
\note{
Class used internally by geometa
}
\section{Methods}{

\describe{
 \item{\code{new(xml, element, attrs, defaults)}}{
   This method is used to instantiate an ISOTemporalPrimitive
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\references{
ISO 19136:2007 Geographic Information -- Geographic Markup Language.
  http://www.iso.org/iso/iso_catalogue/catalogue_tc/catalogue_detail.htm?csnumber=32554

  OGC Geography Markup Language. http://www.opengeospatial.org/standards/gml
}
\keyword{GML}
\keyword{ISO}

