% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMLAbstractRing.R
\docType{class}
\name{GMLAbstractRing}
\alias{GMLAbstractRing}
\title{GMLAbstractRing}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an GML abstract ring
}
\description{
GMLAbstractRing
}
\note{
Class used internally by geometa
}
\section{Methods}{

\describe{
 \item{\code{new(xml, element, attrs, defaults)}}{
   This method is used to instantiate a GML abstract ring
 }
}
}

\references{
ISO 19136:2007 Geographic Information -- Geographic Markup Language.
  http://www.iso.org/iso/iso_catalogue/catalogue_tc/catalogue_detail.htm?csnumber=32554 
  
  OGC Geography Markup Language. http://www.opengeospatial.org/standards/gml
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{GML}
\keyword{Geometric}
\keyword{ISO}
\keyword{ring}
