% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMLAbstractSingleCRS.R
\docType{class}
\name{GMLAbstractSingleCRS}
\alias{GMLAbstractSingleCRS}
\title{GMLAbstractSingleCRS}
\format{\code{\link{R6Class}} object.}
\usage{
GMLAbstractSingleCRS
}
\value{
Object of \code{\link{R6Class}} for modelling an GMLAbstractSingleCRS
}
\description{
GMLAbstractSingleCRS
}
\section{Methods}{

\describe{
 \item{\code{new(xml, defaults, id)}}{
   This method is used to instantiate a GML Abstract single CRS
 }
}
}

\references{
ISO 19136:2007 Geographic Information -- Geographic Markup Language.
  http://www.iso.org/iso/iso_catalogue/catalogue_tc/catalogue_detail.htm?csnumber=32554 
  
  OGC Geography Markup Language. http://www.opengeospatial.org/standards/gml
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{GML}
\keyword{ISO}
\keyword{abstract}
\keyword{crs}
\keyword{single}
