% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMLAbstractSurface.R
\docType{class}
\name{GMLAbstractSurface}
\alias{GMLAbstractSurface}
\title{GMLAbstractSurface}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an GML abstract surface
}
\description{
GMLAbstractSurface
}
\note{
Experimental
}
\section{Methods}{

\describe{
 \item{\code{new(xml, sfg)}}{
   This method is used to instantiate a GML abstract surface
 }
}
}

\references{
ISO 19136:2007 Geographic Information -- Geographic Markup Language.
  http://www.iso.org/iso/iso_catalogue/catalogue_tc/catalogue_detail.htm?csnumber=32554 
  
  OGC Geography Markup Language. http://www.opengeospatial.org/standards/gml
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{GML}
\keyword{ISO}
\keyword{Surface}
