% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMLAbstractTimeGeometricPrimitive.R
\docType{class}
\name{GMLAbstractTimeGeometricPrimitive}
\alias{GMLAbstractTimeGeometricPrimitive}
\title{GMLAbstractTimeGeometricPrimitive}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO GML abstract temporal primitive
}
\description{
GMLAbstractTimeGeometricPrimitive
}
\note{
Class used internally by geometa
}
\section{Methods}{

\describe{
 \item{\code{new(xml, element, namespace, defaults)}}{
   This method is used to instantiate an ISOAbstractTimeGeometricPrimitive
 }
}
}

\references{
ISO 19115:2003 - Geographic information -- Metadata
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{primitive}
\keyword{temporal}
