% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMLAbstractTimePrimitive.R
\docType{class}
\name{GMLAbstractTimePrimitive}
\alias{GMLAbstractTimePrimitive}
\title{GMLAbstractTimePrimitive}
\format{\code{\link{R6Class}} object.}
\usage{
GMLAbstractTimePrimitive
}
\value{
Object of \code{\link{R6Class}} for modelling an GML AbstractTimePrimitive
}
\description{
GMLAbstractTimePrimitive
}
\section{Methods}{

\describe{
 \item{\code{new(xml, defaults)}}{
   This method is used to instantiate a GML AbstractTimePrimitive
 }
 \item{\code{setId(id)}}{
   Sets the id
 }
 \item{\code{addRelatedTime(time)}}{
   Adds related time, object of class among \code{GMLTimeInstant}, \code{GMLTimePeriod},
   \code{GMLTimeNode} or \code{GMLTimeEdge}
 }
 \item{\code{delRelatedTime(time)}}{
   Deletes related time
 }
}
}

\references{
ISO 19136:2007 Geographic Information -- Geographic Markup Language.
  http://www.iso.org/iso/iso_catalogue/catalogue_tc/catalogue_detail.htm?csnumber=32554 
  
  OGC Geography Markup Language. http://www.opengeospatial.org/standards/gml
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{GML}
\keyword{ISO}
\keyword{abstract}
\keyword{primitive}
\keyword{time}
