% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMLAffineCS.R
\docType{class}
\name{GMLAffineCS}
\alias{GMLAffineCS}
\title{GMLAffineCS}
\format{\code{\link{R6Class}} object.}
\usage{
GMLAffineCS
}
\value{
Object of \code{\link{R6Class}} for modelling an GMLAffineCS
}
\description{
GMLAffineCS
}
\section{Inherited Methods}{

\describe{
 \item{\code{new(xml, defaults, id)}}{
   This method is used to instantiate a GML Abstract CRS
 }
 \item{\code{addAxis(axis)}}{
   Adds an axis, object of class \code{GMLCoordinateSystemAxis}
 }
 \item{\code{delAxis(axis)}}{
   Deletes an axis, object of class \code{GMLCoordinateSystemAxis}
 }
}
}

\references{
ISO 19136:2007 Geographic Information -- Geographic Markup Language.
  http://www.iso.org/iso/iso_catalogue/catalogue_tc/catalogue_detail.htm?csnumber=32554 
  
  OGC Geography Markup Language. http://www.opengeospatial.org/standards/gml
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{GML}
\keyword{ISO}
\keyword{affine}
\keyword{coordinate}
\keyword{system}
