% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMLCodeType.R
\docType{class}
\name{GMLCodeType}
\alias{GMLCodeType}
\title{GMLCodeType}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling a GML code type
}
\description{
GMLCodeType
}
\references{
ISO 19136:2007 Geographic Information -- Geographic Markup Language.
  http://www.iso.org/iso/iso_catalogue/catalogue_tc/catalogue_detail.htm?csnumber=32554 
  
  OGC Geography Markup Language. http://www.opengeospatial.org/standards/gml
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{GML}
\keyword{codetype}
