% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMLCompoundCRS.R
\docType{class}
\name{GMLCompoundCRS}
\alias{GMLCompoundCRS}
\title{GMLCompoundCRS}
\format{\code{\link{R6Class}} object.}
\usage{
GMLCompoundCRS
}
\value{
Object of \code{\link{R6Class}} for modelling an GMLCompoundCRS
}
\description{
GMLCompoundCRS
}
\section{Fields}{

\describe{
\item{\code{componentReferenceSystem}}{}
}}

\section{Methods}{

\describe{
 \item{\code{new(xml, defaults, id)}}{
   This method is used to instantiate a GML Abstract CRS
 }
 \item{\code{addComponentReferenceSystem(referenceSystem)}}{
   Adds a reference system
 }
 \item{\code{delComponentReferenceSystem(referenceSystem)}}{
   Deletes a reference system
 }
}
}

\references{
ISO 19136:2007 Geographic Information -- Geographic Markup Language.
  http://www.iso.org/iso/iso_catalogue/catalogue_tc/catalogue_detail.htm?csnumber=32554 
  
  OGC Geography Markup Language. http://www.opengeospatial.org/standards/gml
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{CRS}
\keyword{GML}
\keyword{ISO}
\keyword{compound}
