% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMLCOVAbstractCoverage.R
\docType{class}
\name{GMLCOVAbstractCoverage}
\alias{GMLCOVAbstractCoverage}
\title{GMLCOVAbstractCoverage}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling a GMLCOV Abstract Coverage
}
\description{
GMLCOVAbstractCoverage
}
\section{Methods}{

\describe{
 \item{\code{new(xml, element, attrs, defaults, wrap)}}{
   This method is used to create an GMLCOV Abstract Coverage
 }
}
}

\references{
GML 3.2.1 Application Schema for Coverages http://www.opengis.net/gmlcov/1.0
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{Coverage}
\keyword{GML}
\keyword{GMLCOV}
