% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GMLDefinition.R
\docType{class}
\name{GMLDefinition}
\alias{GMLDefinition}
\title{GMLDefinition}
\format{\code{\link{R6Class}} object.}
\usage{
GMLDefinition
}
\value{
Object of \code{\link{R6Class}} for modelling an GML definition
}
\description{
GMLDefinition
}
\section{Fields}{

\describe{
\item{\code{remarks}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(xml, defaults)}}{
   This method is used to instantiate a GML Definition
 }
 \item{\code{setId(id)}}{
   Sets the id
 }
 \item{\code{addRemark(remark)}}{
   Adds a remark
 }
 \item{\code{delRemark(remark)}}{
   Deletes a remark
 }
}
}
\examples{
gml <- GMLDefinition$new()
  gml$setDescriptionReference("someref")
  gml$setIdentifier("identifier", "codespace")
  gml$addName("name1", "codespace")
  gml$addName("name2", "codespace")
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\references{
ISO 19136:2007 Geographic Information -- Geographic Markup Language.
  http://www.iso.org/iso/iso_catalogue/catalogue_tc/catalogue_detail.htm?csnumber=32554

  OGC Geography Markup Language. http://www.opengeospatial.org/standards/gml
}
\keyword{GML}
\keyword{ISO}
\keyword{definition}

