% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GMLDerivedCRS.R
\docType{class}
\name{GMLDerivedCRS}
\alias{GMLDerivedCRS}
\title{GMLDerivedCRS}
\format{\code{\link{R6Class}} object.}
\usage{
GMLDerivedCRS
}
\value{
Object of \code{\link{R6Class}} for modelling an GMLDerivedCRS
}
\description{
GMLDerivedCRS
}
\section{Fields}{

\describe{
\item{\code{baseCRS}}{}

\item{\code{derivedCRSType}}{}

\item{\code{coordinateSystem}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(xml, defaults, id)}}{
   This method is used to instantiate a GML derived CRS
 }
 \item{\code{setBaseCRS(crs)}}{
   Sets the base CRS, one object of class inherited from \code{GMLAbstractSingleCRS}
 }
 \item{\code{setDerivedCRSType(type, codeSpace)}}{
   Sets a derived CRS type
 }
 \item{\code{setCoordinateSystem(cs)}}{
   Sets the coordinate system
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\references{
ISO 19136:2007 Geographic Information -- Geographic Markup Language.
  http://www.iso.org/iso/iso_catalogue/catalogue_tc/catalogue_detail.htm?csnumber=32554

  OGC Geography Markup Language. http://www.opengeospatial.org/standards/gml
}
\keyword{GML}
\keyword{ISO}
\keyword{crs}
\keyword{derived}

