% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMLElement.R
\docType{class}
\name{GMLElement}
\alias{GMLElement}
\title{GMLElement}
\format{\code{\link{R6Class}} object.}
\usage{
GMLElement
}
\value{
Object of \code{\link{R6Class}} for modelling an GML element
}
\description{
GMLElement
}
\note{
Class used by geometa internal XML decoder/encoder
}
\section{Methods}{

\describe{
 \item{\code{new(xml, element, attrs, defaults)}}{
   This method is used to instantiate a GML element
 }
}
}

\references{
ISO/TS 19103:2005 Geographic information -- Conceptual schema language
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{GML}
\keyword{ISO}
\keyword{element}
