% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMLEnvelope.R
\docType{class}
\name{GMLEnvelope}
\alias{GMLEnvelope}
\title{GMLEnvelope}
\format{\code{\link{R6Class}} object.}
\usage{
GMLEnvelope
}
\value{
Object of \code{\link{R6Class}} for modelling an GML envelope
}
\description{
GMLEnvelope
}
\section{Fields}{

\describe{
\item{\code{posList}}{}
}}

\note{
Experimental
}
\section{Methods}{

\describe{
 \item{\code{new(xml, bbox)}}{
   This method is used to instantiate a GML envelope. The argument 'bbox'
   should be a matrix of dim 2,2 giving the x/y min/max values of a bouding box,
   as returned by \code{bbox} function in package \pkg{sp}
 }
}
}

\references{
ISO 19136:2007 Geographic Information -- Geographic Markup Language.
  http://www.iso.org/iso/iso_catalogue/catalogue_tc/catalogue_detail.htm?csnumber=32554 
  
  OGC Geography Markup Language. http://www.opengeospatial.org/standards/gml
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{Envelope}
\keyword{GML}
\keyword{ISO}
