% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMLGrid.R
\docType{class}
\name{GMLGrid}
\alias{GMLGrid}
\title{GMLGrid}
\format{\code{\link{R6Class}} object.}
\usage{
GMLGrid
}
\value{
Object of \code{\link{R6Class}} for modelling an GML grid
}
\description{
GMLGrid
}
\note{
Class used internally by geometa
}
\section{Methods}{

\describe{
 \item{\code{new(xml, element, attrs, defaults)}}{
   This method is used to instantiate a GML grid
 }
 \item{\code{setGridEnvelope(m)}}{
   Set the grid envelope limits as object.
 }
 \item{\code{setAxislabels(xlabel,ylabel)}}{
   Set the Axis labels
 }
 \item{\code{addAxisName(axisName)}}{
   Adds an axis name
 }
 \item{\code{delAxisName(axisName)}}{
   Deletes an axis name
 }
}
}

\references{
ISO 19136:2007 Geographic Information -- Geographic Markup Language.
  http://www.iso.org/iso/iso_catalogue/catalogue_tc/catalogue_detail.htm?csnumber=32554 
  
  OGC Geography Markup Language. http://www.opengeospatial.org/standards/gml
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{GML}
\keyword{Grid}
\keyword{ISO}
