% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMLGridFunction.R
\docType{class}
\name{GMLGridFunction}
\alias{GMLGridFunction}
\title{GMLGridFunction}
\format{\code{\link{R6Class}} object.}
\usage{
GMLGridFunction
}
\value{
Object of \code{\link{R6Class}} for modelling an GML grid function
}
\description{
GMLGridFunction
}
\note{
Class used internally by geometa
}
\section{Methods}{

\describe{
 \item{\code{new(xml, element, attrs, defaults)}}{
   This method is used to instantiate a GML grid function
 }
 \item{\code{setSequenceRule(sequenceRule)}}{
   Set the sequence rule, as object of class \code{character}
 }
 \item{\code{setStartPoint(x,y)}}{
   Set the start point
 }
}
}

\references{
ISO 19136:2007 Geographic Information -- Geographic Markup Language.
  http://www.iso.org/iso/iso_catalogue/catalogue_tc/catalogue_detail.htm?csnumber=32554 
  OGC Geography Markup Language. http://www.opengeospatial.org/standards/gml
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{GML}
\keyword{GridFunction}
\keyword{ISO}
