% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMLMultiPoint.R
\docType{class}
\name{GMLMultiPoint}
\alias{GMLMultiPoint}
\title{GMLMultiPoint}
\format{\code{\link{R6Class}} object.}
\usage{
GMLMultiPoint
}
\value{
Object of \code{\link{R6Class}} for modelling an GML multipoint
}
\description{
GMLMultiPoint
}
\section{Fields}{

\describe{
\item{\code{pointMember}}{}
}}

\note{
Experimental
}
\section{Methods}{

\describe{
 \item{\code{new(xml, sfg)}}{
   This method is used to instantiate a GML multipoint
 }
 \item{\code{addPointMember(point)}}{
   Add a point member
 }
 \item{\code{delPointMember(point)}}{
   Deletes a point member
 }
}
}

\references{
ISO 19136:2007 Geographic Information -- Geographic Markup Language.
  http://www.iso.org/iso/iso_catalogue/catalogue_tc/catalogue_detail.htm?csnumber=32554 
  
  OGC Geography Markup Language. http://www.opengeospatial.org/standards/gml
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{GML}
\keyword{ISO}
\keyword{MultiPoint}
