% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMLMultiSurface.R
\docType{class}
\name{GMLMultiSurface}
\alias{GMLMultiSurface}
\title{GMLMultiSurface}
\format{\code{\link{R6Class}} object.}
\usage{
GMLMultiSurface
}
\value{
Object of \code{\link{R6Class}} for modelling an GML multisurface
}
\description{
GMLMultiSurface
}
\section{Fields}{

\describe{
\item{\code{surfaceMember}}{}
}}

\note{
Experimental
}
\section{Methods}{

\describe{
 \item{\code{new(xml, sfg)}}{
   This method is used to instantiate a GML multisurface
 }
 \item{\code{addSurfaceMember(surface)}}{
   Add a surface member
 }
 \item{\code{delSurfaceMember(surface)}}{
   Deletes a surface member
 }
}
}

\references{
ISO 19136:2007 Geographic Information -- Geographic Markup Language.
  http://www.iso.org/iso/iso_catalogue/catalogue_tc/catalogue_detail.htm?csnumber=32554 
  
  OGC Geography Markup Language. http://www.opengeospatial.org/standards/gml
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{GML}
\keyword{ISO}
\keyword{MultiSurface}
