% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMLOperationParameter.R
\docType{class}
\name{GMLOperationParameter}
\alias{GMLOperationParameter}
\title{GMLOperationParameter}
\format{\code{\link{R6Class}} object.}
\usage{
GMLOperationParameter
}
\value{
Object of \code{\link{R6Class}} for modelling an GMLOperationParameter
}
\description{
GMLOperationParameter
}
\section{Inherited methods}{

\describe{
  \code{GMLAbstractGeneralOperationParameter}
}
}

\section{Methods}{

\describe{
 \item{\code{new(xml, defaults, id)}}{
   This method is used to instantiate a GML OperationParameter
 }
}
}

\references{
ISO 19136:2007 Geographic Information -- Geographic Markup Language.
  http://www.iso.org/iso/iso_catalogue/catalogue_tc/catalogue_detail.htm?csnumber=32554 
  
  OGC Geography Markup Language. http://www.opengeospatial.org/standards/gml
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{GML}
\keyword{ISO}
\keyword{operation}
\keyword{parameter}
