% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMLOperationParameterGroup.R
\docType{class}
\name{GMLOperationParameterGroup}
\alias{GMLOperationParameterGroup}
\title{GMLOperationParameterGroup}
\format{\code{\link{R6Class}} object.}
\usage{
GMLOperationParameterGroup
}
\value{
Object of \code{\link{R6Class}} for modelling an GMLOperationParameterGroup
}
\description{
GMLOperationParameterGroup
}
\section{Fields}{

\describe{
\item{\code{maximumOccurs}}{[\code{\link{GMLElement}}]}

\item{\code{parameter}}{[\code{\link{GMLOperationParameter}}|\code{\link{GMLOperationParameterGroup}}]}
}}

\section{Inherited methods}{

\describe{
  from \code{GMLAbstractGeneralOperationParameter}
}
}

\section{Methods}{

\describe{
 \item{\code{new(xml, defaults, id)}}{
   This method is used to instantiate a GML OperationParameterGroup
 }
 \item{\code{setMaximumOccurs(maximumOccurs)}}{
   Sets the maximum occurs, object of class \code{integer}
 }
 \item{\code{addParameter(parameter)}}{
   Adds a parameter or parameter group, object of class \code{GMLOperationParameter}
   or \code{GMLOperationParameterGroup}
 }
 \item{\code{delParameter(parameter)}}{
   Deletes a parameter or parameter group, object of class \code{GMLOperationParameter}
   or \code{GMLOperationParameterGroupo}
 }
}
}

\references{
ISO 19136:2007 Geographic Information -- Geographic Markup Language.
  http://www.iso.org/iso/iso_catalogue/catalogue_tc/catalogue_detail.htm?csnumber=32554 
  
  OGC Geography Markup Language. http://www.opengeospatial.org/standards/gml
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{GML}
\keyword{ISO}
\keyword{group}
\keyword{operation}
\keyword{parameter}
