% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMLPoint.R
\docType{class}
\name{GMLPoint}
\alias{GMLPoint}
\title{GMLPoint}
\format{\code{\link{R6Class}} object.}
\usage{
GMLPoint
}
\value{
Object of \code{\link{R6Class}} for modelling an GML point
}
\description{
GMLPoint
}
\section{Fields}{

\describe{
\item{\code{pos}}{}
}}

\note{
Experimental
}
\section{Methods}{

\describe{
 \item{\code{new(xml, sfg = NULL, m = NULL)}}{
   This method is used to instantiate a GML point either using an object
   of class \code{sfg} (from \pkg{sf}), or using a \code{matrix} object
 }
}
}

\references{
ISO 19136:2007 Geographic Information -- Geographic Markup Language.
  http://www.iso.org/iso/iso_catalogue/catalogue_tc/catalogue_detail.htm?csnumber=32554 
  
  OGC Geography Markup Language. http://www.opengeospatial.org/standards/gml
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{GML}
\keyword{ISO}
\keyword{Point}
