% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMLProjectedCRS.R
\docType{class}
\name{GMLProjectedCRS}
\alias{GMLProjectedCRS}
\title{GMLProjectedCRS}
\format{\code{\link{R6Class}} object.}
\usage{
GMLProjectedCRS
}
\value{
Object of \code{\link{R6Class}} for modelling an GMLProjectedCRS
}
\description{
GMLProjectedCRS
}
\section{Fields}{

\describe{
\item{\code{baseGeodeticCRS}}{[\code{\link{GMLGeodeticCRS}}]}

\item{\code{cartesianCS}}{[\code{\link{GMLCartesianCS}}]}
}}

\section{Methods}{

\describe{
 \item{\code{new(xml, defaults, id)}}{
   This method is used to instantiate a GML projected CRS
 }
 \item{\code{setBaseGeodeticCRS(crs)}}{
   Sets the base geodetic CRS, object of class \code{GMLBaseGeodeticCRS}
 }
 \item{\code{setCartesianCS(cs)}}{
   Sets the cartesianCS, object of class \code{GMLCartesianCS}
 }
}
}

\references{
ISO 19136:2007 Geographic Information -- Geographic Markup Language.
  http://www.iso.org/iso/iso_catalogue/catalogue_tc/catalogue_detail.htm?csnumber=32554 
  
  OGC Geography Markup Language. http://www.opengeospatial.org/standards/gml
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{GML}
\keyword{ISO}
\keyword{crs}
\keyword{projected}
