% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMLReferenceableGridByArray.R
\docType{class}
\name{GMLReferenceableGridByArray}
\alias{GMLReferenceableGridByArray}
\title{GMLReferenceableGridByArray}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an GML ReferenceableGridByArray
}
\description{
GMLReferenceableGridByArray
}
\note{
Class used internally by geometa
}
\section{Methods}{

\describe{
 \item{\code{new(xml, element, attrs, defaults)}}{
   This method is used to instantiate a GML ReferenceableGridByArray
 }
}
}

\references{
OGC Geography Markup Language. http://www.opengeospatial.org/standards/gml
  
  OGC GML 3.3 Schema. http://schemas.opengis.net/gml/3.3/referenceableGrid.xsd
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{GML}
\keyword{Grid}
\keyword{ISO}
