% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMLReferenceableGridByVectors.R
\docType{class}
\name{GMLReferenceableGridByVectors}
\alias{GMLReferenceableGridByVectors}
\title{GMLReferenceableGridByVectors}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an GML ReferenceableGridByVectors
}
\description{
GMLReferenceableGridByVectors
}
\note{
Class used internally by geometa
}
\section{Methods}{

\describe{
 \item{\code{new(xml, element, attrs, defaults)}}{
   This method is used to instantiate a GML ReferenceableGridByVectors
 }
 \item{\code{setOrigin(coords)}}{
   Set the grid origin coordinates
 }
 \item{\code{addGeneralGridAxis(axis)}}{
   Add a general grid axis definition, object of class \code{GMLGeneralGridAxis}.
   Returns \code{TRUE} if added, \code{FALSE} otherwise.
 }
 \item{\code{delGeneralGridAxis(axis)}}{
   Deletes a ageneral grid axis definition,  object of class \code{GMLGeneralGridAxis}.
   Returns \code{TRUE} if deleted, \code{FALSE} otherwise.
 }
}
}

\references{
OGC Geography Markup Language. http://www.opengeospatial.org/standards/gml
  
  OGC GML 3.3 Schema. http://schemas.opengis.net/gml/3.3/referenceableGrid.xsd
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{GML}
\keyword{Grid}
\keyword{ISO}
