% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GMLTemporalPrimitive.R
\docType{class}
\name{GMLTemporalPrimitive}
\alias{GMLTemporalPrimitive}
\title{GMLTemporalPrimitive}
\format{\code{\link{R6Class}} object.}
\usage{
GMLTemporalPrimitive
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO GML abstract temporal primitive
}
\description{
GMLTemporalPrimitive
}
\note{
Class used internally by geometa
}
\section{Methods}{

\describe{
 \item{\code{new(xml, element, namespace, defaults)}}{
   This method is used to instantiate an ISOTemporalPrimitive
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\references{
ISO 19115:2003 - Geographic information -- Metadata
}
\keyword{ISO}
\keyword{primitive}
\keyword{temporal}

