% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEtoExpr.R
\name{GSE_GPL}
\alias{GSE_GPL}
\title{Get GPL id for multiple GSE}
\usage{
GSE_GPL(GSE, destdir = "temp-data")
}
\arguments{
\item{GSE}{A character, the number of GSE.}

\item{destdir}{A character, the path to download GSE related files.}
}
\value{
A data frame with GSE and GPL.
}
\description{
Get GPL id for multiple GSE
}
