% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEtoExpr.R
\name{GSEtoExpr}
\alias{GSEtoExpr}
\title{Download GSE expression matrix and convert probe to gene symbol}
\usage{
GSEtoExpr(GSE, destdir = "tmp")
}
\arguments{
\item{GSE}{A character, the number of GSE.}

\item{destdir}{A character, the path to download GSE related files.}
}
\value{
Save the files of expression matrix and phenotype information.
}
\description{
Given a GSE number, download related GSE data and convert probe to gene symbol, and return the gene matrix with gene symbol.
}
\details{
nothing
}
