% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEtoExpr.R
\name{GSEtoGPL}
\alias{GSEtoGPL}
\title{Get GPL id for multiple GSE}
\usage{
GSEtoGPL(GSE, destdir = "tmp")
}
\arguments{
\item{GSE}{A character, the number of GSE.}

\item{destdir}{A character, the path to download GSE related files.}
}
\value{
A data frame with GSE and GPL.
}
\description{
Get GPL id for multiple GSE
}
