% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOAbsoluteExternalPositionalAccuracy.R
\docType{class}
\name{ISOAbsoluteExternalPositionalAccuracy}
\alias{ISOAbsoluteExternalPositionalAccuracy}
\title{ISOAbsoluteExternalPositionalAccuracy}
\format{\code{\link{R6Class}} object.}
\usage{
ISOAbsoluteExternalPositionalAccuracy
}
\value{
Object of \code{\link{R6Class}} for modelling an ISOAbsoluteExternalPositionalAccuracy
}
\description{
ISOAbsoluteExternalPositionalAccuracy
}
\section{Inherited methods from \code{ISODataQualityAbstractElement}}{

}

\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to instantiate an ISOAbsoluteExternalPositionalAccuracy
 }
}
}
\examples{
#encoding
  dq <- ISOAbsoluteExternalPositionalAccuracy$new()
  dq$addNameOfMeasure("measure")
  metaId <- ISOMetaIdentifier$new(code = "measure-id")
  dq$setMeasureIdentification(metaId)
  dq$setMeasureDescription("description")
  dq$setEvaluationMethodDescription("method description")
  dq$setEvaluationMethodType("indirect")
  dq$setDateTime(ISOdate(2015,1,1,12,10,49))
  spec <- ISOCitation$new()
  spec$setTitle("specification title")
  spec$setAlternateTitle("specification alternate title")
  d <- ISODate$new()
  d$setDate(ISOdate(2015, 1, 1, 1))
  d$setDateType("publication")
  spec$addDate(d)
  dq$setEvaluationProcedure(spec)
  result <- ISOConformanceResult$new()
  result$setSpecification(spec)
  result$setExplanation("some explanation about the conformance")
  result$setPass(TRUE)
  dq$addResult(result)
  xml <- dq$encode()
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\references{
ISO 19115:2003 - Geographic information -- Metadata
}
\keyword{ISO}
\keyword{absolute}
\keyword{accuracy}
\keyword{data}
\keyword{external}
\keyword{positional}
\keyword{quality}

