% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOAbstractCatalogue.R
\docType{class}
\name{ISOAbstractCatalogue}
\alias{ISOAbstractCatalogue}
\title{ISOAbstractCatalogue}
\format{\code{\link{R6Class}} object.}
\usage{
ISOAbstractCatalogue
}
\value{
Object of \code{\link{R6Class}} for modelling an ISOAbstracCatalogue
}
\description{
ISOAbstractCatalogue
}
\section{Fields}{

\describe{
\item{\code{name}}{}

\item{\code{scope}}{}

\item{\code{fieldOfApplication}}{}

\item{\code{versionNumber}}{}

\item{\code{versionDate}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to instantiate an ISOAbstractCatalogue
 }
 \item{\code{setName(name)}}{
   Sets the name
 }
 \item{\code{addScope(scope)}}{
   Adds scope (object of class \code{character})
 }
 \item{\code{delScope(scope)}}{
   Deletes scope
 }
 \item{\code{addFieldOfApplication(fieldOfApplication)}}{
   Adds a field of application (object of class \code{character})
 }
 \item{\code{delFieldOfApplication(fieldOfApplication)}}{
   Deletes fieldOfApplication
 }
 \item{\code{setVersionNumber(versionNumber)}}{
   Sets version number (object of class \code{character})
 }
 \item{\code{setVersionDate(versionDate)}}{
   Sets version date
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\references{
ISO 19139:2007 Metadata - XML schema implementation
}
\keyword{ISO}
\keyword{abstract}
\keyword{catalogue}

