% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISODataQualityCompleteness.R
\docType{class}
\name{ISOAbstractCompleteness}
\alias{ISOAbstractCompleteness}
\title{ISOAbstractCompleteness}
\format{\code{\link{R6Class}} object.}
\usage{
ISOAbstractCompleteness
}
\value{
Object of \code{\link{R6Class}} for modelling an ISOAbstractCompleteness
}
\description{
ISOAbstractCompleteness
}
\section{Inherited methods}{

\describe{
  from \code{\link{ISODataQualityAbstractElement}}
}
}

\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to instantiate an \code{\link{ISOAbstractCompleteness}}
 }
}
}

\references{
ISO 19115:2003 - Geographic information -- Metadata
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{abstract}
\keyword{completeness}
\keyword{data}
\keyword{quality}
