% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISODataQualityLogicalConsistency.R
\docType{class}
\name{ISOAbstractLogicalConsistency}
\alias{ISOAbstractLogicalConsistency}
\title{ISOAbstractLogicalConsistency}
\format{\code{\link{R6Class}} object.}
\usage{
ISOAbstractLogicalConsistency
}
\value{
Object of \code{\link{R6Class}} for modelling an ISOAbstractLogicalConsistency
}
\description{
ISOAbstractLogicalConsistency
}
\section{Inherited methods}{

from \code{\link{ISODataQualityAbstractElement}}
}

\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to instantiate an \code{\link{ISOAbstractLogicalConsistency}}
 }
}
}

\references{
ISO 19115:2003 - Geographic information -- Metadata
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{abstract}
\keyword{consistency}
\keyword{data}
\keyword{logical}
\keyword{quality}
