% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISODataQualityPositionalAccuracy.R
\docType{class}
\name{ISOAbstractPositionalAccuracy}
\alias{ISOAbstractPositionalAccuracy}
\title{ISOAbstractPositionalAccuracy}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an ISOAbstractPositionalAccuracy
}
\description{
ISOAbstractPositionalAccuracy
}
\section{Inherited methods}{

from \code{\link{ISODataQualityAbstractElement}}
}

\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to instantiate an \code{\link{ISOAbstractPositionalAccuracy}}
 }
}
}

\references{
ISO 19115:2003 - Geographic information -- Metadata
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{abstract}
\keyword{accuracy}
\keyword{data}
\keyword{positional}
\keyword{quality}
