% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOAbstractResult.R
\docType{class}
\name{ISOAbstractResult}
\alias{ISOAbstractResult}
\title{ISOAbstractResult}
\format{\code{\link{R6Class}} object.}
\usage{
ISOAbstractResult
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO Result
}
\description{
ISOAbstractResult
}
\note{
abstract class
}
\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to instantiate an ISOResult
 }
}
}

\references{
ISO 19115:2003 - Geographic information -- Metadata
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{data}
\keyword{quality}
\keyword{result}
