% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISODataQualityTemporalAccuracy.R
\docType{class}
\name{ISOAbstractTemporalAccuracy}
\alias{ISOAbstractTemporalAccuracy}
\title{ISOAbstractTemporalAccuracy}
\format{\code{\link{R6Class}} object.}
\usage{
ISOAbstractTemporalAccuracy
}
\value{
Object of \code{\link{R6Class}} for modelling an ISOAbstractTemporalAccuracy
}
\description{
ISOAbstractTemporalAccuracy
}
\section{Inherited methods}{

\describe{
  from \code{ISODataQualityAbstractElement}
}
}

\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to instantiate an ISOAbstractTemporalAccuracy
 }
}
}

\references{
ISO 19115:2003 - Geographic information -- Metadata
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{abstract}
\keyword{accuracy}
\keyword{data}
\keyword{quality}
\keyword{temporal}
