% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISODataQualityThematicAccuracy.R
\docType{class}
\name{ISOAbstractThematicAccuracy}
\alias{ISOAbstractThematicAccuracy}
\title{ISOAbstractThematicAccuracy}
\format{\code{\link{R6Class}} object.}
\usage{
ISOAbstractThematicAccuracy
}
\value{
Object of \code{\link{R6Class}} for modelling an ISOAbstractThematicAccuracy
}
\description{
ISOAbstractThematicAccuracy
}
\section{Inherited methods}{

\describe{
  from \code{\link{ISODataQualityAbstractElement}}
}
}

\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to instantiate an \code{\link{ISOAbstractThematicAccuracy}}
 }
}
}

\references{
ISO 19115:2003 - Geographic information -- Metadata
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{abstract}
\keyword{accuracy}
\keyword{data}
\keyword{quality}
\keyword{thematic}
