% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISODataQualityTemporalAccuracy.R
\docType{class}
\name{ISOAccuracyOfATimeMeasurement}
\alias{ISOAccuracyOfATimeMeasurement}
\title{ISOAccuracyOfATimeMeasurement}
\format{\code{\link{R6Class}} object.}
\usage{
ISOAccuracyOfATimeMeasurement
}
\value{
Object of \code{\link{R6Class}} for modelling an ISOAccuracyOfATimeMeasurement
}
\description{
ISOAccuracyOfATimeMeasurement
}
\section{Inherited methods}{

\describe{
  from \code{ISODataQualityAbstractElement}
}
}

\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to instantiate an ISOAccuracyOfATimeMeasurement
 }
}
}

\examples{
  #encoding
  dq <- ISOAccuracyOfATimeMeasurement$new()
  dq$addNameOfMeasure("measure")
  metaId <- ISOMetaIdentifier$new(code = "measure-id")
  dq$setMeasureIdentification(metaId)
  dq$setMeasureDescription("description")
  dq$setEvaluationMethodDescription("method description")
  dq$setEvaluationMethodType("indirect")
  dq$setDateTime(ISOdate(2015,1,1,12,10,49))
  spec <- ISOCitation$new()
  spec$setTitle("specification title")
  spec$setAlternateTitle("specification alternate title")
  d <- ISODate$new()
  d$setDate(ISOdate(2015, 1, 1, 1))
  d$setDateType("publication")
  spec$addDate(d)
  dq$setEvaluationProcedure(spec)
  result <- ISOConformanceResult$new()
  result$setSpecification(spec)
  result$setExplanation("some explanation about the conformance")
  result$setPass(TRUE)
  dq$addResult(result)
  xml <- dq$encode()
  
}
\references{
ISO 19115:2003 - Geographic information -- Metadata
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{accuracy}
\keyword{data}
\keyword{measurement}
\keyword{quality}
\keyword{temporal}
\keyword{time}
