% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOAddress.R
\docType{class}
\name{ISOAddress}
\alias{ISOAddress}
\title{ISOAddress}
\format{\code{\link{R6Class}} object.}
\usage{
ISOAddress
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO Address
}
\description{
ISOAddress
}
\section{Fields}{

\describe{
\item{\code{deliveryPoint}}{}

\item{\code{city}}{}

\item{\code{postalCode}}{}

\item{\code{country}}{}

\item{\code{electronicMailAddress}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to instantiate an ISOAddress
 }
 \item{\code{setDeliveryPoint(deliveryPoint)}}{
   Sets the delivery point
 }
 \item{\code{setCity(city)}}{
   Sets the city
 }
 \item{\code{setPostalCode(postalCode)}}{
   Sets the postal code
 }
 \item{\code{setCountry(country)}}{
   Sets the country
 }
 \item{\code{setEmail(email)}}{
   Sets the electronic Mail address
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{address}

