% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOAnchor.R
\docType{class}
\name{ISOAnchor}
\alias{ISOAnchor}
\title{ISOAnchor}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO Anchor
}
\description{
ISOAnchor
}
\section{Methods}{

\describe{
 \item{\code{new(xml, name, ...)}}{
   This method is used to instantiate an \code{\link{ISOAnchor}}
 }
}
}

\examples{
  md <- ISOAnchor$new(name = "some entity name", href = "someentityuri")
  xml <- md$encode()

}
\references{
ISO/TS 19139:2007 Geographic information -- XML
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{anchor}
