% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOAngle.R
\docType{class}
\name{ISOAngle}
\alias{ISOAngle}
\title{ISOAngle}
\format{\code{\link{R6Class}} object.}
\usage{
ISOAngle
}
\value{
Object of \code{\link{R6Class}} for modelling an ISOAngle measure
}
\description{
ISOAngle
}
\section{Fields}{

\describe{
\item{\code{value}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(xml,value)}}{
   This method is used to instantiate an ISOAngle measure
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\references{
ISO/TS 19103:2005 Geographic information -- Conceptual schema language
}
\keyword{ISO}
\keyword{angle}
\keyword{measure}

