% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOAssociation.R
\docType{class}
\name{ISOAssociation}
\alias{ISOAssociation}
\title{ISOAssociation}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an ISOAssociation
}
\description{
ISOAssociation
}
\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to instantiate an \code{\link{ISOAssociation}}
 }
}
}

\references{
ISO 19115:2003 - Geographic information -- Metadata
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{association}
