% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOAssociationRole.R
\docType{class}
\name{ISOAssociationRole}
\alias{ISOAssociationRole}
\title{ISOAssociationRole}
\format{\code{\link{R6Class}} object.}
\usage{
ISOAssociationRole
}
\value{
Object of \code{\link{R6Class}} for modelling an ISOAssociationRole
}
\description{
ISOAssociationRole
}
\section{Fields}{

\describe{
\item{\code{type}}{[\code{\link{ISORoleType}}]}

\item{\code{isOrdered}}{[\code{\link{logical}}]}

\item{\code{isNavigable}}{[\code{\link{logical}}]}

\item{\code{relation}}{[\code{\link{ISOFeatureAssociation}}]}

\item{\code{rolePlayer}}{[\code{\link{ISOFeatureType}}]}
}}

\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to instantiate an \code{\link{ISOAssociationRole}}
 }
 \item{\code{setRoleType(roleType)}}{
   Sets the role type, object of class \code{\link{ISORoleType}} or any \code{character}
   value among \code{ISORoleType$values()}.
 }
 \item{\code{setIsOrdered(isOrdered)}}{
   Sets \code{TRUE} if ordered, \code{FALSE} otherwise
 }
 \item{\code{setIsNavigable(isNavigable)}}{
   Sets \code{TRUE} if navigable, \code{FALSE} otherwise
 }
 \item{code{setRelation(relation)}}{
   Sets an object of class \code{\link{ISOFeatureAssocation}} as relation
 }
 \item{\code{addRolePlayer(rolePlayer)}}{
   Adds a role player, object of class \code{\link{ISOFeatureType}}
 }
 \item{\code{delRolePlayer(rolePlayer)}}{
   Deletes a role player, object of class \code{\link{ISOFeatureType}}
 }
}
}

\references{
ISO 19110:2005 Methodology for Feature cataloguing
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{association}
\keyword{role}
