% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOAssociationType.R
\docType{class}
\name{ISOAssociationType}
\alias{ISOAssociationType}
\title{ISOAssociationType}
\format{\code{\link{R6Class}} object.}
\usage{
ISOAssociationType
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO AssociationType
}
\description{
ISOAssociationType
}
\section{Fields}{

\describe{
\item{\code{value}}{}
}}

\section{Methods}{

\describe{
 \item{\code{new(xml,value, description)}}{
   This method is used to instantiate an ISOAssociationType
 }
}
}

\examples{
  #possible values
  values <- ISOAssociationType$values(labels = TRUE)
  
  #geomOnly
  geomOnly <- ISOAssociationType$new(value = "source")
  
}
\references{
ISO 19115:2003 - Geographic information -- Metadata
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{association}
\keyword{type}
