% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOBaseCharacterString.R
\docType{class}
\name{ISOBaseCharacterString}
\alias{ISOBaseCharacterString}
\title{ISOBaseCharacterString}
\format{\code{\link{R6Class}} object.}
\usage{
ISOBaseCharacterString
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO BaseCharacterString
}
\description{
ISOBaseCharacterString
}
\section{Fields}{

\describe{
\item{\code{value}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(xml,value)}}{
   This method is used to instantiate an ISOBaseCharacterString
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{character}
\keyword{string}

