% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOBaseDate.R
\docType{class}
\name{ISOBaseDate}
\alias{ISOBaseDate}
\title{ISOBaseDate}
\format{\code{\link{R6Class}} object.}
\usage{
ISOBaseDate
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO Date
}
\description{
ISOBaseDate
}
\section{Fields}{

\describe{
\item{\code{value}}{}
}}
\note{
Class used by geometa internal XML decoder/encoder
}
\section{Methods}{

\describe{
 \item{\code{new(xml,value)}}{
   This method is used to instantiate an ISOBaseDate
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{date}

