% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOBaseDateTime.R
\docType{class}
\name{ISOBaseDateTime}
\alias{ISOBaseDateTime}
\title{ISOBaseDateTime}
\format{\code{\link{R6Class}} object.}
\usage{
ISOBaseDateTime
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO DateTime
}
\description{
ISOBaseDateTime
}
\section{Fields}{

\describe{
\item{\code{value}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(xml,value)}}{
   This method is used to instantiate an ISOBaseDateTime
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{datetime}

