% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOBaseReal.R
\docType{class}
\name{ISOBaseReal}
\alias{ISOBaseReal}
\title{ISOBaseReal}
\format{\code{\link{R6Class}} object.}
\usage{
ISOBaseReal
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO Real
}
\description{
ISOBaseReal
}
\section{Fields}{

\describe{
\item{\code{value}}{}
}}
\note{
Class used by geometa internal XML decoder/encoder
}
\section{Methods}{

\describe{
 \item{\code{new(xml,value)}}{
   This method is used to instantiate an ISOBaseReal
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{real}

