% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOBaseTimeBeginPosition.R
\docType{class}
\name{ISOBaseTimeBeginPosition}
\alias{ISOBaseTimeBeginPosition}
\title{ISOBaseTimeBeginPosition}
\format{\code{\link{R6Class}} object.}
\usage{
ISOBaseTimeBeginPosition
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO Time begin position
}
\description{
ISOBaseTimeBeginPosition
}
\section{Fields}{

\describe{
\item{\code{value}}{}
}}
\note{
Class used by geometa internal XML decoder/encoder
}
\section{Methods}{

\describe{
 \item{\code{new(xml,value)}}{
   This method is used to instantiate an ISOBaseTimeBeginPosition
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{begin}
\keyword{position}

