% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOBaseTimeEndPosition.R
\docType{class}
\name{ISOBaseTimeEndPosition}
\alias{ISOBaseTimeEndPosition}
\title{ISOBaseTimeEndPosition}
\format{\code{\link{R6Class}} object.}
\usage{
ISOBaseTimeEndPosition
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO Time end position
}
\description{
ISOBaseTimeEndPosition
}
\section{Fields}{

\describe{
\item{\code{value}}{}
}}
\note{
Class used by geometa internal XML decoder/encoder
}
\section{Methods}{

\describe{
 \item{\code{new(xml,value)}}{
   This method is used to instantiate an ISOBaseTimeEndPosition
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{end}
\keyword{position}

