% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOBaseURL.R
\docType{class}
\name{ISOBaseURL}
\alias{ISOBaseURL}
\title{ISOBaseURL}
\format{\code{\link{R6Class}} object.}
\usage{
ISOBaseURL
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO BaseURL
}
\description{
ISOBaseURL
}
\section{Fields}{

\describe{
\item{\code{value}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(value)}}{
   This method is used to instantiate an ISOBaseURL
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{URL}

