% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOBinary.R
\docType{class}
\name{ISOBinary}
\alias{ISOBinary}
\title{ISOBinary}
\format{\code{\link{R6Class}} object.}
\usage{
ISOBinary
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO UnlimitedInteger
}
\description{
ISOBinary
}
\section{Fields}{

\describe{
\item{\code{value}}{[\code{\link{character}}] binary}
}}

\section{Methods}{

\describe{
 \item{\code{new(xml,value)}}{
   This method is used to instantiate an \code{\link{ISOBinary}}
 }
}
}

\examples{
  bin <- ISOBinary$new(value = "http://someuri")
 
}
\references{
ISO/TS 19103:2005 Geographic information -- Conceptual schema language
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{binary}
