% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOBoundAssociationRole.R
\docType{class}
\name{ISOBoundAssociationRole}
\alias{ISOBoundAssociationRole}
\title{ISOBoundAssociationRole}
\format{\code{\link{R6Class}} object.}
\usage{
ISOBoundAssociationRole
}
\value{
Object of \code{\link{R6Class}} for modelling an ISOBoundAssociationRole
}
\description{
ISOBoundAssociationRole
}
\section{Fields}{

\describe{
\item{\code{rolePlayer}}{[\code{\link{ISOFeatureType}}] feature type}
}}

\section{Methods}{

\describe{
 \item{\code{new(xml, defaults)}}{
   This method is used to instantiate an \code{\link{ISOBoundAssociationRole}}
 }
 \item{\code{setFeatureType(featureType)}}{
   Set feature type, object of class \code{\link{ISOFeatureType}}
 }
}
}

\references{
ISO 19110:2005 Methodology for Feature cataloguing
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{BoundAssociationRole}
\keyword{ISO}
