% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOBoundFeatureAttribute.R
\docType{class}
\name{ISOBoundFeatureAttribute}
\alias{ISOBoundFeatureAttribute}
\title{ISOBoundFeatureAttribute}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an ISOBoundFeatureAttribute
}
\description{
ISOBoundFeatureAttribute
}
\section{Fields}{

\describe{
\item{\code{valueType}}{[\code{\link{ISOTypeName}}] value type}
}}

\section{Methods}{

\describe{
 \item{\code{new(xml, defaults)}}{
   This method is used to instantiate an \code{\link{ISOBoundFeatureAttribute}}
 }
 \item{\code{setTypeName(typeName)}}{
   Set typeName, object of class \code{\link{ISOTypeName}} or \code{character}
 }
}
}

\references{
ISO 19110:2005 Methodology for Feature cataloguing
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{BoundFeatureAttribute}
\keyword{ISO}
