% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOBoundingPolygon.R
\docType{class}
\name{ISOBoundingPolygon}
\alias{ISOBoundingPolygon}
\title{ISOBoundingPolygon}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO BoundingPolygon
}
\description{
ISOBoundingPolygon
}
\section{Fields}{

\describe{
\item{\code{polygon}}{[\code{\link{GMLAbstractGeometry}}] a (list of) geometrie(s)}
}}

\note{
Experimental
}
\section{Methods}{

\describe{
 \item{\code{new(xml, sfg)}}{
   This method is used to instantiate an \code{\link{ISOBoundingPolygon}}
 }
}
}

\references{
ISO 19115:2003 - Geographic information -- Metadata
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{bounding}
\keyword{polygon}
