% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOBrowseGraphic.R
\docType{class}
\name{ISOBrowseGraphic}
\alias{ISOBrowseGraphic}
\title{ISOBrowseGraphic}
\format{\code{\link{R6Class}} object.}
\usage{
ISOBrowseGraphic
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO BrowseGraphic
}
\description{
ISOBrowseGraphic
}
\section{Fields}{

\describe{
\item{\code{value}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(value, fileName, fileDescription, fileType)}}{
   This method is used to instantiate an ISOBrowseGraphic
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{browse}
\keyword{graphic}

