% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOCellGeometry.R
\docType{class}
\name{ISOCellGeometry}
\alias{ISOCellGeometry}
\title{ISOCellGeometry}
\format{\code{\link{R6Class}} object.}
\usage{
ISOCellGeometry
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO CellGeometryCode
}
\description{
ISOCellGeometry
}
\section{Fields}{

\describe{
\item{\code{value}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(xml,value, description)}}{
   This method is used to instantiate an ISOCellGeometry
 }
}
}
\examples{
#possible values
  values <- ISOCellGeometry$values(labels = TRUE)

  #example of 'point' cell geometry code
  pointCode <- ISOCellGeometry$new(value = "point")
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\references{
ISO 19115:2003 - Geographic information -- Metadata
}
\keyword{ISO}
\keyword{cell}
\keyword{geometry}

