% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOCellGeometryCode.R
\docType{class}
\name{ISOCellGeometryCode}
\alias{ISOCellGeometryCode}
\title{ISOCellGeometryCode}
\format{\code{\link{R6Class}} object.}
\usage{
ISOCellGeometryCode
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO CellGeometryCode
}
\description{
ISOCellGeometryCode
}
\section{Fields}{

\describe{
\item{\code{value}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(xml,value)}}{
   This method is used to instantiate an ISOCellGeometryCode
 }
}
}
\examples{
#possible values
  values <- ISOCellGeometryCode$values(labels = TRUE)

  #example of 'point' cell geometry code
  pointCode <- ISOCellGeometryCode$new(value = "point")
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{cell}
\keyword{code}
\keyword{geometry}

