% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOCharacterSet.R
\docType{class}
\name{ISOCharacterSet}
\alias{ISOCharacterSet}
\title{ISOCharacterSet}
\format{\code{\link{R6Class}} object.}
\usage{
ISOCharacterSet
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO CharacterSet
}
\description{
ISOCharacterSet
}
\section{Fields}{

\describe{
\item{\code{value}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(value)}}{
   This method is used to instantiate an ISOCharacterSet
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{characterSet}
\keyword{charset}

