% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOCitation.R
\docType{class}
\name{ISOCitation}
\alias{ISOCitation}
\title{ISOCitation}
\format{\code{\link{R6Class}} object.}
\usage{
ISOCitation
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO Citation
}
\description{
ISOCitation
}
\section{Fields}{

\describe{
\item{\code{title}}{}

\item{\code{alternateTitle}}{}

\item{\code{date}}{}

\item{\code{edition}}{}

\item{\code{editionDate}}{}

\item{\code{identifier}}{}

\item{\code{presentationForm}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to instantiate an ISOCitation
 }
 \item{\code{setTitle(title)}}{
   Sets the title
 }
 \item{\code{setAlternateTitle(alternateTitle)}}{
   Sets an alternate title
 }
 \item{\code{addDate(date)}}{
   Adds the date (ISODate object containing date and dateType)
 }
 \item{\code{setEdition(edition)}}{
   Sets the edition
 }
 \item{\code{setEditionDate(editionDate)}}{
   Sets the edition date (ISODate object containing date and dateType)
 }
 \item{\code{setIdentifier(code, codeSpace)}}{
   Sets the identifier
 }
 \item{\code{setCitedResponsibleParty(rp)}}{
   Sets the cited responsiblep party
 }
 \item{\code{setPresentationForm}}{
   Sets the presentation form
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{citation}

