% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOCodelist.R
\docType{class}
\name{ISOCodelist}
\alias{ISOCodelist}
\title{ISOCodelist}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO codelist
}
\description{
ISOCodelist
}
\section{Fields}{

\describe{
\item{\code{id}}{[\code{\link{character}}] codelist id}

\item{\code{refFile}}{[\code{\link{character}}] reference XML file}

\item{\code{codeSpace}}{[\code{\link{character}}] codelist codeSpace}

\item{\code{identifier}}{[\code{\link{character}}] codelist identifier}

\item{\code{description}}{[\code{\link{character}}] codelist description}

\item{\code{entries}}{[\code{\link{data.frame}}] the list of codelist entries}
}}

\note{
Class used by geometa internal codelist XML decoder/encoder
}
\section{Methods}{

\describe{
 \item{\code{new(refFile, id)}}{
   This method is used to instantiate an \code{\link{ISOCodelist}}
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{codelist}
