% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISODataQualityLogicalConsistency.R
\docType{class}
\name{ISOConceptualConsistency}
\alias{ISOConceptualConsistency}
\title{ISOConceptualConsistency}
\format{
\code{\link[R6]{R6Class}} object.
}
\value{
Object of \code{\link[R6]{R6Class}} for modelling an ISOConceptualConsistency
}
\description{
ISOConceptualConsistency

ISOConceptualConsistency
}
\examples{
  #encoding
  dq <- ISOConceptualConsistency$new()
  dq$addNameOfMeasure("measure")
  metaId <- ISOMetaIdentifier$new(code = "measure-id")
  dq$setMeasureIdentification(metaId)
  dq$setMeasureDescription("description")
  dq$setEvaluationMethodDescription("method description")
  dq$setEvaluationMethodType("indirect")
  dq$setDateTime(ISOdate(2015,1,1,12,10,49))
  spec <- ISOCitation$new()
  spec$setTitle("specification title")
  spec$addAlternateTitle("specification alternate title")
  d <- ISODate$new()
  d$setDate(ISOdate(2015, 1, 1, 1))
  d$setDateType("publication")
  spec$addDate(d)
  dq$setEvaluationProcedure(spec)
  result <- ISOConformanceResult$new()
  result$setSpecification(spec)
  result$setExplanation("some explanation about the conformance")
  result$setPass(TRUE)
  dq$addResult(result)
  xml <- dq$encode()
  
}
\references{
- ISO 19139 \url{https://schemas.isotc211.org/19139/-/gmd/1.0/gmd/#element_DQ_ConceptualConsistency}
  
  - ISO 19115-3 \url{https://schemas.isotc211.org/19157/-/mdq/1.2/mdq/#element_DQ_ConceptualConsistency}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{conceptual}
\keyword{consistency}
\keyword{data}
\keyword{quality}
\section{Super classes}{
\code{\link[geometa:geometaLogger]{geometa::geometaLogger}} -> \code{\link[geometa:ISOAbstractObject]{geometa::ISOAbstractObject}} -> \code{\link[geometa:ISOAbstractQualityElement]{geometa::ISOAbstractQualityElement}} -> \code{\link[geometa:ISODataQualityAbstractElement]{geometa::ISODataQualityAbstractElement}} -> \code{\link[geometa:ISOAbstractLogicalConsistency]{geometa::ISOAbstractLogicalConsistency}} -> \code{ISOConceptualConsistency}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ISOConceptualConsistency-clone}{\code{ISOConceptualConsistency$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="geometa" data-topic="geometaLogger" data-id="ERROR"><a href='../../geometa/html/geometaLogger.html#method-geometaLogger-ERROR'><code>geometa::geometaLogger$ERROR()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="geometaLogger" data-id="INFO"><a href='../../geometa/html/geometaLogger.html#method-geometaLogger-INFO'><code>geometa::geometaLogger$INFO()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="geometaLogger" data-id="WARN"><a href='../../geometa/html/geometaLogger.html#method-geometaLogger-WARN'><code>geometa::geometaLogger$WARN()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="addFieldAttrs"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-addFieldAttrs'><code>geometa::ISOAbstractObject$addFieldAttrs()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="addListElement"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-addListElement'><code>geometa::ISOAbstractObject$addListElement()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="checkMetadataStandardCompliance"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-checkMetadataStandardCompliance'><code>geometa::ISOAbstractObject$checkMetadataStandardCompliance()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="contains"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-contains'><code>geometa::ISOAbstractObject$contains()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="createLocalisedProperty"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-createLocalisedProperty'><code>geometa::ISOAbstractObject$createLocalisedProperty()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="decode"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-decode'><code>geometa::ISOAbstractObject$decode()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="delListElement"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-delListElement'><code>geometa::ISOAbstractObject$delListElement()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="encode"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-encode'><code>geometa::ISOAbstractObject$encode()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="getClass"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-getClass'><code>geometa::ISOAbstractObject$getClass()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="getClassName"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-getClassName'><code>geometa::ISOAbstractObject$getClassName()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="getNamespaceDefinition"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-getNamespaceDefinition'><code>geometa::ISOAbstractObject$getNamespaceDefinition()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="isDocument"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-isDocument'><code>geometa::ISOAbstractObject$isDocument()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="isFieldInheritedFrom"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-isFieldInheritedFrom'><code>geometa::ISOAbstractObject$isFieldInheritedFrom()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="print"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-print'><code>geometa::ISOAbstractObject$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="save"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-save'><code>geometa::ISOAbstractObject$save()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="setAttr"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-setAttr'><code>geometa::ISOAbstractObject$setAttr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="setCodeList"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-setCodeList'><code>geometa::ISOAbstractObject$setCodeList()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="setCodeListValue"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-setCodeListValue'><code>geometa::ISOAbstractObject$setCodeListValue()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="setCodeSpace"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-setCodeSpace'><code>geometa::ISOAbstractObject$setCodeSpace()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="setHref"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-setHref'><code>geometa::ISOAbstractObject$setHref()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="setId"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-setId'><code>geometa::ISOAbstractObject$setId()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="setIsNull"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-setIsNull'><code>geometa::ISOAbstractObject$setIsNull()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="setValue"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-setValue'><code>geometa::ISOAbstractObject$setValue()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="stopIfMetadataStandardIsNot"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-stopIfMetadataStandardIsNot'><code>geometa::ISOAbstractObject$stopIfMetadataStandardIsNot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="validate"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-validate'><code>geometa::ISOAbstractObject$validate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="wrapBaseElement"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-wrapBaseElement'><code>geometa::ISOAbstractObject$wrapBaseElement()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISODataQualityAbstractElement" data-id="addDerivedElement"><a href='../../geometa/html/ISODataQualityAbstractElement.html#method-ISODataQualityAbstractElement-addDerivedElement'><code>geometa::ISODataQualityAbstractElement$addDerivedElement()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISODataQualityAbstractElement" data-id="addNameOfMeasure"><a href='../../geometa/html/ISODataQualityAbstractElement.html#method-ISODataQualityAbstractElement-addNameOfMeasure'><code>geometa::ISODataQualityAbstractElement$addNameOfMeasure()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISODataQualityAbstractElement" data-id="addResult"><a href='../../geometa/html/ISODataQualityAbstractElement.html#method-ISODataQualityAbstractElement-addResult'><code>geometa::ISODataQualityAbstractElement$addResult()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISODataQualityAbstractElement" data-id="delDerivedElement"><a href='../../geometa/html/ISODataQualityAbstractElement.html#method-ISODataQualityAbstractElement-delDerivedElement'><code>geometa::ISODataQualityAbstractElement$delDerivedElement()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISODataQualityAbstractElement" data-id="delNameOfMeasure"><a href='../../geometa/html/ISODataQualityAbstractElement.html#method-ISODataQualityAbstractElement-delNameOfMeasure'><code>geometa::ISODataQualityAbstractElement$delNameOfMeasure()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISODataQualityAbstractElement" data-id="delResult"><a href='../../geometa/html/ISODataQualityAbstractElement.html#method-ISODataQualityAbstractElement-delResult'><code>geometa::ISODataQualityAbstractElement$delResult()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISODataQualityAbstractElement" data-id="initialize"><a href='../../geometa/html/ISODataQualityAbstractElement.html#method-ISODataQualityAbstractElement-initialize'><code>geometa::ISODataQualityAbstractElement$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISODataQualityAbstractElement" data-id="setDateTime"><a href='../../geometa/html/ISODataQualityAbstractElement.html#method-ISODataQualityAbstractElement-setDateTime'><code>geometa::ISODataQualityAbstractElement$setDateTime()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISODataQualityAbstractElement" data-id="setEvaluationMethod"><a href='../../geometa/html/ISODataQualityAbstractElement.html#method-ISODataQualityAbstractElement-setEvaluationMethod'><code>geometa::ISODataQualityAbstractElement$setEvaluationMethod()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISODataQualityAbstractElement" data-id="setEvaluationMethodDescription"><a href='../../geometa/html/ISODataQualityAbstractElement.html#method-ISODataQualityAbstractElement-setEvaluationMethodDescription'><code>geometa::ISODataQualityAbstractElement$setEvaluationMethodDescription()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISODataQualityAbstractElement" data-id="setEvaluationMethodType"><a href='../../geometa/html/ISODataQualityAbstractElement.html#method-ISODataQualityAbstractElement-setEvaluationMethodType'><code>geometa::ISODataQualityAbstractElement$setEvaluationMethodType()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISODataQualityAbstractElement" data-id="setEvaluationProcedure"><a href='../../geometa/html/ISODataQualityAbstractElement.html#method-ISODataQualityAbstractElement-setEvaluationProcedure'><code>geometa::ISODataQualityAbstractElement$setEvaluationProcedure()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISODataQualityAbstractElement" data-id="setMeasure"><a href='../../geometa/html/ISODataQualityAbstractElement.html#method-ISODataQualityAbstractElement-setMeasure'><code>geometa::ISODataQualityAbstractElement$setMeasure()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISODataQualityAbstractElement" data-id="setMeasureDescription"><a href='../../geometa/html/ISODataQualityAbstractElement.html#method-ISODataQualityAbstractElement-setMeasureDescription'><code>geometa::ISODataQualityAbstractElement$setMeasureDescription()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISODataQualityAbstractElement" data-id="setMeasureIdentification"><a href='../../geometa/html/ISODataQualityAbstractElement.html#method-ISODataQualityAbstractElement-setMeasureIdentification'><code>geometa::ISODataQualityAbstractElement$setMeasureIdentification()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISODataQualityAbstractElement" data-id="setStandaloneQualityReportDetails"><a href='../../geometa/html/ISODataQualityAbstractElement.html#method-ISODataQualityAbstractElement-setStandaloneQualityReportDetails'><code>geometa::ISODataQualityAbstractElement$setStandaloneQualityReportDetails()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ISOConceptualConsistency-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ISOConceptualConsistency-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ISOConceptualConsistency$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
