% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOConstraint.R
\docType{class}
\name{ISOConstraint}
\alias{ISOConstraint}
\title{ISOConstraint}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an ISOConstraint
}
\description{
ISOConstraint
}
\section{Fields}{

\describe{
\item{\code{description}}{[\code{\link{character}}]}
}}

\section{Methods}{

\describe{
 \item{\code{new(xml, description)}}{
   This method is used to instantiate an \code{\link{ISOConstraint}}
 }
 \item{\code{setDescription(description, locales)}}{
   Sets the description. Locale names can be specified 
   as \code{list} with the \code{locales} argument.
 }
}
}

\examples{
  md <- ISOConstraint$new(description = "description")
  xml <- md$encode()
 
}
\references{
ISO 19110:2005 Methodology for Feature cataloguing
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{constraint}
\keyword{feature}
