% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOConstraints.R
\docType{class}
\name{ISOConstraints}
\alias{ISOConstraints}
\title{ISOConstraints}
\format{\code{\link{R6Class}} object.}
\usage{
ISOConstraints
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO abstract Constraints
}
\description{
ISOConstraints
}
\section{Fields}{

\describe{
\item{\code{useLimitation}}{}
}}

\note{
Abstract ISO class
}
\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to instantiate an ISOLegalConstraints
 }
 \item{\code{addUseLimitation(useLimitation, locales)}}{
   Adds a use limitation. Locale names can be specified 
   as \code{list} with the \code{locales} argument.
 }
 \item{\code{setUseLimitation(useLimitation, locales)}}{
   Sets a use limitation. Locale names can be specified 
   as \code{list} with the \code{locales} argument.
 }
 \item{\code{delUseLimitation(useLimitation, locales)}}{
   Deletes a use limitation. Locale names can be specified 
   as \code{list} with the \code{locales} argument.
 }
}
}

\references{
ISO 19115:2003 - Geographic information -- Metadata
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{constraints}
