% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOContentInformation.R
\docType{class}
\name{ISOContentInformation}
\alias{ISOContentInformation}
\title{ISOContentInformation}
\format{\code{\link{R6Class}} object.}
\usage{
ISOContentInformation
}
\value{
Object of \code{\link{R6Class}} for modelling an ISOContentInformation
}
\description{
ISOContentInformation
}
\note{
Abstract class. Used internally by \pkg{geometa}
}
\section{Abstract Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to instantiate an \code{\link{ISOContentInformation}}
 }
}
}

\references{
ISO 19115:2003 - Geographic information -- Metadata
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{content}
\keyword{information}
