% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOCountry.R
\docType{class}
\name{ISOCountry}
\alias{ISOCountry}
\title{ISOCountry}
\format{\code{\link{R6Class}} object.}
\usage{
ISOCountry
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO Country
}
\description{
ISOCountry
}
\section{Methods}{

\describe{
 \item{\code{new(xml,value)}}{
   This method is used to instantiate an \code{\link{ISOCountry}}
 }
}
}

\examples{
  #possible values
  values <- ISOCountry$values(labels = TRUE)
  
  #some charset
  charset <- ISOCountry$new(value = "utf8")

}
\references{
ISO 19115:2003 - Geographic information -- Metadata
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{country}
