% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOCouplingType.R
\docType{class}
\name{ISOCouplingType}
\alias{ISOCouplingType}
\title{ISOCouplingType}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an ISOCouplingType
}
\description{
ISOCouplingType
}
\section{Methods}{

\describe{
 \item{\code{new(xml,value, description)}}{
   This method is used to instantiate an \code{\link{ISOCouplingType}}
 }
}
}

\examples{
  #possible values
  values <- ISOCouplingType$values(labels = TRUE)
  
  #couplingType
  couplingType <- ISOCouplingType$new(value = "loose")

}
\references{
ISO 19119:2005 - Geographic information -- Services
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{coupling}
\keyword{type}
