% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOCoverageContentType.R
\docType{class}
\name{ISOCoverageContentType}
\alias{ISOCoverageContentType}
\title{ISOCoverageContentType}
\format{\code{\link{R6Class}} object.}
\usage{
ISOCoverageContentType
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO CoverageContentType
}
\description{
ISOCoverageContentType
}
\section{Methods}{

\describe{
 \item{\code{new(xml,value, description)}}{
   This method is used to instantiate an \code{\link{ISOCoverageContentType}}
 }
}
}

\examples{
  #possible values
  values <- ISOCoverageContentType$values(labels = TRUE)
  
  #example of CoverageContentType
  modelResultType <- ISOCoverageContentType$new(value = "modelResult")
  
}
\references{
ISO 19115:2003 - Geographic information -- Metadata
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{CoverageContentType}
\keyword{ISO}
