% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISODataIdentification19115_3.R
\docType{class}
\name{ISODataIdentification19115_3}
\alias{ISODataIdentification19115_3}
\title{ISODataIdentification}
\format{
\code{\link[R6]{R6Class}} object.
}
\value{
Object of \code{\link[R6]{R6Class}} for modelling an ISO DataIdentification
}
\description{
ISODataIdentification

ISODataIdentification
}
\references{
- ISO 19115-3 \url{https://schemas.isotc211.org/19115/-3/mri/1.0/mri/#element_MD_DataIdentification}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{data}
\keyword{identification}
\section{Super classes}{
\code{\link[geometa:geometaLogger]{geometa::geometaLogger}} -> \code{\link[geometa:ISOAbstractObject]{geometa::ISOAbstractObject}} -> \code{\link[geometa:ISOIdentification]{geometa::ISOIdentification}} -> \code{\link[geometa:ISOIdentification19115_3]{geometa::ISOIdentification19115_3}} -> \code{ISODataIdentification19115_3}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{defaultLocale}}{defaultLocale [0..1]: ISOLocale}

\item{\code{otherLocale}}{otherLocale [0..*]: ISOLocale}

\item{\code{environmentDescription}}{environment description [0..1]: character}

\item{\code{supplementalInformation}}{supplementalInformation [0..1]: character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ISODataIdentification19115_3-new}{\code{ISODataIdentification19115_3$new()}}
\item \href{#method-ISODataIdentification19115_3-setDefaultLocale}{\code{ISODataIdentification19115_3$setDefaultLocale()}}
\item \href{#method-ISODataIdentification19115_3-addOtherLocale}{\code{ISODataIdentification19115_3$addOtherLocale()}}
\item \href{#method-ISODataIdentification19115_3-delOtherLocale}{\code{ISODataIdentification19115_3$delOtherLocale()}}
\item \href{#method-ISODataIdentification19115_3-setEnvironmentDescription}{\code{ISODataIdentification19115_3$setEnvironmentDescription()}}
\item \href{#method-ISODataIdentification19115_3-setSupplementalInformation}{\code{ISODataIdentification19115_3$setSupplementalInformation()}}
\item \href{#method-ISODataIdentification19115_3-clone}{\code{ISODataIdentification19115_3$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="geometa" data-topic="geometaLogger" data-id="ERROR"><a href='../../geometa/html/geometaLogger.html#method-geometaLogger-ERROR'><code>geometa::geometaLogger$ERROR()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="geometaLogger" data-id="INFO"><a href='../../geometa/html/geometaLogger.html#method-geometaLogger-INFO'><code>geometa::geometaLogger$INFO()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="geometaLogger" data-id="WARN"><a href='../../geometa/html/geometaLogger.html#method-geometaLogger-WARN'><code>geometa::geometaLogger$WARN()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="addFieldAttrs"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-addFieldAttrs'><code>geometa::ISOAbstractObject$addFieldAttrs()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="addListElement"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-addListElement'><code>geometa::ISOAbstractObject$addListElement()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="checkMetadataStandardCompliance"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-checkMetadataStandardCompliance'><code>geometa::ISOAbstractObject$checkMetadataStandardCompliance()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="contains"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-contains'><code>geometa::ISOAbstractObject$contains()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="createLocalisedProperty"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-createLocalisedProperty'><code>geometa::ISOAbstractObject$createLocalisedProperty()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="decode"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-decode'><code>geometa::ISOAbstractObject$decode()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="delListElement"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-delListElement'><code>geometa::ISOAbstractObject$delListElement()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="encode"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-encode'><code>geometa::ISOAbstractObject$encode()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="getClass"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-getClass'><code>geometa::ISOAbstractObject$getClass()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="getClassName"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-getClassName'><code>geometa::ISOAbstractObject$getClassName()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="getNamespaceDefinition"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-getNamespaceDefinition'><code>geometa::ISOAbstractObject$getNamespaceDefinition()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="isDocument"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-isDocument'><code>geometa::ISOAbstractObject$isDocument()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="isFieldInheritedFrom"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-isFieldInheritedFrom'><code>geometa::ISOAbstractObject$isFieldInheritedFrom()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="print"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-print'><code>geometa::ISOAbstractObject$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="save"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-save'><code>geometa::ISOAbstractObject$save()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="setAttr"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-setAttr'><code>geometa::ISOAbstractObject$setAttr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="setCodeList"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-setCodeList'><code>geometa::ISOAbstractObject$setCodeList()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="setCodeListValue"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-setCodeListValue'><code>geometa::ISOAbstractObject$setCodeListValue()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="setCodeSpace"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-setCodeSpace'><code>geometa::ISOAbstractObject$setCodeSpace()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="setHref"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-setHref'><code>geometa::ISOAbstractObject$setHref()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="setId"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-setId'><code>geometa::ISOAbstractObject$setId()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="setIsNull"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-setIsNull'><code>geometa::ISOAbstractObject$setIsNull()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="setValue"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-setValue'><code>geometa::ISOAbstractObject$setValue()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="stopIfMetadataStandardIsNot"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-stopIfMetadataStandardIsNot'><code>geometa::ISOAbstractObject$stopIfMetadataStandardIsNot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="validate"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-validate'><code>geometa::ISOAbstractObject$validate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="wrapBaseElement"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-wrapBaseElement'><code>geometa::ISOAbstractObject$wrapBaseElement()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOIdentification19115_3" data-id="addAdditionalDocumentation"><a href='../../geometa/html/ISOIdentification19115_3.html#method-ISOIdentification19115_3-addAdditionalDocumentation'><code>geometa::ISOIdentification19115_3$addAdditionalDocumentation()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOIdentification19115_3" data-id="addAssociatedResource"><a href='../../geometa/html/ISOIdentification19115_3.html#method-ISOIdentification19115_3-addAssociatedResource'><code>geometa::ISOIdentification19115_3$addAssociatedResource()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOIdentification19115_3" data-id="addCredit"><a href='../../geometa/html/ISOIdentification19115_3.html#method-ISOIdentification19115_3-addCredit'><code>geometa::ISOIdentification19115_3$addCredit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOIdentification19115_3" data-id="addExtent"><a href='../../geometa/html/ISOIdentification19115_3.html#method-ISOIdentification19115_3-addExtent'><code>geometa::ISOIdentification19115_3$addExtent()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOIdentification19115_3" data-id="addFormat"><a href='../../geometa/html/ISOIdentification19115_3.html#method-ISOIdentification19115_3-addFormat'><code>geometa::ISOIdentification19115_3$addFormat()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOIdentification19115_3" data-id="addGraphicOverview"><a href='../../geometa/html/ISOIdentification19115_3.html#method-ISOIdentification19115_3-addGraphicOverview'><code>geometa::ISOIdentification19115_3$addGraphicOverview()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOIdentification19115_3" data-id="addKeywords"><a href='../../geometa/html/ISOIdentification19115_3.html#method-ISOIdentification19115_3-addKeywords'><code>geometa::ISOIdentification19115_3$addKeywords()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOIdentification19115_3" data-id="addPointOfContact"><a href='../../geometa/html/ISOIdentification19115_3.html#method-ISOIdentification19115_3-addPointOfContact'><code>geometa::ISOIdentification19115_3$addPointOfContact()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOIdentification19115_3" data-id="addResourceConstraints"><a href='../../geometa/html/ISOIdentification19115_3.html#method-ISOIdentification19115_3-addResourceConstraints'><code>geometa::ISOIdentification19115_3$addResourceConstraints()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOIdentification19115_3" data-id="addResourceMaintenance"><a href='../../geometa/html/ISOIdentification19115_3.html#method-ISOIdentification19115_3-addResourceMaintenance'><code>geometa::ISOIdentification19115_3$addResourceMaintenance()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOIdentification19115_3" data-id="addResourceSpecificUsage"><a href='../../geometa/html/ISOIdentification19115_3.html#method-ISOIdentification19115_3-addResourceSpecificUsage'><code>geometa::ISOIdentification19115_3$addResourceSpecificUsage()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOIdentification19115_3" data-id="addSpatialRepresentationType"><a href='../../geometa/html/ISOIdentification19115_3.html#method-ISOIdentification19115_3-addSpatialRepresentationType'><code>geometa::ISOIdentification19115_3$addSpatialRepresentationType()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOIdentification19115_3" data-id="addSpatialResolution"><a href='../../geometa/html/ISOIdentification19115_3.html#method-ISOIdentification19115_3-addSpatialResolution'><code>geometa::ISOIdentification19115_3$addSpatialResolution()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOIdentification19115_3" data-id="addStatus"><a href='../../geometa/html/ISOIdentification19115_3.html#method-ISOIdentification19115_3-addStatus'><code>geometa::ISOIdentification19115_3$addStatus()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOIdentification19115_3" data-id="addTemporalResolution"><a href='../../geometa/html/ISOIdentification19115_3.html#method-ISOIdentification19115_3-addTemporalResolution'><code>geometa::ISOIdentification19115_3$addTemporalResolution()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOIdentification19115_3" data-id="addTopicCategory"><a href='../../geometa/html/ISOIdentification19115_3.html#method-ISOIdentification19115_3-addTopicCategory'><code>geometa::ISOIdentification19115_3$addTopicCategory()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOIdentification19115_3" data-id="delAdditionalDocumentation"><a href='../../geometa/html/ISOIdentification19115_3.html#method-ISOIdentification19115_3-delAdditionalDocumentation'><code>geometa::ISOIdentification19115_3$delAdditionalDocumentation()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOIdentification19115_3" data-id="delAssociatedResource"><a href='../../geometa/html/ISOIdentification19115_3.html#method-ISOIdentification19115_3-delAssociatedResource'><code>geometa::ISOIdentification19115_3$delAssociatedResource()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOIdentification19115_3" data-id="delCredit"><a href='../../geometa/html/ISOIdentification19115_3.html#method-ISOIdentification19115_3-delCredit'><code>geometa::ISOIdentification19115_3$delCredit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOIdentification19115_3" data-id="delExtent"><a href='../../geometa/html/ISOIdentification19115_3.html#method-ISOIdentification19115_3-delExtent'><code>geometa::ISOIdentification19115_3$delExtent()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOIdentification19115_3" data-id="delFormat"><a href='../../geometa/html/ISOIdentification19115_3.html#method-ISOIdentification19115_3-delFormat'><code>geometa::ISOIdentification19115_3$delFormat()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOIdentification19115_3" data-id="delGraphicOverview"><a href='../../geometa/html/ISOIdentification19115_3.html#method-ISOIdentification19115_3-delGraphicOverview'><code>geometa::ISOIdentification19115_3$delGraphicOverview()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOIdentification19115_3" data-id="delKeywords"><a href='../../geometa/html/ISOIdentification19115_3.html#method-ISOIdentification19115_3-delKeywords'><code>geometa::ISOIdentification19115_3$delKeywords()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOIdentification19115_3" data-id="delPointOfContact"><a href='../../geometa/html/ISOIdentification19115_3.html#method-ISOIdentification19115_3-delPointOfContact'><code>geometa::ISOIdentification19115_3$delPointOfContact()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOIdentification19115_3" data-id="delResourceConstraints"><a href='../../geometa/html/ISOIdentification19115_3.html#method-ISOIdentification19115_3-delResourceConstraints'><code>geometa::ISOIdentification19115_3$delResourceConstraints()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOIdentification19115_3" data-id="delResourceMaintenance"><a href='../../geometa/html/ISOIdentification19115_3.html#method-ISOIdentification19115_3-delResourceMaintenance'><code>geometa::ISOIdentification19115_3$delResourceMaintenance()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOIdentification19115_3" data-id="delResourceSpecificUsage"><a href='../../geometa/html/ISOIdentification19115_3.html#method-ISOIdentification19115_3-delResourceSpecificUsage'><code>geometa::ISOIdentification19115_3$delResourceSpecificUsage()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOIdentification19115_3" data-id="delSpatialRepresentationType"><a href='../../geometa/html/ISOIdentification19115_3.html#method-ISOIdentification19115_3-delSpatialRepresentationType'><code>geometa::ISOIdentification19115_3$delSpatialRepresentationType()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOIdentification19115_3" data-id="delSpatialResolution"><a href='../../geometa/html/ISOIdentification19115_3.html#method-ISOIdentification19115_3-delSpatialResolution'><code>geometa::ISOIdentification19115_3$delSpatialResolution()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOIdentification19115_3" data-id="delStatus"><a href='../../geometa/html/ISOIdentification19115_3.html#method-ISOIdentification19115_3-delStatus'><code>geometa::ISOIdentification19115_3$delStatus()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOIdentification19115_3" data-id="delTemporalResolution"><a href='../../geometa/html/ISOIdentification19115_3.html#method-ISOIdentification19115_3-delTemporalResolution'><code>geometa::ISOIdentification19115_3$delTemporalResolution()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOIdentification19115_3" data-id="delTopicCategory"><a href='../../geometa/html/ISOIdentification19115_3.html#method-ISOIdentification19115_3-delTopicCategory'><code>geometa::ISOIdentification19115_3$delTopicCategory()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOIdentification19115_3" data-id="setAbstract"><a href='../../geometa/html/ISOIdentification19115_3.html#method-ISOIdentification19115_3-setAbstract'><code>geometa::ISOIdentification19115_3$setAbstract()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOIdentification19115_3" data-id="setCitation"><a href='../../geometa/html/ISOIdentification19115_3.html#method-ISOIdentification19115_3-setCitation'><code>geometa::ISOIdentification19115_3$setCitation()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOIdentification19115_3" data-id="setProcessingLevel"><a href='../../geometa/html/ISOIdentification19115_3.html#method-ISOIdentification19115_3-setProcessingLevel'><code>geometa::ISOIdentification19115_3$setProcessingLevel()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOIdentification19115_3" data-id="setPurpose"><a href='../../geometa/html/ISOIdentification19115_3.html#method-ISOIdentification19115_3-setPurpose'><code>geometa::ISOIdentification19115_3$setPurpose()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ISODataIdentification19115_3-new"></a>}}
\if{latex}{\out{\hypertarget{method-ISODataIdentification19115_3-new}{}}}
\subsection{Method \code{new()}}{
Initializes object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ISODataIdentification19115_3$new(xml = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xml}}{object of class \link{XMLInternalNode-class}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ISODataIdentification19115_3-setDefaultLocale"></a>}}
\if{latex}{\out{\hypertarget{method-ISODataIdentification19115_3-setDefaultLocale}{}}}
\subsection{Method \code{setDefaultLocale()}}{
Set default locale
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ISODataIdentification19115_3$setDefaultLocale(locale)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{locale}}{object of class \link{ISOLocale}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ISODataIdentification19115_3-addOtherLocale"></a>}}
\if{latex}{\out{\hypertarget{method-ISODataIdentification19115_3-addOtherLocale}{}}}
\subsection{Method \code{addOtherLocale()}}{
Adds locale
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ISODataIdentification19115_3$addOtherLocale(locale)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{locale}}{object of class \link{ISOLocale}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{TRUE} if added, \code{FALSE} otherwise
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ISODataIdentification19115_3-delOtherLocale"></a>}}
\if{latex}{\out{\hypertarget{method-ISODataIdentification19115_3-delOtherLocale}{}}}
\subsection{Method \code{delOtherLocale()}}{
Deletes locale
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ISODataIdentification19115_3$delOtherLocale(locale)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{locale}}{object of class \link{ISOLocale}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{TRUE} if deleted, \code{FALSE} otherwise
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ISODataIdentification19115_3-setEnvironmentDescription"></a>}}
\if{latex}{\out{\hypertarget{method-ISODataIdentification19115_3-setEnvironmentDescription}{}}}
\subsection{Method \code{setEnvironmentDescription()}}{
Set environment description
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ISODataIdentification19115_3$setEnvironmentDescription(
  environmentDescription,
  locales = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{environmentDescription}}{environment description}

\item{\code{locales}}{a list of localized information. Default is \code{NULL}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ISODataIdentification19115_3-setSupplementalInformation"></a>}}
\if{latex}{\out{\hypertarget{method-ISODataIdentification19115_3-setSupplementalInformation}{}}}
\subsection{Method \code{setSupplementalInformation()}}{
Set supplemental information
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ISODataIdentification19115_3$setSupplementalInformation(
  supplementalInformation,
  locales = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{supplementalInformation}}{supplemental information}

\item{\code{locales}}{a list of localized information. Default is \code{NULL}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ISODataIdentification19115_3-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ISODataIdentification19115_3-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ISODataIdentification19115_3$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
