% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISODataQuality.R
\docType{class}
\name{ISODataQuality}
\alias{ISODataQuality}
\title{ISODataQuality}
\format{\code{\link{R6Class}} object.}
\usage{
ISODataQuality
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO DataQuality
}
\description{
ISODataQuality
}
\section{Fields}{

\describe{
\item{\code{scope}}{}

\item{\code{report}}{}

\item{\code{lineage}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(value)}}{
   This method is used to instantiate an ISODataQuality
 }
 \item{\code{setScope(scope)}}{
   Sets the scope
 }
 \item{\code{addReport(report)}}{
   Adds a report
 }
 \item{\code{setLineage(lineage)}}{
   Sets the lineage
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{data}
\keyword{quality}

