% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISODatatype.R
\docType{class}
\name{ISODatatype}
\alias{ISODatatype}
\title{ISODatatype}
\format{\code{\link{R6Class}} object.}
\usage{
ISODatatype
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO Datatype
}
\description{
ISODatatype
}
\section{Fields}{

\describe{
\item{\code{value}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(xml,value, description)}}{
   This method is used to instantiate an ISODatatype
 }
}
}
\examples{
#possible values
  values <- ISODatatype$values(labels = TRUE)

  #string Datatype
  stringType <- ISODatatype$new(value = "characterString")
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\references{
ISO 19115:2003 - Geographic information -- Metadata
}
\keyword{Datatype}
\keyword{ISO}

