% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISODimensionNameType.R
\docType{class}
\name{ISODimensionNameType}
\alias{ISODimensionNameType}
\title{ISODimensionNameType}
\format{\code{\link{R6Class}} object.}
\usage{
ISODimensionNameType
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO DimensionNameType
}
\description{
ISODimensionNameType
}
\section{Methods}{

\describe{
 \item{\code{new(xml,value, description)}}{
   This method is used to instantiate an \code{\link{ISODimensionNameType}}
 }
}
}

\examples{
  #possible values
  values <- ISODimensionNameType$values(labels = TRUE)
  
  #row DimensionNameType
  rowType <- ISODimensionNameType$new(value = "row")
  
}
\references{
ISO 19115:2003 - Geographic information -- Metadata
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{DimensionNameType}
\keyword{ISO}
