% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISODistributionUnits.R
\docType{class}
\name{ISODistributionUnits}
\alias{ISODistributionUnits}
\title{ISODistributionUnits}
\format{\code{\link{R6Class}} object.}
\usage{
ISODistributionUnits
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO DistributionUnits
}
\description{
ISODistributionUnits
}
\section{Fields}{

\describe{
\item{\code{value}}{}
}}

\section{Methods}{

\describe{
 \item{\code{new(xml,value, description)}}{
   This method is used to instantiate an ISODistributionUnits
 }
}
}

\examples{
  unit <- ISODistributionUnits$new(value = "unit")
  
}
\references{
ISO 19115:2003 - Geographic information -- Metadata
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{distribution}
\keyword{units}
