% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOElementSequence.R
\docType{class}
\name{ISOElementSequence}
\alias{ISOElementSequence}
\title{ISOElementSequence}
\format{\code{\link{R6Class}} object.}
\usage{
ISOElementSequence
}
\value{
Object of \code{\link{R6Class}} for modelling an ISOElementSequence
}
\description{
ISOElementSequence
}
\section{Fields}{

\describe{
\item{\code{aName}}{}

\item{\code{attributeType}}{}
}}
\note{
This class is used internally by geometa to deal with simple type not
handled by proper class element. e.g. \code{name} property of \code{ISOParameter}
class from ISO 19119:2005
}
\section{Methods}{

\describe{
 \item{\code{new(xml, ...)}}{
   This method is used to instantiate an element sequence
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\references{
ISO/TS 19103:2005 Geographic information -- Conceptual schema language
}
\keyword{ISO}
\keyword{record}

