% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOEvaluationMethodType.R
\docType{class}
\name{ISOEvaluationMethodType}
\alias{ISOEvaluationMethodType}
\title{ISOEvaluationMethodType}
\format{\code{\link{R6Class}} object.}
\usage{
ISOEvaluationMethodType
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO EvaluationMethodType
}
\description{
ISOEvaluationMethodType
}
\section{Fields}{

\describe{
\item{\code{value}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(xml,value)}}{
   This method is used to instantiate an ISOEvaluationMethodType
 }
}
}
\examples{
#possible values
  values <- ISOEvaluationMethodType$values(labels = TRUE)

  #example of EvaluationMethodType
  indirect <- ISOEvaluationMethodType$new(value = "indirect")
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\references{
ISO 19115:2003 - Geographic information -- Metadata
}
\keyword{EvaluationMethodType}
\keyword{ISO}

