% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOFeatureAssociation.R
\docType{class}
\name{ISOFeatureAssociation}
\alias{ISOFeatureAssociation}
\title{ISOFeatureAssociation}
\format{\code{\link{R6Class}} object.}
\usage{
ISOFeatureAssociation
}
\value{
Object of \code{\link{R6Class}} for modelling an ISOFeatureAssociation
}
\description{
ISOFeatureAssociation
}
\section{Fields}{

\describe{
\item{\code{roleName}}{[\code{\link{ISOAssociationRole}}] association role(s)}
}}

\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to instantiate an ISOFeatureAssociation
 }
 \item{\code{addRoleName(associationRole)}}{
   Adds an association role, object of class \code{ISOAssociationRole}
 }
 \item{\code{delRoleName(associationRole)}}{
   Deletes an association role, object of class \code{ISOAssociationRole}
 }
}
}

\references{
ISO 19110:2005 Methodology for Feature cataloguing
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{association}
\keyword{feature}
