% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOFeatureAssociation.R
\docType{class}
\name{ISOFeatureAssociation19139}
\alias{ISOFeatureAssociation19139}
\title{ISOFeatureAssociation19139}
\format{
\code{\link[R6]{R6Class}} object.
}
\value{
Object of \code{\link[R6]{R6Class}} for modelling an ISOFeatureAssociation in ISO 19139
}
\description{
ISOFeatureAssociation19139

ISOFeatureAssociation19139
}
\references{
ISO 19110:2005 Methodology for Feature cataloguing
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{association}
\keyword{feature}
\section{Super classes}{
\code{\link[geometa:geometaLogger]{geometa::geometaLogger}} -> \code{\link[geometa:ISOAbstractObject]{geometa::ISOAbstractObject}} -> \code{\link[geometa:ISOFeatureType]{geometa::ISOFeatureType}} -> \code{\link[geometa:ISOFeatureType19139]{geometa::ISOFeatureType19139}} -> \code{ISOFeatureAssociation19139}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{roleName}}{roleName [2..*]: ISOAssociationRole}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ISOFeatureAssociation19139-new}{\code{ISOFeatureAssociation19139$new()}}
\item \href{#method-ISOFeatureAssociation19139-addRoleName}{\code{ISOFeatureAssociation19139$addRoleName()}}
\item \href{#method-ISOFeatureAssociation19139-delRoleName}{\code{ISOFeatureAssociation19139$delRoleName()}}
\item \href{#method-ISOFeatureAssociation19139-clone}{\code{ISOFeatureAssociation19139$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="geometa" data-topic="geometaLogger" data-id="ERROR"><a href='../../geometa/html/geometaLogger.html#method-geometaLogger-ERROR'><code>geometa::geometaLogger$ERROR()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="geometaLogger" data-id="INFO"><a href='../../geometa/html/geometaLogger.html#method-geometaLogger-INFO'><code>geometa::geometaLogger$INFO()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="geometaLogger" data-id="WARN"><a href='../../geometa/html/geometaLogger.html#method-geometaLogger-WARN'><code>geometa::geometaLogger$WARN()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="addFieldAttrs"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-addFieldAttrs'><code>geometa::ISOAbstractObject$addFieldAttrs()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="addListElement"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-addListElement'><code>geometa::ISOAbstractObject$addListElement()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="checkMetadataStandardCompliance"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-checkMetadataStandardCompliance'><code>geometa::ISOAbstractObject$checkMetadataStandardCompliance()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="contains"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-contains'><code>geometa::ISOAbstractObject$contains()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="createLocalisedProperty"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-createLocalisedProperty'><code>geometa::ISOAbstractObject$createLocalisedProperty()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="decode"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-decode'><code>geometa::ISOAbstractObject$decode()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="delListElement"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-delListElement'><code>geometa::ISOAbstractObject$delListElement()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="encode"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-encode'><code>geometa::ISOAbstractObject$encode()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="getClass"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-getClass'><code>geometa::ISOAbstractObject$getClass()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="getClassName"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-getClassName'><code>geometa::ISOAbstractObject$getClassName()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="getNamespaceDefinition"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-getNamespaceDefinition'><code>geometa::ISOAbstractObject$getNamespaceDefinition()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="isDocument"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-isDocument'><code>geometa::ISOAbstractObject$isDocument()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="isFieldInheritedFrom"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-isFieldInheritedFrom'><code>geometa::ISOAbstractObject$isFieldInheritedFrom()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="print"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-print'><code>geometa::ISOAbstractObject$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="save"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-save'><code>geometa::ISOAbstractObject$save()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="setAttr"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-setAttr'><code>geometa::ISOAbstractObject$setAttr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="setCodeList"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-setCodeList'><code>geometa::ISOAbstractObject$setCodeList()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="setCodeListValue"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-setCodeListValue'><code>geometa::ISOAbstractObject$setCodeListValue()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="setCodeSpace"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-setCodeSpace'><code>geometa::ISOAbstractObject$setCodeSpace()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="setHref"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-setHref'><code>geometa::ISOAbstractObject$setHref()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="setId"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-setId'><code>geometa::ISOAbstractObject$setId()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="setIsNull"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-setIsNull'><code>geometa::ISOAbstractObject$setIsNull()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="setValue"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-setValue'><code>geometa::ISOAbstractObject$setValue()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="stopIfMetadataStandardIsNot"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-stopIfMetadataStandardIsNot'><code>geometa::ISOAbstractObject$stopIfMetadataStandardIsNot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="validate"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-validate'><code>geometa::ISOAbstractObject$validate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOAbstractObject" data-id="wrapBaseElement"><a href='../../geometa/html/ISOAbstractObject.html#method-ISOAbstractObject-wrapBaseElement'><code>geometa::ISOAbstractObject$wrapBaseElement()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOFeatureType19139" data-id="addAlias"><a href='../../geometa/html/ISOFeatureType19139.html#method-ISOFeatureType19139-addAlias'><code>geometa::ISOFeatureType19139$addAlias()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOFeatureType19139" data-id="addCharacteristic"><a href='../../geometa/html/ISOFeatureType19139.html#method-ISOFeatureType19139-addCharacteristic'><code>geometa::ISOFeatureType19139$addCharacteristic()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOFeatureType19139" data-id="addConstraint"><a href='../../geometa/html/ISOFeatureType19139.html#method-ISOFeatureType19139-addConstraint'><code>geometa::ISOFeatureType19139$addConstraint()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOFeatureType19139" data-id="addInheritsFrom"><a href='../../geometa/html/ISOFeatureType19139.html#method-ISOFeatureType19139-addInheritsFrom'><code>geometa::ISOFeatureType19139$addInheritsFrom()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOFeatureType19139" data-id="addInheritsTo"><a href='../../geometa/html/ISOFeatureType19139.html#method-ISOFeatureType19139-addInheritsTo'><code>geometa::ISOFeatureType19139$addInheritsTo()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOFeatureType19139" data-id="delAlias"><a href='../../geometa/html/ISOFeatureType19139.html#method-ISOFeatureType19139-delAlias'><code>geometa::ISOFeatureType19139$delAlias()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOFeatureType19139" data-id="delCharacteristic"><a href='../../geometa/html/ISOFeatureType19139.html#method-ISOFeatureType19139-delCharacteristic'><code>geometa::ISOFeatureType19139$delCharacteristic()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOFeatureType19139" data-id="delConstraint"><a href='../../geometa/html/ISOFeatureType19139.html#method-ISOFeatureType19139-delConstraint'><code>geometa::ISOFeatureType19139$delConstraint()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOFeatureType19139" data-id="delInheritsFrom"><a href='../../geometa/html/ISOFeatureType19139.html#method-ISOFeatureType19139-delInheritsFrom'><code>geometa::ISOFeatureType19139$delInheritsFrom()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOFeatureType19139" data-id="delInheritsTo"><a href='../../geometa/html/ISOFeatureType19139.html#method-ISOFeatureType19139-delInheritsTo'><code>geometa::ISOFeatureType19139$delInheritsTo()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOFeatureType19139" data-id="setCode"><a href='../../geometa/html/ISOFeatureType19139.html#method-ISOFeatureType19139-setCode'><code>geometa::ISOFeatureType19139$setCode()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOFeatureType19139" data-id="setDefinition"><a href='../../geometa/html/ISOFeatureType19139.html#method-ISOFeatureType19139-setDefinition'><code>geometa::ISOFeatureType19139$setDefinition()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOFeatureType19139" data-id="setDefinitionReference"><a href='../../geometa/html/ISOFeatureType19139.html#method-ISOFeatureType19139-setDefinitionReference'><code>geometa::ISOFeatureType19139$setDefinitionReference()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOFeatureType19139" data-id="setFeatureCatalogue"><a href='../../geometa/html/ISOFeatureType19139.html#method-ISOFeatureType19139-setFeatureCatalogue'><code>geometa::ISOFeatureType19139$setFeatureCatalogue()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOFeatureType19139" data-id="setIsAbstract"><a href='../../geometa/html/ISOFeatureType19139.html#method-ISOFeatureType19139-setIsAbstract'><code>geometa::ISOFeatureType19139$setIsAbstract()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geometa" data-topic="ISOFeatureType19139" data-id="setTypeName"><a href='../../geometa/html/ISOFeatureType19139.html#method-ISOFeatureType19139-setTypeName'><code>geometa::ISOFeatureType19139$setTypeName()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ISOFeatureAssociation19139-new"></a>}}
\if{latex}{\out{\hypertarget{method-ISOFeatureAssociation19139-new}{}}}
\subsection{Method \code{new()}}{
Initializes object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ISOFeatureAssociation19139$new(xml = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xml}}{object of class \link[XML]{XMLInternalNode-class}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ISOFeatureAssociation19139-addRoleName"></a>}}
\if{latex}{\out{\hypertarget{method-ISOFeatureAssociation19139-addRoleName}{}}}
\subsection{Method \code{addRoleName()}}{
Adds role name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ISOFeatureAssociation19139$addRoleName(associationRole)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{associationRole}}{object of class \link{ISOAssociationRole}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{TRUE} if added, \code{FALSE} otherwise
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ISOFeatureAssociation19139-delRoleName"></a>}}
\if{latex}{\out{\hypertarget{method-ISOFeatureAssociation19139-delRoleName}{}}}
\subsection{Method \code{delRoleName()}}{
Deletes role name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ISOFeatureAssociation19139$delRoleName(associationRole)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{associationRole}}{object of class \link{ISOAssociationRole}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{TRUE} if deleted, \code{FALSE} otherwise
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ISOFeatureAssociation19139-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ISOFeatureAssociation19139-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ISOFeatureAssociation19139$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
