% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOFileName.R
\docType{class}
\name{ISOFileName}
\alias{ISOFileName}
\title{ISOFileName}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO FileName
}
\description{
ISOFileName
}
\section{Methods}{

\describe{
 \item{\code{new(xml, file, name)}}{
   This method is used to instantiate an \code{\link{ISOFileName}}
 }
}
}

\examples{
  md <- ISOFileName$new(file = "someuri", name = "filename")
  xml <- md$encode()

}
\references{
ISO/TS 19139:2007 Geographic information -- XML
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{file}
\keyword{name}
