% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOGeometricObjectType.R
\docType{class}
\name{ISOGeometricObjectType}
\alias{ISOGeometricObjectType}
\title{ISOGeometricObjectType}
\format{\code{\link{R6Class}} object.}
\usage{
ISOGeometricObjectType
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO GeometricObjectType
}
\description{
ISOGeometricObjectType
}
\section{Fields}{

\describe{
\item{\code{value}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(xml,value)}}{
   This method is used to instantiate an ISOGeometricObjectType
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{geometric}
\keyword{object}
\keyword{type}

