% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOIdentifier.R
\docType{class}
\name{ISOIdentifier}
\alias{ISOIdentifier}
\title{ISOIdentifier}
\format{\code{\link{R6Class}} object.}
\usage{
ISOIdentifier
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO Identifier
}
\description{
ISOIdentifier
}
\section{Fields}{

\describe{
\item{\code{authority}}{[\code{\link{ISOCitation}}] authority}

\item{\code{code}}{[\code{\link{character}}] code}
}}

\note{
Abstract ISO class
}
\section{Methods}{

\describe{
 \item{\code{new(xml, code, codeSpace)}}{
   This method is used to instantiate an \code{\link{ISOIdentifier}}
 }
 \item{\code{setAuthority(authority)}}{
   Sets an authority object of class \code{\link{ISOCitation}}
 }
}
}

\references{
ISO 19115:2003 - Geographic information -- Metadata
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{identifier}
