% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOImageryGeolocationInformation.R
\docType{class}
\name{ISOImageryAbstractGeolocationInformation}
\alias{ISOImageryAbstractGeolocationInformation}
\title{ISOImageryAbstractGeolocationInformation}
\format{\code{\link{R6Class}} object.}
\usage{
ISOImageryAbstractGeolocationInformation
}
\value{
Object of \code{\link{R6Class}} for modelling an ISOimagery geolocation information
}
\description{
ISOImageryAbstractGeolocationInformation
}
\note{
abstract class
}
\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to instantiate an \code{\link{ISOImageryAbstractGeolocationInformation}}
 }
}
}

\references{
ISO 19115-2:2009 - Geographic information -- Metadata Part 2: Extensions for imagery and gridded data
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{abstract}
\keyword{geolocation}
\keyword{imagery}
\keyword{information}
