% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOImageryBandDefinition.R
\docType{class}
\name{ISOImageryBandDefinition}
\alias{ISOImageryBandDefinition}
\title{ISOImageryBandDefinition}
\format{\code{\link{R6Class}} object.}
\usage{
ISOImageryBandDefinition
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO Imagery Band definition
}
\description{
ISOImageryBandDefinition
}
\section{Methods}{

\describe{
 \item{\code{new(xml,value)}}{
   This method is used to instantiate an \code{\link{ISOImageryBandDefinition}}
 }
}
}

\examples{
  #possible values
  values <- ISOImageryBandDefinition$values(labels = TRUE)
  
  #some def
  fiftyp <- ISOImageryBandDefinition$new(value = "fiftyPercent")

}
\references{
ISO 19115-2:2009 - Geographic information -- Metadata Part 2: Extensions for imagery and gridded data
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{band}
\keyword{definition}
\keyword{imagery}
