% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOImageryContext.R
\docType{class}
\name{ISOImageryContext}
\alias{ISOImageryContext}
\title{ISOImageryContext}
\format{\code{\link{R6Class}} object.}
\usage{
ISOImageryContext
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO Imagery Context
}
\description{
ISOImageryContext
}
\section{Methods}{

\describe{
 \item{\code{new(xml,value)}}{
   This method is used to instantiate an \code{\link{ISOImageryContext}}
 }
}
}

\examples{
  #possible values
  values <- ISOImageryContext$values(labels = TRUE)
  
  #some def
  acquisition <- ISOImageryContext$new(value = "acquisition")

}
\references{
ISO 19115-2:2009 - Geographic information -- Metadata Part 2: Extensions for imagery and gridded data
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{context}
\keyword{imagery}
