% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOImageryCoverageResult.R
\docType{class}
\name{ISOImageryCoverageResult}
\alias{ISOImageryCoverageResult}
\title{ISOImageryCoverageResult}
\format{\code{\link{R6Class}} object.}
\usage{
ISOImageryCoverageResult
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO imagery coverage result
}
\description{
ISOImageryCoverageResult
}
\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to instantiate an \code{\link{ISOImageryCoverageResult}}
 }
}

 @section Methods inherited from \code{\link{ISOAbstractResult}}:
\describe{
 See methods description at \code{\link{ISOAbstractResult}}
}
}

\references{
ISO 19115-2:2009 - Geographic information -- Metadata Part 2: Extensions for imagery and gridded data
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{coverage}
\keyword{imagery}
\keyword{result}
