% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOImageryGeometryType.R
\docType{class}
\name{ISOImageryGeometryType}
\alias{ISOImageryGeometryType}
\title{ISOImageryGeometryType}
\format{\code{\link{R6Class}} object.}
\usage{
ISOImageryGeometryType
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO Imagery geometry type
}
\description{
ISOImageryGeometryType
}
\section{Methods}{

\describe{
 \item{\code{new(xml,value)}}{
   This method is used to instantiate an \code{\link{ISOImageryGeometryType}}
 }
}
}

\examples{
  #possible values
  values <- ISOImageryGeometryType$values(labels = TRUE)
  
  #some def
  point <- ISOImageryGeometryType$new(value = "point")

}
\references{
ISO 19115-2:2009 - Geographic information -- Metadata Part 2: Extensions for imagery and gridded data
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{geometry}
\keyword{imagery}
\keyword{type}
