% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOImageryGeorectified.R
\docType{class}
\name{ISOImageryGeorectified}
\alias{ISOImageryGeorectified}
\title{ISOImageryGeorectified}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO image Georectified
}
\description{
ISOImageryGeorectified
}
\section{Fields}{

\describe{
\item{\code{checkPoint}}{[\code{list} of \code{\link{ISOImageryGCP}}]}
}}

\section{Methods}{

\describe{
 \item{\code{new(xml,value)}}{
   This method is used to instantiate an \code{\link{ISOImageryGeorectified}}
 }
 \item{\code{addCheckPoint(sfg,m)}}{
   Adds a check point, either an object of class 'sfg' (from \pkg{sf}) or a 'matrix'
 }
 \item{\code{delCheckPoint(sfg,m)}}{
   Deletes a check point, either an object of class 'sfg' (from \pkg{sf}) or a 'matrix'
 }
}
}

\section{Methods inherited from \code{\link{ISOGridSpatialRepresentation}}}{

See \code{\link{ISOGridSpatialRepresentation}}
}

\section{Methods inherited from \code{\link{ISOGeorectified}}}{

See \code{\link{ISOGeorectified}}
}

\references{
ISO 19115-2:2009 - Geographic information -- Metadata -- Part 2: Extensions for imagery and gridded data
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{georectified}
\keyword{image}
