% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOImageryObjective.R
\docType{class}
\name{ISOImageryObjectiveType}
\alias{ISOImageryObjectiveType}
\title{ISOImageryObjectiveType}
\format{\code{\link{R6Class}} object.}
\usage{
ISOImageryObjectiveType
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO imagery ObjectiveType
}
\description{
ISOImageryObjectiveType
}
\section{Methods}{

\describe{
 \item{\code{new(xml,value)}}{
   This method is used to instantiate an \code{\link{ISOImageryObjectiveType}}
 }
}
}

\examples{
  #possible values
  values <- ISOImageryObjectiveType$values(labels = TRUE)
  
  #some def
  survey <- ISOImageryObjectiveType$new(value = "survey")

}
\references{
ISO 19115-2:2009 - Geographic information -- Metadata Part 2: Extensions for imagery and gridded data
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{ObjectiveType}
\keyword{imagery}
