% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOImageryOperationType.R
\docType{class}
\name{ISOImageryOperationType}
\alias{ISOImageryOperationType}
\title{ISOImageryOperationType}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO Imagery Operation type
}
\description{
ISOImageryOperationType
}
\section{Methods}{

\describe{
 \item{\code{new(xml,value)}}{
   This method is used to instantiate an \code{\link{ISOImageryOperationType}}
 }
}
}

\examples{
  #possible values
  values <- ISOImageryOperationType$values(labels = TRUE)
  
  #some def
  real <- ISOImageryOperationType$new(value = "real")

}
\references{
ISO 19115-2:2009 - Geographic information -- Metadata Part 2: Extensions for imagery and gridded data
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{Operation}
\keyword{imagery}
\keyword{type}
