% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOImageryPlatform.R
\docType{class}
\name{ISOImageryPlatform}
\alias{ISOImageryPlatform}
\title{ISOImageryPlatform}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO imagery platform
}
\description{
ISOImageryPlatform
}
\section{Fields}{

\describe{
\item{\code{citation}}{[\code{list} of \code{\link{ISOCitation}}]}

\item{\code{identifier}}{[\code{\link{ISOMetaIdentifier}}]}

\item{\code{description}}{[\code{\link{character}}|\code{\link{ISOLocalisedCharacterString}}]}

\item{\code{sponsor}}{[\code{list} of \code{\link{ISOResponsibleParty}}]}

\item{\code{instrument}}{[\code{list} of \code{\link{ISOImageryInstrument}}]}
}}

\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to instantiate an \code{\link{ISOImageryPlatform}}
 }
 \item{\code{addCitation(citation)}}{
   Adds citation, object of class \code{\link{ISOCitation}}
 }
 \item{\code{delCitation(citation)}}{
   Deletes a citation, object of class \code{\link{ISOCitation}}
 }
 \item{\code{setIdentifier(identifier)}}{
   Sets an identifier, object of class \code{character} or \code{\link{ISOMetaIdentifier}}
 }
 \item{\code{setDescription(description, locales)}}{
   Sets a description (object of class "character"). Locale names can be 
   specified as \code{list} with the \code{locales} argument.
 }
 \item{\code{addSponsor(sponsor)}}{
   Adds a sponsor, object of class \code{\link{ISOResponsibleParty}}
 }
 \item{\code{delSponsor(sponsor)}}{
   Deletes a sponsor, object of class \code{\link{ISOResponsibleParty}}
 }
 \item{\code{addInstrument(instrument)}}{
   Add a instrument, object of class \code{\link{ISOImageryInstrument}}
 }
 \item{\code{delInstrument(instrument)}}{
   Deletes a instrument, object of class \code{\link{ISOImageryInstrument}}
 }
}
}

\examples{
   md <- ISOImageryPlatform$new()
   
   #add citation
   rp1 <- ISOResponsibleParty$new()
   rp1$setIndividualName("someone1")
   rp1$setOrganisationName("somewhere1")
   rp1$setPositionName("someposition1")
   rp1$setRole("pointOfContact")
   contact1 <- ISOContact$new()
   phone1 <- ISOTelephone$new()
   phone1$setVoice("myphonenumber1")
   phone1$setFacsimile("myfacsimile1")
   contact1$setPhone(phone1)
   address1 <- ISOAddress$new()
   address1$setDeliveryPoint("theaddress1")
   address1$setCity("thecity1")
   address1$setPostalCode("111")
   address1$setCountry("France")
   address1$setEmail("someone1@theorg.org")
   contact1$setAddress(address1)
   res <- ISOOnlineResource$new()
   res$setLinkage("http://www.somewhereovertheweb.org")
   res$setName("somename")
   contact1$setOnlineResource(res)
   rp1$setContactInfo(contact1)
   
   #citation
   ct <- ISOCitation$new()
   ct$setTitle("sometitle")
   d <- ISODate$new()
   d$setDate(ISOdate(2015, 1, 1, 1))
   d$setDateType("publication")
   ct$addDate(d)
   ct$setEdition("1.0")
   ct$setEditionDate(ISOdate(2015,1,1))
   ct$addIdentifier(ISOMetaIdentifier$new(code = "identifier"))
   ct$addPresentationForm("mapDigital")
   ct$addCitedResponsibleParty(rp1)
   md$addCitation(ct)
   
   md$setIdentifier("identifier")
   md$setDescription("some description")
   
   xml <- md$encode()

}
\references{
ISO 19115-2:2009 - Geographic information -- Metadata Part 2: Extensions for imagery and gridded data
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{imagery}
\keyword{platform}
