% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOImageryPlatformPass.R
\docType{class}
\name{ISOImageryPlatformPass}
\alias{ISOImageryPlatformPass}
\title{ISOImageryPlatformPass}
\format{\code{\link{R6Class}} object.}
\usage{
ISOImageryPlatformPass
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO imagery PlatformPass
}
\description{
ISOImageryPlatformPass
}
\section{Fields}{

\describe{
\item{\code{identifier}}{[\code{\link{ISOMetaIdentifier}}]}

\item{\code{extent}}{[?]}

\item{\code{relatedEvent}}{[\code{list} of \code{\link{ISOImageryEvent}}]}
}}

\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to instantiate an \code{\link{ISOImageryPlatformPass}}
 }
 \item{\code{setIdentifier(identifier)}}{
   Sets an identifier, object of class \code{character} or \code{\link{ISOMetaIdentifier}}
 }
 \item{\code{setExtent(extent)}}{
   Set the extent
 }
 \item{\code{addEvent(event)}}{
   Add a event, object of class \code{\link{ISOImageryEvent}}
 }
 \item{\code{delEvent(event)}}{
   Deletes a event, object of class \code{\link{ISOImageryEvent}}
 }
}
}

\examples{
   md <- ISOImageryPlatformPass$new()
   
   xml <- md$encode()

}
\references{
ISO 19115-2:2009 - Geographic information -- Metadata Part 2: Extensions for imagery and gridded data
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{PlatformPass}
\keyword{imagery}
