% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOImageryPolarizationOrientation.R
\docType{class}
\name{ISOImageryPolarizationOrientation}
\alias{ISOImageryPolarizationOrientation}
\title{ISOImageryPolarizationOrientation}
\format{\code{\link{R6Class}} object.}
\usage{
ISOImageryPolarizationOrientation
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO Imagery polarization orientation
}
\description{
ISOImageryPolarizationOrientation
}
\section{Methods}{

\describe{
 \item{\code{new(xml,value)}}{
   This method is used to instantiate an \code{\link{ISOImageryPolarizationOrientation}}
 }
}
}

\examples{
  #possible values
  values <- ISOImageryPolarizationOrientation$values(labels = TRUE)
  
  #some def
  h <- ISOImageryPolarizationOrientation$new(value = "horizontal")

}
\references{
ISO 19115-2:2009 - Geographic information -- Metadata Part 2: Extensions for imagery and gridded data
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{imagery}
\keyword{orientation}
\keyword{polarization}
