% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOImageryPriority.R
\docType{class}
\name{ISOImageryPriority}
\alias{ISOImageryPriority}
\title{ISOImageryPriority}
\format{\code{\link{R6Class}} object.}
\usage{
ISOImageryPriority
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO imagery priority
}
\description{
ISOImageryPriority
}
\section{Methods}{

\describe{
 \item{\code{new(xml,value)}}{
   This method is used to instantiate an \code{\link{ISOImageryPriority}}
 }
}
}

\examples{
  #possible values
  values <- ISOImageryPriority$values(labels = TRUE)
  
  #some def
  highImp <- ISOImageryPriority$new(value = "highImportance")

}
\references{
ISO 19115-2:2009 - Geographic information -- Metadata Part 2: Extensions for imagery and gridded data
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{imagery}
\keyword{priority}
