% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOImagerySensorType.R
\docType{class}
\name{ISOImagerySensorType}
\alias{ISOImagerySensorType}
\title{ISOImagerySensorType}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO imagery sensor type
}
\description{
ISOImagerySensorType
}
\section{Methods}{

\describe{
 \item{\code{new(xml,value)}}{
   This method is used to instantiate an \code{\link{ISOImagerySensorType}}
 }
}
}

\examples{
  md <- ISOImagerySensorType$new()

}
\references{
ISO 19115-2:2009 - Geographic information -- Metadata Part 2: Extensions for imagery and gridded data
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{imagery}
\keyword{sensor}
\keyword{type}
