% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOImagerySequence.R
\docType{class}
\name{ISOImagerySequence}
\alias{ISOImagerySequence}
\title{ISOImagerySequence}
\format{\code{\link{R6Class}} object.}
\usage{
ISOImagerySequence
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO imagery sequence
}
\description{
ISOImagerySequence
}
\section{Methods}{

\describe{
 \item{\code{new(xml,value)}}{
   This method is used to instantiate an \code{\link{ISOImagerySequence}}
 }
}
}

\examples{
  #possible values
  values <- ISOImagerySequence$values(labels = TRUE)
  
  #some def
  inst <- ISOImagerySequence$new(value = "instantaneous")

}
\references{
ISO 19115-2:2009 - Geographic information -- Metadata Part 2: Extensions for imagery and gridded data
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{imagery}
\keyword{sequence}
