% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOImageryTransferFunctionType.R
\docType{class}
\name{ISOImageryTransferFunctionType}
\alias{ISOImageryTransferFunctionType}
\title{ISOImageryTransferFunctionType}
\format{\code{\link{R6Class}} object.}
\usage{
ISOImageryTransferFunctionType
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO imagery transfer function type
}
\description{
ISOImageryTransferFunctionType
}
\section{Methods}{

\describe{
 \item{\code{new(xml,value)}}{
   This method is used to instantiate an \code{\link{ISOImageryTransferFunctionType}}
 }
}
}

\examples{
  #possible values
  values <- ISOImageryTransferFunctionType$values(labels = TRUE)
  
  #some def
  log <- ISOImageryTransferFunctionType$new(value = "logarithmic")

}
\references{
ISO 19115-2:2009 - Geographic information -- Metadata Part 2: Extensions for imagery and gridded data
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{function}
\keyword{imagery}
\keyword{transfer}
\keyword{type}
