% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOImageryTrigger.R
\docType{class}
\name{ISOImageryTrigger}
\alias{ISOImageryTrigger}
\title{ISOImageryTrigger}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO imagery trigger
}
\description{
ISOImageryTrigger
}
\section{Methods}{

\describe{
 \item{\code{new(xml,value)}}{
   This method is used to instantiate an \code{\link{ISOImageryTrigger}}
 }
}
}

\examples{
  #possible values
  values <- ISOImageryTrigger$values(labels = TRUE)
  
  #some def
  auto <- ISOImageryTrigger$new(value = "automatic")

}
\references{
ISO 19115-2:2009 - Geographic information -- Metadata Part 2: Extensions for imagery and gridded data
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{imagery}
\keyword{trigger}
