% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOImageryUsability.R
\docType{class}
\name{ISOImageryUsability}
\alias{ISOImageryUsability}
\title{ISOImageryUsability}
\format{\code{\link{R6Class}} object.}
\usage{
ISOImageryUsability
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO imagery usability
}
\description{
ISOImageryUsability
}
\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to instantiate an \code{\link{ISOImageryUsability}}
 }
}

 @section Methods inherited from \code{\link{ISODataQualityAbstractElement}}:
\describe{
 See methods description at \code{\link{ISODataQualityAbstractElement}}
}
}

\references{
ISO 19115-2:2009 - Geographic information -- Metadata Part 2: Extensions for imagery and gridded data
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{imagery}
\keyword{usability}
