% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOImagingCondition.R
\docType{class}
\name{ISOImagingCondition}
\alias{ISOImagingCondition}
\title{ISOImagingCondition}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an ISOImagingCondition
}
\description{
ISOImagingCondition
}
\section{Methods}{

\describe{
 \item{\code{new(xml,value, description)}}{
   This method is used to instantiate an \code{\link{ISOImagingCondition}}
 }
}
}

\examples{
  #possible values
  values <- ISOImagingCondition$values(labels = TRUE)
  
  #ImagingCondition
  ImagingCondition <- ISOImagingCondition$new(value = "rain")

}
\references{
ISO 19115:2003 - Geographic information -- Metadata
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{condition}
\keyword{imaging}
