% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOInitiativeType.R
\docType{class}
\name{ISOInitiativeType}
\alias{ISOInitiativeType}
\title{ISOInitiativeType}
\format{\code{\link{R6Class}} object.}
\usage{
ISOInitiativeType
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO InitiativeType
}
\description{
ISOInitiativeType
}
\section{Methods}{

\describe{
 \item{\code{new(xml,value, description)}}{
   This method is used to instantiate an \code{\link{ISOInitiativeType}}
 }
}
}

\examples{
  #possible values
  values <- ISOInitiativeType$values(labels = TRUE)
  
  #geomOnly
  geomOnly <- ISOInitiativeType$new(value = "campaign")
  
}
\references{
ISO 19115:2003 - Geographic information -- Metadata
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{initative}
\keyword{type}
