% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOLanguage.R
\docType{class}
\name{ISOLanguage}
\alias{ISOLanguage}
\title{ISOLanguage}
\format{\code{\link{R6Class}} object.}
\usage{
ISOLanguage
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO Language
}
\description{
ISOLanguage
}
\section{Fields}{

\describe{
\item{\code{value}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(xml,value)}}{
   This method is used to instantiate an ISOLanguage
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{language}

