% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOLegalConstraints.R
\docType{class}
\name{ISOLegalConstraints}
\alias{ISOLegalConstraints}
\title{ISOLegalConstraints}
\format{\code{\link{R6Class}} object.}
\usage{
ISOLegalConstraints
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO LegalConstraints
}
\description{
ISOLegalConstraints
}
\section{Fields}{

\describe{
\item{\code{useLimitation}}{}

\item{\code{accessConstraints}}{}

\item{\code{useConstraints}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to instantiate an ISOLegalConstraints
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{constraints}
\keyword{legal}

