% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOLength.R
\docType{class}
\name{ISOLength}
\alias{ISOLength}
\title{ISOLength}
\format{\code{\link{R6Class}} object.}
\usage{
ISOLength
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO Length measure
}
\description{
ISOLength
}
\section{Fields}{

\describe{
\item{\code{value}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(xml,value, uom, useUomURI)}}{
   This method is used to instantiate an ISOLength. The \code{uom} argument represents
   the symbol of unit of measure used. The parameter  \code{useUomURI} can be used to
   set the uom as URI, its default value is \code{FALSE}.
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\references{
ISO/TS 19103:2005 Geographic information -- Conceptual schema language
}
\keyword{ISO}
\keyword{length}
\keyword{measure}

