% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOLineage.R
\docType{class}
\name{ISOLineage}
\alias{ISOLineage}
\title{ISOLineage}
\format{\code{\link{R6Class}} object.}
\usage{
ISOLineage
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO Lineage
}
\description{
ISOLineage
}
\section{Fields}{

\describe{
\item{\code{statement}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to instantiate an ISOLineage
 }
 \item{\code{setStatement(statement)}}{
   Sets the statement
 }
}
}
\examples{
lineage <- ISOLineage$new()
  lineage$setStatement("statement")
  xml <- lineage$encode()
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{lineage}

