% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOLocalisedCharacterString.R
\docType{class}
\name{ISOLocalisedCharacterString}
\alias{ISOLocalisedCharacterString}
\title{ISOLocalisedCharacterString}
\format{\code{\link{R6Class}} object.}
\usage{
ISOLocalisedCharacterString
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO LocalisedCharacterString
}
\description{
ISOLocalisedCharacterString
}
\section{Methods}{

\describe{
 \item{\code{new(xml, locale, value)}}{
   This method is used to instantiate an \code{\link{ISOLocalisedCharacterString}}
 }
}
}

\examples{
  str <- ISOLocalisedCharacterString$new(locale = "FR", value = "ma description")
  str$encode()

}
\references{
ISO/TS 19103:2005 Geographic information -- Conceptual schema language
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{character}
\keyword{localised}
\keyword{string}
