% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOLocalName.R
\docType{class}
\name{ISOLocalName}
\alias{ISOLocalName}
\title{ISOLocalName}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO LocalName
}
\description{
ISOLocalName
}
\section{Fields}{

\describe{
\item{\code{value}}{[\code{\link{character}}] local name}
}}

\section{Methods}{

\describe{
 \item{\code{new(xml,value)}}{
   This method is used to instantiate an ISOLocalName
 }
}
}

\references{
ISO/TS 19103:2005 Geographic information -- Conceptual schema language
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{local}
\keyword{name}
