% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOMaintenanceInformation.R
\docType{class}
\name{ISOMaintenanceInformation}
\alias{ISOMaintenanceInformation}
\title{ISOMaintenanceInformation}
\format{\code{\link{R6Class}} object.}
\usage{
ISOMaintenanceInformation
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO MaintenanceInformation
}
\description{
ISOMaintenanceInformation
}
\section{Fields}{

\describe{
\item{\code{maintenanceAndUpdateFrequency}}{[\code{\link{ISOMaintenanceFrequency}}|\code{\link{character}}] frequency}
}}

\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to instantiate an \code{\link{ISOMaintenanceInformation}}
 }
 \item{\code{setMaintenanceFrequency(frequency)}}{
   Sets the maintenance and update frequency
 }
}
}

\examples{
  md <- ISOMaintenanceInformation$new()
  md$setMaintenanceFrequency("daily")
  xml <- md$encode()
  
}
\references{
ISO 19115:2003 - Geographic information -- Metadata
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{information}
\keyword{maintenance}
