% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOMeasure.R
\docType{class}
\name{ISOMeasure}
\alias{ISOMeasure}
\title{ISOMeasure}
\format{\code{\link{R6Class}} object.}
\usage{
ISOMeasure
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO Measure
}
\description{
ISOMeasure
}
\section{Fields}{

\describe{
\item{\code{value}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(xml,value)}}{
   This method is used to instantiate an ISOMeasure
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\references{
ISO/TS 19103:2005 Geographic information -- Conceptual schema language
}
\keyword{ISO}
\keyword{measure}

