% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOMediumFormat.R
\docType{class}
\name{ISOMediumFormat}
\alias{ISOMediumFormat}
\title{ISOMediumFormat}
\format{\code{\link{R6Class}} object.}
\usage{
ISOMediumFormat
}
\value{
Object of \code{\link{R6Class}} for modelling an ISOMediumFormat
}
\description{
ISOMediumFormat
}
\section{Fields}{

\describe{
\item{\code{value}}{}
}}

\section{Methods}{

\describe{
 \item{\code{new(xml,value, description)}}{
   This method is used to instantiate an ISOMediumFormat
 }
}
}

\examples{
  #possible values
  values <- ISOMediumFormat$values(labels = TRUE)
  
  #MediumFormat
  MediumFormat <- ISOMediumFormat$new(value = "tar")

}
\references{
ISO 19115:2003 - Geographic information -- Metadata
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{format}
\keyword{medium}
