% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOMediumName.R
\docType{class}
\name{ISOMediumName}
\alias{ISOMediumName}
\title{ISOMediumName}
\format{\code{\link{R6Class}} object.}
\usage{
ISOMediumName
}
\value{
Object of \code{\link{R6Class}} for modelling an ISOMediumName
}
\description{
ISOMediumName
}
\section{Fields}{

\describe{
\item{\code{value}}{}
}}

\section{Methods}{

\describe{
 \item{\code{new(xml,value, description)}}{
   This method is used to instantiate an ISOMediumName
 }
}
}

\examples{
  #possible values
  values <- ISOMediumName$values(labels = TRUE)
  
  #MediumName
  MediumName <- ISOMediumName$new(value = "satellite")

}
\references{
ISO 19115:2003 - Geographic information -- Metadata
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{medium}
\keyword{name}
