% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOMemberName.R
\docType{class}
\name{ISOMemberName}
\alias{ISOMemberName}
\title{ISOMemberName}
\format{\code{\link{R6Class}} object.}
\usage{
ISOMemberName
}
\value{
Object of \code{\link{R6Class}} for modelling an ISOMemberName
}
\description{
ISOMemberName
}
\section{Fields}{

\describe{
\item{\code{aName}}{}

\item{\code{attributeType}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(xml, aName, attributeType)}}{
   This method is used to instantiate an ISOMemberName
 }
 \item{\code{setName(aName)}}{
   Set the aName, object of class \code{character}
 }
 \item{\code{setAttributeType(attributeType)}}{
   Set the attribute type, object of class \code{ISOTypeName} or \code{character}
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\references{
ISO/TS 19103:2005 Geographic information -- Conceptual schema language
}
\keyword{ISO}
\keyword{record}

