% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOMetadata.R
\docType{class}
\name{ISOMetadata}
\alias{ISOMetadata}
\title{ISOMetadataElement}
\format{\code{\link{R6Class}} object.}
\usage{
ISOMetadata
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO Metadata Element
}
\description{
ISOMetadataElement
}
\section{Fields}{

\describe{
\item{\code{fileIdentifier}}{}

\item{\code{anguage}}{}

\item{\code{characterSet}}{}

\item{\code{parentIdentifier}}{}

\item{\code{hierarchyLevel}}{}

\item{\code{contact}}{}

\item{\code{dateStamp}}{}

\item{\code{metadataStandardName}}{}

\item{\code{metadataStandardVersion}}{}

\item{\code{spatialRepresentationInfo}}{}

\item{\code{referenceSystemInfo}}{}

\item{\code{identificationInfo}}{}

\item{\code{distributionInfo}}{}

\item{\code{dataQualityInfo}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new()}}{
   This method is used to instantiate an ISOMetadata
 }
 \item{\code{setFileIdentifier(fileIdentifier)}}{
   Sets the file identifier
 }
 \item{\code{setParentIdentifier(parentIdentifier)}}{
   Sets the parentIdentifier
 }
 \item{\code{setLanguage{language}}}{
   Sets the language
 }
 \item{\code{setCharacterSet(charset)}}{
   Sets the character set
 }
 \item{\code{setHierarchyLevel(level)}}{
   Sets the hierarchy level
 }
 \item{\code{setDateStamp(date)}}{
   Sets the date stamp
 }
 \item{\code{setMetadataStandardName(name)}}{
   Sets the metadata standard name
 }
 \item{\code{setMetadataStandardVersion(version)}}{
   Sets the metadata standard version
 }
 \item{\code{addContact(contact)}}{
   Adds a contact (responsible party)
 }
 \item{\code{setIdentificationInfo(identificationInfo)}}{
   Sets the data identification
 }
 \item{\code{setDistributionInfo(distributionInfo)}}{
   Sets the distribution
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{element}
\keyword{metadata}

