% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOMetadataCodelistElement.R
\docType{class}
\name{ISOMetadataCodelistElement}
\alias{ISOMetadataCodelistElement}
\title{ISOMetadataCodelistElement}
\format{\code{\link{R6Class}} object.}
\usage{
ISOMetadataCodelistElement
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO Metadata codelist element
}
\description{
ISOMetadataCodelistElement
}
\section{Fields}{

\describe{
\item{\code{value}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(xml,value)}}{
   This method is used to instantiate an ISOMetadataCodelistElement
 }
 \item{\code{getAcceptedValues()}}{
   This method allows to get the codelist accepted values
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{code}
\keyword{element}

