% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOMetadataElement.R
\docType{class}
\name{ISOMetadataElement}
\alias{ISOMetadataElement}
\title{ISOMetadataElement}
\format{\code{\link{R6Class}} object.}
\usage{
ISOMetadataElement
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO Metadata Element
}
\description{
ISOMetadataElement
}
\section{Static Methods}{

\describe{
 \item{\code{getISOClassByNode(node)}}{
   Inherit the ISO class matching an XML document or node
 }
}
}

\section{Abstract Methods}{

\describe{
 \item{\code{new(xml, element, namespace)}}{
   This method is used to instantiate an ISOMetadataElement
 }
 \item{\code{decode(xml)}}{
   Decodes a ISOMetadata* R6 object from XML representation
 }
 \item{\code{encode()}}{
   Encodes a ISOMetadata* R6 object to XML representation
 }
 \item{\code{wrapBaseElement(field, fieldObj)}}{
   Wraps a base element type
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{element}
\keyword{metadata}

