% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOMetaIdentifier.R
\docType{class}
\name{ISOMetaIdentifier}
\alias{ISOMetaIdentifier}
\title{ISOMetaIdentifier}
\format{\code{\link{R6Class}} object.}
\usage{
ISOMetaIdentifier
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO MetaIdentifier
}
\description{
ISOMetaIdentifier
}
\section{Methods}{

\describe{
 \item{\code{new(xml, code)}}{
   This method is used to instantiate an ISOMetaIdentifier
 }
}
}

\examples{
  md <- ISOMetaIdentifier$new(code = "identifier")
  xml <- md$encode()
  
}
\references{
ISO 19115:2003 - Geographic information -- Metadata
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{identifier}
\keyword{meta}
