% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOMimeFileType.R
\docType{class}
\name{ISOMimeFileType}
\alias{ISOMimeFileType}
\title{ISOMimeFileType}
\format{\code{\link{R6Class}} object.}
\usage{
ISOMimeFileType
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO MimeFileType
}
\description{
ISOMimeFileType
}
\section{Methods}{

\describe{
 \item{\code{new(xml, type, name)}}{
   This method is used to instantiate an ISOMimeFileType
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\references{
ISO/TS 19139:2007 Geographic information -- XML
}
\keyword{ISO}
\keyword{file}
\keyword{mime}
\keyword{type}

