% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOMultiplicityRange.R
\docType{class}
\name{ISOMultiplicityRange}
\alias{ISOMultiplicityRange}
\title{ISOMultiplicityRange}
\format{\code{\link{R6Class}} object.}
\usage{
ISOMultiplicityRange
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO MultiplicityRange
}
\description{
ISOMultiplicityRange
}
\section{Fields}{

\describe{
\item{\code{value}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(xml, lower, upper)}}{
   This method is used to instantiate an ISOMultiplicityRange
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{multiplicity}
\keyword{range}

