% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOOnLineFunction.R
\docType{class}
\name{ISOOnLineFunction}
\alias{ISOOnLineFunction}
\title{ISOOnLineFunction}
\format{\code{\link{R6Class}} object.}
\usage{
ISOOnLineFunction
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO OnLineFunction
}
\description{
ISOOnLineFunction
}
\section{Fields}{

\describe{
\item{\code{value}}{}
}}

\section{Methods}{

\describe{
 \item{\code{new(xml,value, description)}}{
   This method is used to instantiate an ISOOnLineFunction
 }
}
}

\examples{
  #possible values
  values <- ISOOnLineFunction$values(labels = TRUE)
  
  #example
  download <- ISOOnLineFunction$new(value = "download")
  
}
\references{
ISO 19115:2003 - Geographic information -- Metadata
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{OnLineFunction}
