% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOParameterDirection.R
\docType{class}
\name{ISOParameterDirection}
\alias{ISOParameterDirection}
\title{ISOParameterDirection}
\format{\code{\link{R6Class}} object.}
\usage{
ISOParameterDirection
}
\value{
Object of \code{\link{R6Class}} for modelling an ISOParameterDirection
}
\description{
ISOParameterDirection
}
\section{Fields}{

\describe{
\item{\code{value}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(xml,value)}}{
   This method is used to instantiate an ISOParameterDirection
 }
}
}
\examples{
#possible values
  values <- ISOParameterDirection$values(labels = TRUE)

  #paramDir
  paramDir <- ISOParameterDirection$new(value = "in")
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\references{
ISO 19119:2005 - Geographic information -- Services
}
\keyword{ISO}
\keyword{direction}
\keyword{parameter}

