% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOPixelOrientation.R
\docType{class}
\name{ISOPixelOrientation}
\alias{ISOPixelOrientation}
\title{ISOPixelOrientation}
\format{\code{\link{R6Class}} object.}
\usage{
ISOPixelOrientation
}
\value{
Object of \code{\link{R6Class}} for modelling an ISOPixelOrientation
}
\description{
ISOPixelOrientation
}
\section{Methods}{

\describe{
 \item{\code{new(xml,value, description)}}{
   This method is used to instantiate an \code{\link{ISOPixelOrientation}}
 }
}
}

\examples{
  #possible values
  values <- ISOPixelOrientation$values(labels = TRUE)
  
  #PixelOrientation
  PixelOrientation <- ISOPixelOrientation$new(value = "center")

}
\references{
ISO 19115:2003 - Geographic information -- Metadata
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{orientation}
\keyword{pixel}
