% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOPresentationForm.R
\docType{class}
\name{ISOPresentationForm}
\alias{ISOPresentationForm}
\title{ISOPresentationForm}
\format{\code{\link{R6Class}} object.}
\usage{
ISOPresentationForm
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO PresentationForm
}
\description{
ISOPresentationForm
}
\section{Fields}{

\describe{
\item{\code{value}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(xml,value)}}{
   This method is used to instantiate an ISOPresentationForm
 }
}
}
\examples{
#possible values
  values <- ISOPresentationForm$values(labels = TRUE)

  #mapDigital type
  map <- ISOPresentationForm$new(value = "mapDigital")
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{form}
\keyword{presentation}

