% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOProductionSeries.R
\docType{class}
\name{ISOProductionSeries}
\alias{ISOProductionSeries}
\title{ISOProductionSeries}
\format{\code{\link{R6Class}} object.}
\usage{
ISOProductionSeries
}
\value{
Object of \code{\link{R6Class}} for modelling an ISOProductionSeries
}
\description{
ISOProductionSeries
}
\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to instantiate an ISOProductionSeries
 }
}
}

\section{Inherited methods from \code{ISOAbstractAggregate}}{

\describe{
 \item{\code{addComposedOf(composedOf)}}{
   Adds a dataset, object of class \code{ISODataSet}
 }
 \item{\code{delComposedOf(composedOf)}}{
   Deletes a dataset, object of class \code{ISODataSet}
 }
 \item{\code{addProductionSeriesMetadata(metadata)}}{
   Adds a ProductionSeries metadata, object of class \code{ISOMetadata}
 }
 \item{\code{delProductionSeriesMetadata(metadata)}}{
   Deletes a ProductionSeries metadata, object of class \code{ISOMetadata}
 }
 \item{\code{addSubset(subset)}}{
   Adds a subset, object that inherits from \code{ISOAbstractAggregate}
 }
 \item{\code{delSubset(subset)}}{
   Deletes a subset, object that inherits from \code{ISOAbstractAggregate}
 }
 \item{\code{addSuperset(superset)}}{
   Adds a superset, object that inherits from \code{ISOAbstractAggregate}
 }
 \item{\code{delSuperset(superset)}}{
   Deletes a superset, object that inherits from \code{ISOAbstractAggregate}
 }
}
}

\references{
ISO 19115:2003 - Geographic information -- Metadata
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{production}
\keyword{series}
