% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOPropertyType.R
\docType{class}
\name{ISOPropertyType}
\alias{ISOPropertyType}
\title{ISOPropertyType}
\format{\code{\link{R6Class}} object.}
\usage{
ISOPropertyType
}
\value{
Object of \code{\link{R6Class}} for modelling an ISOPropertyType
}
\description{
ISOPropertyType
}
\section{Fields}{

\describe{
\item{\code{memberName}}{}

\item{\code{definition}}{}

\item{\code{cardinality}}{}

\item{\code{definitionReference}}{}
}}
\section{Inherited methods from \code{ISOCarrierOfCharacteristics}}{

\describe{
 \item{\code{setFeatureType(featureType)}}{
   Sets a feature type, object of class \code{ISOFeatureType}
 }
 \item{\code{addConstraint(constraint)}}{
   Adds a constraint, object of class \code{ISOConstraint} or \code{character}
 }
 \item{\code{delConstraint(constraint)}}{
   Deletes a constraint, object of class \code{ISOConstraint} or \code{character}
 }
}
}

\section{Methods}{

\describe{
 \item{\code{new(xml, defaults)}}{
   This method is used to instantiate an ISOPropertyType
 }
 \item{\code{setMemberName(memberName)}}{
   Sets the member name. Object of class \code{ISOLocalName} or \code{"character"}
 }
 \item{\code{setDefinition(definition)}}{
   Sets the definition
 }
 \item{\code{setCardinality(lower, upper)}}{
   Sets the cardinality boundaries lower and upper of class \code{numeric}
 }
 \item{\code{setDefinitionReference(definitionReference)}}{
   Sets the definition Reference, object of class \code{ISODefinitionReference}
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\references{
ISO 19110:2005 Methodology for Feature cataloguing
}
\keyword{ISO}
\keyword{property}
\keyword{type}

