% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISORangeDimension.R
\docType{class}
\name{ISORangeDimension}
\alias{ISORangeDimension}
\title{ISORangeDimension}
\format{\code{\link{R6Class}} object.}
\usage{
ISORangeDimension
}
\value{
Object of \code{\link{R6Class}} for modelling an ISORangeDimension
}
\description{
ISORangeDimension
}
\section{Fields}{

\describe{
\item{\code{sequenceIdentifier}}{}

\item{\code{descriptor}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to instantiate an ISORangeDimension
 }
 \item{\code{setSequenceIdentifier(memberName)}}{
   Sets the sequence identifier, object of class \code{ISOMemberName}
 }
 \item{\code{setDescriptor(descriptor)}}{
   Sets the descriptor, object of class \code{character}
 }
}
}
\examples{
#create dimension
   md <- ISORangeDimension$new()
   md$setSequenceIdentifier(ISOMemberName$new(aName = "name", attributeType = "type"))
   md$setDescriptor("descriptor")
   xml <- md$encode()
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{dimension}
\keyword{range}

