% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISORecord.R
\docType{class}
\name{ISORecord}
\alias{ISORecord}
\title{ISORecord}
\format{\code{\link{R6Class}} object.}
\usage{
ISORecord
}
\value{
Object of \code{\link{R6Class}} for modelling an ISORecord
}
\description{
ISORecord
}
\section{Fields}{

\describe{
\item{\code{value}}{}
}}

\section{Methods}{

\describe{
 \item{\code{new(xml,value)}}{
   This method is used to instantiate an ISORecord
 }
}
}

\references{
ISO/TS 19103:2005 Geographic information -- Conceptual schema language
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{record}
