% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISORecordType.R
\docType{class}
\name{ISORecordType}
\alias{ISORecordType}
\title{ISORecordType}
\format{\code{\link{R6Class}} object.}
\usage{
ISORecordType
}
\value{
Object of \code{\link{R6Class}} for modelling an ISORecordType
}
\description{
ISORecordType
}
\section{Fields}{

\describe{
\item{\code{value}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(xml,value)}}{
   This method is used to instantiate an ISORecordType
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\references{
ISO/TS 19103:2005 Geographic information -- Conceptual schema language
}
\keyword{ISO}
\keyword{record}

