% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOReferenceIdentifier.R
\docType{class}
\name{ISOReferenceIdentifier}
\alias{ISOReferenceIdentifier}
\title{ISOReferenceIdentifier}
\format{\code{\link{R6Class}} object.}
\usage{
ISOReferenceIdentifier
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO ReferenceIdentifier
}
\description{
ISOReferenceIdentifier
}
\section{Methods}{

\describe{
 \item{\code{new(xml, code, codeSpace)}}{
   This method is used to instantiate an ISOReferenceIdentifier
 }
 \item{\code{setCodeSpace(codeSpace)}}{
   Sets a codeSpace
 }
 \item{\code{setVersion(version)}}{
   Sets a version
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{identifier}
\keyword{reference}

