% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISORestriction.R
\docType{class}
\name{ISORestriction}
\alias{ISORestriction}
\title{ISOHierarchyLevel}
\format{\code{\link{R6Class}} object.}
\usage{
ISORestriction
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO Restriction
}
\description{
ISOHierarchyLevel
}
\section{Methods}{

\describe{
 \item{\code{new(xml,value, description)}}{
   This method is used to instantiate an \code{\link{ISORestriction}}
 }
}
}

\examples{
  #possible values
  values <- ISORestriction$values(labels = TRUE)
  
  #copyright restriction
  cr <- ISORestriction$new(value = "copyright")
  
}
\references{
ISO 19115:2003 - Geographic information -- Metadata
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{Restriction}
