% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISORole.R
\docType{class}
\name{ISORole}
\alias{ISORole}
\title{ISORole}
\format{\code{\link{R6Class}} object.}
\usage{
ISORole
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO Role
}
\description{
ISORole
}
\section{Fields}{

\describe{
\item{\code{value}}{}
}}

\section{Methods}{

\describe{
 \item{\code{new(xml,value)}}{
   This method is used to instantiate an ISORole
 }
}
}

\examples{
  #possible values
  values <- ISORole$values(labels = TRUE)
  
  #publisher restriction
  role <- ISORole$new(value = "publisher")
  
}
\references{
ISO 19115:2003 - Geographic information -- Metadata
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{role}
