% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISORoleType.R
\docType{class}
\name{ISORoleType}
\alias{ISORoleType}
\title{ISORoleType}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO RoleType
}
\description{
ISORoleType
}
\section{Methods}{

\describe{
 \item{\code{new(xml,value, description)}}{
   This method is used to instantiate an \code{\link{ISORoleType}}
 }
}
}

\examples{
  #possible values
  values <- ISORoleType$values(labels = TRUE)
  
  #some charset
  ordinaryType <- ISORoleType$new(value = "ordinary")

}
\references{
ISO 19110:2005 Methodology for Feature cataloguing
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{roleType}
