% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOScale.R
\docType{class}
\name{ISOScale}
\alias{ISOScale}
\title{ISOScale}
\format{\code{\link{R6Class}} object.}
\usage{
ISOScale
}
\value{
Object of \code{\link{R6Class}} for modelling an ISOScale measure
}
\description{
ISOScale
}
\section{Fields}{

\describe{
\item{\code{value}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(xml,value)}}{
   This method is used to instantiate an ISOScale measure
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\references{
ISO/TS 19103:2005 Geographic information -- Conceptual schema language
}
\keyword{ISO}
\keyword{measure}
\keyword{scale}

