% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOScopeDescription.R
\docType{class}
\name{ISOScopeDescription}
\alias{ISOScopeDescription}
\title{ISOScopeDescription}
\format{\code{\link{R6Class}} object.}
\usage{
ISOScopeDescription
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO ScopeDescription
}
\description{
ISOScopeDescription
}
\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to instantiate an \code{\link{ISOScopeDescription}}
 }
 \item{\code{addAttribute(attribute)}}{
   Adds an attribute
 }
 \item{\code{delAttribute(attribute)}}{
   Deletes an attribute
 }
 \item{\code{addAttributeInstance(attributeInstance)}}{
   Adds an attribute instance
 }
 \item{\code{delAttributeInstance(attributeInstance)}}{
   Deletes an attribute instance
 }
 \item{\code{addFeature(feature)}}{
   Adds an feature
 }
 \item{\code{delFeature(feature)}}{
   Deletes an feature
 }
 \item{\code{addFeatureInstance(featureInstance)}}{
   Adds an feature instance
 }
 \item{\code{delFeatureInstance(featureInstance)}}{
   Deletes an feature instance
 }
 \item{\code{setDataset(dataset)}}{
   Set the dataset
 }
 \item{\code{setOther(other)}}{
   Set other description
 }
}
}

\examples{
 md <- ISOScopeDescription$new()
 xml <- md$encode()

}
\references{
ISO 19115:2003 - Geographic information -- Metadata
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{browse}
\keyword{graphic}
