% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOScopedName.R
\docType{class}
\name{ISOScopedName}
\alias{ISOScopedName}
\title{ISOScopedName}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO ScopedName
}
\description{
ISOScopedName
}
\section{Fields}{

\describe{
\item{\code{value}}{[\code{\link{character}}] scope name}
}}

\section{Methods}{

\describe{
 \item{\code{new(xml,value)}}{
   This method is used to instantiate an \code{\link{ISOScopedName}}
 }
}
}

\references{
ISO/TS 19103:2005 Geographic information -- Conceptual schema language
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{name}
\keyword{scoped}
