% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOSeries.R
\docType{class}
\name{ISOSeries}
\alias{ISOSeries}
\title{ISOSeries}
\format{\code{\link{R6Class}} object.}
\usage{
ISOSeries
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO Extent
}
\description{
ISOSeries
}
\section{Fields}{

\describe{
\item{\code{page}}{}
}}

\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to instantiate an ISOSeries
 }
 \item{\code{setName(name, locales)}}{
   Sets a name (object of class "character"). Locale names can be 
   specified as \code{list} with the \code{locales} argument.
 }
 \item{\code{setIssueIdentification(issueId, locales)}}{
   Sets an issue identification (object of class "character"). Locale 
   names can be specified as \code{list} with the \code{locales} argument.
 }
 \item{\code{setPage(page, locales)}}{
   Sets a page (object of class "character"). Locale names can be 
   specified as \code{list} with the \code{locales} argument.
 }
}
}

\references{
ISO 19115:2003 - Geographic information -- Metadata
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{series}
