% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOSpatialRepresentation.R
\docType{class}
\name{ISOSpatialRepresentation}
\alias{ISOSpatialRepresentation}
\title{ISOSpatialRepresentation}
\format{\code{\link{R6Class}} object.}
\usage{
ISOSpatialRepresentation
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO abstract SpatialRepresentation
}
\description{
ISOSpatialRepresentation
}
\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to instantiate an ISOSpatialRepresentation
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{address}

