% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOSpatialRepresentationType.R
\docType{class}
\name{ISOSpatialRepresentationType}
\alias{ISOSpatialRepresentationType}
\title{ISOSpatialRepresentationType}
\format{\code{\link{R6Class}} object.}
\usage{
ISOSpatialRepresentationType
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO SpatialRepresentationType
}
\description{
ISOSpatialRepresentationType
}
\section{Fields}{

\describe{
\item{\code{value}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(xml,value)}}{
   This method is used to instantiate an ISOSpatialRepresentationType
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{representation}
\keyword{spatial}
\keyword{type}

