% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOSpatialTemporalExtent.R
\docType{class}
\name{ISOSpatialTemporalExtent}
\alias{ISOSpatialTemporalExtent}
\title{ISOSpatialTemporalExtent}
\format{\code{\link{R6Class}} object.}
\usage{
ISOSpatialTemporalExtent
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO SpatialTemporalExtent
}
\description{
ISOSpatialTemporalExtent
}
\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to instantiate an ISOTemporalExtent
 }
 \item{\code{addSpatialExtent(spatialExtent)}}{
   Adds an object of class \code{ISOGeographicExtent}
 }
 \item{\code{setSpatialExtent(spatialExtent)}}{
   Sets an object of class \code{ISOGeographicExtent}
 }
 \item{\code{delSpatialExtent(spatialExtent)}}{
   Deletes an object of class \code{ISOGeographicExtent}
 }
}
}
\examples{
#create object
  md <- ISOSpatialTemporalExtent$new()
  start <- ISOdate(2000, 1, 12, 12, 59, 45)
  end <- ISOdate(2010, 8, 22, 13, 12, 43)
  tp <- GMLTimePeriod$new(beginPosition = start, endPosition = end)
  md$setTimePeriod(tp)
  spatialExtent <- ISOGeographicBoundingBox$new(minx = -180, miny = -90, maxx = 180, maxy = 90)
  md$setSpatialExtent(spatialExtent)

  xml <- md$encode()
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\references{
ISO 19115:2003 - Geographic information -- Metadata
}
\keyword{ISO}
\keyword{extent}
\keyword{spatialtemporal}

