% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOTemporalExtent.R
\docType{class}
\name{ISOTemporalExtent}
\alias{ISOTemporalExtent}
\title{ISOTemporalExtent}
\format{\code{\link{R6Class}} object.}
\usage{
ISOTemporalExtent
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO TemporalExtent
}
\description{
ISOTemporalExtent
}
\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to instantiate an ISOTemporalExtent
 }
 \item{\code{setTimeInstant(timeInstant)}}{
   Sets a time instant
 }
 \item{\code{setTimePeriod(timePeriod)}}{
   Sets a time period
 }
}
}
\examples{
te <- ISOTemporalExtent$new()
   start <- ISOdate(2000, 1, 12, 12, 59, 45)
   end <- ISOdate(2010, 8, 22, 13, 12, 43)
   tp <- ISOTimePeriod$new(beginPosition = start, endPosition = end)
   te$setTimePeriod(tp)
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\references{
ISO 19115:2003 - Geographic information -- Metadata
}
\keyword{ISO}
\keyword{extent}
\keyword{temporal}

