% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOTemporalPrimitive.R
\docType{class}
\name{ISOTemporalPrimitive}
\alias{ISOTemporalPrimitive}
\title{ISOTemporalPrimitive}
\format{\code{\link{R6Class}} object.}
\usage{
ISOTemporalPrimitive
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO abstract temporal primitive
}
\description{
ISOTemporalPrimitive
}
\note{
Class used internally by geometa
}
\section{Methods}{

\describe{
 \item{\code{new(xml, element, namespace, defaults)}}{
   This method is used to instantiate an ISOTemporalPrimitive
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{primitive}
\keyword{temporal}

