% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMLTimePeriod.R
\docType{class}
\name{ISOTimePeriod}
\alias{ISOTimePeriod}
\title{ISOTimePeriod}
\format{\code{\link{R6Class}} object.}
\usage{
ISOTimePeriod
}
\value{
Object of \code{\link{R6Class}} for modelling an GMLTimePeriod
}
\description{
ISOTimePeriod
}
\section{Warning}{
 Deprecated class, use \code{GMLTimePeriod} instead
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{period}
\keyword{time}
