% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOTopicCategory.R
\docType{class}
\name{ISOTopicCategory}
\alias{ISOTopicCategory}
\title{ISOTopicCategory}
\format{\code{\link{R6Class}} object.}
\usage{
ISOTopicCategory
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO TopicCategory
}
\description{
ISOTopicCategory
}
\section{Methods}{

\describe{
 \item{\code{new(xml,value, description)}}{
   This method is used to instantiate an \code{\link{ISOTopicCategory}}
 }
}
}

\examples{
  #possible values
  values <- ISOTopicCategory$values(labels = TRUE)
  
  #biota topic
  biota <- ISOTopicCategory$new(value = "biota")
  
}
\references{
ISO 19115:2003 - Geographic information -- Metadata
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{category}
\keyword{topic}
