% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOTopologyLevel.R
\docType{class}
\name{ISOTopologyLevel}
\alias{ISOTopologyLevel}
\title{ISOTopologyLevel}
\format{\code{\link{R6Class}} object.}
\usage{
ISOTopologyLevel
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO TopologyLevel
}
\description{
ISOTopologyLevel
}
\section{Fields}{

\describe{
\item{\code{value}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(value)}}{
   This method is used to instantiate an ISOTopologyLevel
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{charset}
\keyword{level}
\keyword{topology}

