% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOUnitOfMeasure.R
\docType{class}
\name{ISOUnitOfMeasure}
\alias{ISOUnitOfMeasure}
\title{ISOUnitOfMeasure}
\format{\code{\link{R6Class}} object.}
\usage{
ISOUnitOfMeasure
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO abstract UnitOfMeasure
}
\description{
ISOUnitOfMeasure
}
\section{Fields}{

\describe{
\item{\code{uomName}}{}

\item{\code{uomSymbol}}{}

\item{\code{measureType}}{}

\item{\code{nameStandardUnit}}{}

\item{\code{scaleToStandardUnit}}{}

\item{\code{offsetToStandardUnit}}{}

\item{\code{formula}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(xml,value)}}{
   This method is used to instantiate an ISOUnitOfMeasure
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}

