% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOUnlimitedInteger.R
\docType{class}
\name{ISOUnlimitedInteger}
\alias{ISOUnlimitedInteger}
\title{ISOUnlimitedInteger}
\format{\code{\link{R6Class}} object.}
\usage{
ISOUnlimitedInteger
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO UnlimitedInteger
}
\description{
ISOUnlimitedInteger
}
\section{Fields}{

\describe{
\item{\code{value}}{}
}}
\note{
Class used by geometa internal XML decoder/encoder
}
\section{Methods}{

\describe{
 \item{\code{new(xml,value)}}{
   This method is used to instantiate an ISOUnlimitedInteger
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{integer}
\keyword{unlimited}

