% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOUomAngle.R
\docType{class}
\name{ISOUomAngle}
\alias{ISOUomAngle}
\title{ISOUomAngle}
\format{\code{\link{R6Class}} object.}
\usage{
ISOUomAngle
}
\value{
Object of \code{\link{R6Class}} for modelling an ISOUomAngle
}
\description{
ISOUomAngle
}
\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to instantiate an ISOUomAngle
 }
}
}
\examples{
uom <- ISOUomAngle$new()
  uom$setUomName("Degrees")
  uom$setUomSymbol("deg")
  xml <- uom$encode()
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\references{
ISO/TS 19103:2005 Geographic information -- Conceptual schema language
}
\keyword{ISO}
\keyword{angle}
\keyword{measure}
\keyword{unit}

