% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOUomArea.R
\docType{class}
\name{ISOUomArea}
\alias{ISOUomArea}
\title{ISOUomArea}
\format{\code{\link{R6Class}} object.}
\usage{
ISOUomArea
}
\value{
Object of \code{\link{R6Class}} for modelling an ISOUomArea
}
\description{
ISOUomArea
}
\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to instantiate an ISOUomArea
 }
}
}
\examples{
uom <- ISOUomArea$new()
  uom$setUomName("Square meters")
  uom$setUomSymbol("sqm")
  xml <- uom$encode()
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{area}
\keyword{measure}
\keyword{unit}

