% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOUomCurrency.R
\docType{class}
\name{ISOUomCurrency}
\alias{ISOUomCurrency}
\title{ISOUomCurrency}
\format{\code{\link{R6Class}} object.}
\usage{
ISOUomCurrency
}
\value{
Object of \code{\link{R6Class}} for modelling an ISOUomCurrency
}
\description{
ISOUomCurrency
}
\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to instantiate an ISOUomCurrency
 }
}
}
\examples{
uom <- ISOUomCurrency$new()
  uom$setUomName("Dollar")
  uom$setUomSymbol("$")
  xml <- uom$encode()
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\references{
ISO/TS 19103:2005 Geographic information -- Conceptual schema language
}
\keyword{ISO}
\keyword{currency}
\keyword{measure}
\keyword{unit}

