% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOUomLength.R
\docType{class}
\name{ISOUomLength}
\alias{ISOUomLength}
\title{ISOUomLength}
\format{\code{\link{R6Class}} object.}
\usage{
ISOUomLength
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO UomLength
}
\description{
ISOUomLength
}
\section{Methods}{

\describe{
 \item{\code{new(xml,value)}}{
   This method is used to instantiate an ISOUomLength
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{length}
\keyword{measure}
\keyword{unit}

