% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOUomScale.R
\docType{class}
\name{ISOUomScale}
\alias{ISOUomScale}
\title{ISOUomScale}
\format{\code{\link{R6Class}} object.}
\usage{
ISOUomScale
}
\value{
Object of \code{\link{R6Class}} for modelling an ISOUomScale
}
\description{
ISOUomScale
}
\section{Methods}{

\describe{
 \item{\code{new(xml,value)}}{
   This method is used to instantiate an ISOUomScale
 }
}
}
\examples{
uom <- ISOUomScale$new()
  uom$setUomName("scaleunit")
  uom$setUomSymbol("scale")
  xml <- uom$encode()
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\references{
ISO/TS 19103:2005 Geographic information -- Conceptual schema language
}
\keyword{ISO}
\keyword{measure}
\keyword{scale}
\keyword{unit}

