% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOUomTime.R
\docType{class}
\name{ISOUomTime}
\alias{ISOUomTime}
\title{ISOUomTime}
\format{\code{\link{R6Class}} object.}
\usage{
ISOUomTime
}
\value{
Object of \code{\link{R6Class}} for modelling an ISOUomTime
}
\description{
ISOUomTime
}
\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to instantiate an ISOUomTime
 }
}
}
\examples{
uom <- ISOUomTime$new()
  uom$setUomName("Hours")
  uom$setUomSymbol("h")
  xml <- uom$encode()
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\references{
ISO/TS 19103:2005 Geographic information -- Conceptual schema language
}
\keyword{ISO}
\keyword{measure}
\keyword{time}
\keyword{unit}

