% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOUomVelocity.R
\docType{class}
\name{ISOUomVelocity}
\alias{ISOUomVelocity}
\title{ISOUomVelocity}
\format{\code{\link{R6Class}} object.}
\usage{
ISOUomVelocity
}
\value{
Object of \code{\link{R6Class}} for modelling an ISOUomVelocity
}
\description{
ISOUomVelocity
}
\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to instantiate an ISOUomVelocity
 }
}
}
\examples{
uom <- ISOUomVelocity$new()
  uom$setUomName("meters per second")
  uom$setUomSymbol("m/s")
  xml <- uom$encode()
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{measure}
\keyword{unit}
\keyword{velocity}

