% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOUomVolume.R
\docType{class}
\name{ISOUomVolume}
\alias{ISOUomVolume}
\title{ISOUomVolume}
\format{\code{\link{R6Class}} object.}
\usage{
ISOUomVolume
}
\value{
Object of \code{\link{R6Class}} for modelling an ISOUomVolume
}
\description{
ISOUomVolume
}
\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to instantiate an ISOUomVolume
 }
}
}
\examples{
uom <- ISOUomVolume$new()
  uom$setUomName("cubic meters")
  uom$setUomSymbol("m3")
  xml <- uom$encode()
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{measure}
\keyword{unit}
\keyword{volume}

