% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOUomWeight.R
\docType{class}
\name{ISOUomWeight}
\alias{ISOUomWeight}
\title{ISOUomWeight}
\format{\code{\link{R6Class}} object.}
\usage{
ISOUomWeight
}
\value{
Object of \code{\link{R6Class}} for modelling an ISOUomWeight
}
\description{
ISOUomWeight
}
\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to instantiate an ISOUomWeight
 }
}
}
\examples{
uom <- ISOUomWeight$new()
  uom$setUomName("grams")
  uom$setUomSymbol("g")
  xml <- uom$encode()
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{measure}
\keyword{unit}
\keyword{weight}

