% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOURL.R
\docType{class}
\name{ISOURL}
\alias{ISOURL}
\title{ISOURL}
\format{\code{\link{R6Class}} object.}
\usage{
ISOURL
}
\value{
Object of \code{\link{R6Class}} for modelling an ISOURL
}
\description{
ISOURL
}
\section{Fields}{

\describe{
\item{\code{value}}{}
}}
\note{
Class used by geometa internal XML decoder/encoder
}
\section{Methods}{

\describe{
 \item{\code{new(xml,value)}}{
   This method is used to instantiate an ISOURL
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\references{
ISO 19115:2003 - Geographic information -- Metadata
}
\keyword{ISO}
\keyword{URL}

