% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOUsage.R
\docType{class}
\name{ISOUsage}
\alias{ISOUsage}
\title{ISOUsage}
\format{\code{\link{R6Class}} object.}
\usage{
ISOUsage
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO Usage
}
\description{
ISOUsage
}
\section{Fields}{

\describe{
\item{\code{userContactInfo}}{}
}}

\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to instantiate an ISOUsage
 }
 \item{\code{setSpecificUsage(specificUsage, locales)}}{
   Set the specific usage
 }
 \item{\code{setUsageDateTime(usageDateTime)}}{
   Set the usage date time, object of class \code{POSIXt}
 }
 \item{\code{setUserDeterminedLimitations(userDeterminedLimitations, locales)}}{
   Set the limitations determined by user
 }
 \item{\code{addUserContact(contact)}}{
   Adds user contact, object of class \code{ISOResponsibleParty}
 }
 \item{\code{delUserContact(contact)}}{
   Deletes user contact, object of class \code{ISOResponsibleParty}
 }
}
}

\references{
ISO 19115:2003 - Geographic information -- Metadata
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{usage}
