% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOVerticalExtent.R
\docType{class}
\name{ISOVerticalExtent}
\alias{ISOVerticalExtent}
\title{ISOVerticalExtent}
\format{\code{\link{R6Class}} object.}
\usage{
ISOVerticalExtent
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO VerticalExtent
}
\description{
ISOVerticalExtent
}
\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to instantiate an ISOVerticalExtent
 }
}
}
\examples{
ve <- ISOVerticalExtent$new()
  ve$setMinimumValue(0)
  ve$setMaximumValue(19)
  xml <- ve$encode()
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\references{
ISO 19115:2003 - Geographic information -- Metadata
}
\keyword{ISO}
\keyword{extent}
\keyword{vertical}

