% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineExprs.R
\name{mergeByRowname}
\alias{mergeByRowname}
\title{Merge two matrix by row names}
\usage{
mergeByRowname(x, y, all = TRUE, ...)
}
\arguments{
\item{x, y}{matrix, or objects to be coerced to one.}

\item{all}{logical;all = L is shorthand for all.x = L and all.y = L, where L is either TRUE or FALSE.}

\item{...}{arguments to be passed to or from methods.}
}
\value{
The merged matrix.
}
\description{
Given two matrix with columns and row names, merge them by row names.
The function derives from merge function in base package, so it can receive arguments from merge.
}
\details{
nothing
}
\examples{
x=matrix(1:9,nrow=3,dimnames = list(c("row1","row2","row3"),c("col1","col2","col3")))
y=matrix(1:8,nrow=4,dimnames = list(c("row2","row3","row4","row5"),c("col1","col5")))
mergeByRowname(x,y)
mergeByRowname(x,y,all.x=FALSE)
}
