% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_gene_input.R
\name{meta_gene_input}
\alias{meta_gene_input}
\title{Regional Plot Association Results of Pig}
\usage{
meta_gene_input(metaDat, geneName)
}
\arguments{
\item{metaDat}{a dataframe includes GSE_GPL, GSE, GSM, Group}

\item{geneName}{a character of gene SYMBOL name}
}
\value{
a dataframe of meta input data
}
\description{
This is a plotting tool to simultaneously display the association p-value,
             the LD, recombination rate and annotated gene in region of interesting in pig.
             The interpretation and visualization tool of association results will
             contribute to further identification of causal mutations.
}
\details{
nothing
}
