% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_gene.R
\name{metaGene}
\alias{metaGene}
\title{Perpare the input data of one gene for meta analysis}
\usage{
metaGene(pheGene, geneName)
}
\arguments{
\item{pheGene}{A dataframe, includes GSE, GSM, Group, GeneExpr}

\item{geneName}{a character of gene SYMBOL name}
}
\value{
a dataframe of meta input data
}
\description{
nothing
}
\details{
nothing
}
