% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaGene.R
\name{metaRes}
\alias{metaRes}
\title{Extract the result of data with meta class}
\usage{
metaRes(metaData)
}
\arguments{
\item{metaData}{A dataframe of meta input data}
}
\value{
A data.frame of meta results for one gene
}
\description{
nothing
}
\details{
nothing
}
