% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOCodelist.R
\name{parseISOCodelists}
\alias{parseISOCodelists}
\title{parseISOCodelists}
\usage{
parseISOCodelists(version)
}
\arguments{
\item{version}{the version of the metadata standard}
}
\value{
a \code{list} of \link{ISOCodelist} objects
}
\description{
\code{parseISOCodelists} allows to parse the list of ISO codelists
registered in \pkg{geometa}
}
\note{
Internal method used to parse ISO codelists for package build
}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
