% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineExprs.R
\name{path_tail}
\alias{path_tail}
\title{Convert delimiter to "/" for different system}
\usage{
path_tail(path)
}
\arguments{
\item{path}{A character for the path}
}
\value{
A character for the path
}
\description{
Convert delimiter to "/" for different system, and end of "/"
}
