% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometa_mapping.R
\docType{class}
\name{pivot_converter}
\alias{pivot_converter}
\title{pivot_converter}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling a mapping format converter
}
\description{
pivot_converter

pivot_converter
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{convert}
\keyword{mapping}
\keyword{metadata}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{from}}{from}

\item{\code{to}}{to}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-pivot_converter-new}{\code{pivot_converter$new()}}
\item \href{#method-pivot_converter-clone}{\code{pivot_converter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pivot_converter-new"></a>}}
\if{latex}{\out{\hypertarget{method-pivot_converter-new}{}}}
\subsection{Method \code{new()}}{
Initializes pivot converter
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pivot_converter$new(from, to)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{from}}{from}

\item{\code{to}}{to}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pivot_converter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-pivot_converter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pivot_converter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
