% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEtoExpr.R
\name{probesToGene}
\alias{probesToGene}
\title{Replace the probes id of GPL by gene symbol}
\usage{
probesToGene(exprSet, probe_symbol)
}
\arguments{
\item{exprSet}{matrix, the raw expression matrix with colname of individual ids and rowname of probe ids.}

\item{probe_symbol}{data.frame with probeID and symbolID.}
}
\value{
matrix, the clean expression matrix with colname of individual ids and rowname of gene symbol.
}
\description{
Given a expression matrix and the GPL annotation with probeID and symbolID,
find the probe with maximum of total expression from same gene due to one gene have multiple porbes,
return the expression data with rowname of gene symbol. The row number of is commonly less than the row number of raw expression matrix.
}
\details{
nothing
}
