% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOCodelist.R
\name{registerISOCodelist}
\alias{registerISOCodelist}
\title{registerISOCodelist}
\usage{
getISOCodelist(id)
}
\description{
\code{registerISOCodelist} allows to register a new codelist
registered in \pkg{geometa}
}
\examples{
            
  registerISOCodelist(
   refFile = "http://www.isotc211.org/2005/resources/Codelist/ML_gmxCodelists.xml",
   id = "LanguageCode",
   force = TRUE
 )

}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
