% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOMetadataNamespace.R
\name{registerISOMetadataNamespace}
\alias{registerISOMetadataNamespace}
\title{registerISOMetadataNamespace}
\usage{
registerISOMetadataNamespace(id, uri, force)
}
\arguments{
\item{id}{prefix of the namespace}

\item{uri}{URI of the namespace}

\item{force}{logical parameter indicating if registration has be to be forced
in case the identified namespace is already registered}
}
\description{
\code{registerISOMetadataNamespace} allows to register a new namespace
registered in \pkg{geometa}
}
\examples{
            
  registerISOMetadataNamespace(id = "myprefix", uri = "http://someuri")

}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
