% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOMetadataSchemas.R
\name{registerISOMetadataSchema}
\alias{registerISOMetadataSchema}
\title{registerISOMetadataSchema}
\usage{
registerISOMetadataSchema(xsdFile)
}
\arguments{
\item{xsdFile}{the schema XSD file}
}
\description{
\code{registerISOMetadataSchema} allows to register a new schema
in \pkg{geometa}
}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
