% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometa_mapping.R
\name{registerMappingFormat}
\alias{registerMappingFormat}
\title{registerMappingFormat}
\usage{
registerMappingFormat(mapping_format)
}
\arguments{
\item{mapping_format}{object of class \code{pivot_format}}
}
\description{
\code{registerMappingFormat} allows to register a new mapping
format in \pkg{geometa}
}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
