% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometa_mapping.R
\name{registerMappings}
\alias{registerMappings}
\title{registerMappings}
\usage{
registerMappings(x)
}
\arguments{
\item{x}{a \code{data.frame} containing the metadata mapping rules}
}
\description{
\code{MappingFile} allows to register in  \pkg{geometa} a \code{data.frame} 
containing mappings rules to convert from/to other metadata formats (currently 
EML/emld objects and NetCDF-CF/ncdf4 objects)
}
