% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEtoExpr.R
\name{saveGSE}
\alias{saveGSE}
\title{Download GSE expression matrix and convert probe to gene symbol}
\usage{
saveGSE(GSE, destdir = "tmp", annotSymbol = FALSE)
}
\arguments{
\item{GSE}{A character, the number of GSE.}

\item{destdir}{A character, the path to download GSE related files.}

\item{annotSymbol}{A Boolean. The default is FALSE, do not annotate probe to gene Symbol.}
}
\value{
Save the files of expression matrix and phenotype information.
}
\description{
Given a GSE number, download related GSE data and convert probe to gene symbol, and return the gene matrix with gene symbol.
}
\details{
nothing
}
