% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometa_options.R
\name{setGeometaOption}
\alias{setGeometaOption}
\title{setGeometaOption}
\usage{
setGeometaOption(option, value)
}
\arguments{
\item{option}{the name of the option}

\item{value}{the value to set for the option}
}
\description{
\code{setGeometaOption} allows to set an option from \pkg{geometa}
}
\examples{
            
  setGeometaOption("schemaBaseUrl", "http://somealternativeurl")

}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
