% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOCodelist.R
\name{setISOCodelists}
\alias{setISOCodelists}
\title{setISOCodelists}
\usage{
setISOCodelists(version)
}
\arguments{
\item{version}{the version of the metadata standard}
}
\description{
\code{setISOCodelists} allows to set the list of ISO codelists
registered in \pkg{geometa}
}
\examples{
            
  setISOCodelists(version = "19139")

}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
