% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOMetadataNamespace.R
\name{setISOMetadataNamespaces}
\alias{setISOMetadataNamespaces}
\title{setISOMetadataNamespaces}
\usage{
setISOMetadataNamespaces(version)
}
\arguments{
\item{version}{the standard version}
}
\description{
\code{setISOMetadataNamespaces} sets the default list of namespaces
}
\examples{
            
  getISOMetadataNamespaces()

}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
