% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOMetadataSchemas.R
\name{setISOMetadataSchemas}
\alias{setISOMetadataSchemas}
\title{setISOMetadataSchemas}
\usage{
setISOMetadataSchemas(version)
}
\arguments{
\item{version}{the schema version}
}
\description{
\code{setISOMetadataSchemas} register the schemas in \pkg{geometa}
}
\examples{
            
  setISOMetadataSchemas(version = "19115-3")

}
