% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOMetadataNamespace.R
\name{setMetadataNamespaces}
\alias{setMetadataNamespaces}
\alias{setISOMetadataNamespaces}
\title{setMetadataNamespaces}
\usage{
setMetadataNamespaces(version)
}
\arguments{
\item{version}{the standard version}
}
\description{
\code{setMetadataNamespaces} sets the default list of namespaces
}
\examples{
            
  getISOMetadataNamespaces()

}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
