% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWEAbstractDataComponent.R
\docType{class}
\name{SWEAbstractDataComponent}
\alias{SWEAbstractDataComponent}
\title{SWEAbstractDataComponent}
\format{
\code{\link{R6Class}} object.
}
\arguments{
\item{xml}{object of class \link{XMLInternalNode-class} from \pkg{XML}}

\item{element}{element}

\item{updatable}{updatable}

\item{optional}{optional}

\item{definition}{definition}

\item{name}{name}

\item{codeSpace}{codespace}
}
\value{
Object of \code{\link{R6Class}} for modelling an SWE Abstract data component
}
\description{
Initializes an object of class \link{SWEAbstractDataComponent}

Adds name

Deletes name
}
\section{Fields}{

\describe{
\item{\code{name}}{name}
}}

\note{
Class used internally by geometa
}
\references{
SWE Common Data Model Encoding Standard. https://www.ogc.org/standards/swecommon
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{SWE}
