% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWEAbstractObject.R
\docType{class}
\name{SWEAbstractObject}
\alias{SWEAbstractObject}
\title{SWEAbstractObject}
\format{
\code{\link{R6Class}} object.
}
\arguments{
\item{xml}{object of class \link{XMLInternalNode-class} from \pkg{XML}}

\item{element}{element}

\item{attrs}{attrs}

\item{defaults}{defaults}

\item{wrap}{wrap}

\item{value_as_field}{whether value should be set as field}
}
\value{
Object of \code{\link{R6Class}} for modelling an SWE abstract object
}
\description{
Initializes an object of class \link{SWEAbstractObject}
}
\note{
Class used internally by geometa
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{GML}
\keyword{ISO}
