% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWEAbstractSimpleComponent.R
\docType{class}
\name{SWEAbstractSimpleComponent}
\alias{SWEAbstractSimpleComponent}
\title{SWEAbstractSimpleComponent}
\format{
\code{\link{R6Class}} object.
}
\arguments{
\item{xml}{object of class \link{XMLInternalNode-class} from \pkg{XML}}

\item{element}{element}

\item{updatable}{updatable}

\item{optional}{optional}

\item{definition}{definition}
}
\value{
Object of \code{\link{R6Class}} for modelling an SWE Abstract simple component
}
\description{
Initializes an object of class \link{SWEAbstractSimpleComponent}
}
\references{
OGC Geography Markup Language. https://www.ogc.org/standards/swecommon
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{SWE}
