% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWEAbstractSWEIdentifiable.R
\docType{class}
\name{SWEAbstractSWEIdentifiable}
\alias{SWEAbstractSWEIdentifiable}
\title{SWEAbstractSWEIdentifiable}
\format{
\code{\link{R6Class}} object.
}
\arguments{
\item{xml}{object of class \link{XMLInternalNode-class} from \pkg{XML}}

\item{identifier}{identifier}

\item{label}{label}

\item{description}{description}
}
\value{
Object of \code{\link{R6Class}} for modelling an SWE abstract identifiable
}
\description{
Initializes a SWE Nil Values object

Set identifier

Set label

Set description
}
\section{Fields}{

\describe{
\item{\code{identifier}}{identifier}

\item{\code{label}}{label}

\item{\code{description}}{description}
}}

\references{
SWE Common Data Model Encoding Standard. https://www.ogc.org/standards/swecommon
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{SWE}
