% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWECount.R
\docType{class}
\name{SWECount}
\alias{SWECount}
\title{SWECount}
\format{
\code{\link{R6Class}} object.
}
\arguments{
\item{xml}{object of class \link{XMLInternalNode-class} from \pkg{XML}}

\item{element}{element}

\item{attrs}{attrs}

\item{defaults}{defaults}

\item{wrap}{wrap}
}
\value{
Object of \code{\link{R6Class}} for modelling an SWE Count
}
\description{
Initializes an object of class \link{SWECount}

setConstraint

setValue
}
\section{Fields}{

\describe{
\item{\code{constraint}}{constraint}

\item{\code{value}}{value}
}}

\references{
OGC Geography Markup Language. https://www.ogc.org/standards/swecommon
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{SWE}
