% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWEDataRecord.R
\docType{class}
\name{SWEDataRecord}
\alias{SWEDataRecord}
\title{SWEDataRecord}
\format{
\code{\link{R6Class}} object.
}
\arguments{
\item{xml}{object of class \link{XMLInternalNode-class} from \pkg{XML}}

\item{element}{element}

\item{updatable}{updatable}

\item{optional}{optional}

\item{definition}{definition}

\item{field}{field}
}
\value{
Object of \code{\link{R6Class}} for modelling an SWE data record
}
\description{
Initializes an object of class \link{SWEDataRecord}

Adds field

Deletes field
}
\section{Fields}{

\describe{
\item{\code{field}}{field}
}}

\note{
Class used internally by geometa
}
\references{
SWE Common Data Model Encoding Standard. https://www.ogc.org/standards/swecommon
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{SWE}
