% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWENilValues.R
\docType{class}
\name{SWENilValues}
\alias{SWENilValues}
\title{SWENilValues}
\format{
\code{\link{R6Class}} object.
}
\arguments{
\item{xml}{object of class \link{XMLInternalNode-class} from \pkg{XML}}

\item{values}{vector of numerical values to consider as nil values}

\item{value}{value}

\item{reason}{reason}
}
\value{
Object of \code{\link{R6Class}} for modelling an SWE nil values object
}
\description{
Initializes a SWE Nil Values object

Adds a nil value with a reason
}
\section{Fields}{

\describe{
\item{\code{nilValue}}{nil value}
}}

\references{
SWE Common Data Model Encoding Standard. https://www.ogc.org/standards/swecommon
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{SWE}
