% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWEText.R
\docType{class}
\name{SWEText}
\alias{SWEText}
\title{SWEText}
\format{
\code{\link{R6Class}} object.
}
\arguments{
\item{xml}{object of class \link{XMLInternalNode-class} from \pkg{XML}}

\item{updatable}{updatable}

\item{optional}{optional}

\item{definition}{definition}

\item{constraint}{constraint}

\item{value}{value}
}
\value{
Object of \code{\link{R6Class}} for modelling an SWE Text
}
\description{
Initializes an object of class \link{SWEText}

setConstraint

setValue
}
\section{Fields}{

\describe{
\item{\code{constraint}}{constraint}

\item{\code{value}}{value}
}}

\references{
OGC Geography Markup Language. https://www.ogc.org/standards/swecommon
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{SWE}
