% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWETextEncoding.R
\docType{class}
\name{SWETextEncoding}
\alias{SWETextEncoding}
\title{SWETextEncoding}
\format{
\code{\link{R6Class}} object.
}
\arguments{
\item{xml}{object of class \link{XMLInternalNode-class} from \pkg{XML}}

\item{collapseWhiteSpaces}{Indicates whether white spaces (i.e. space, tab, CR, LF) 
should be collapsed with separators when parsing the data stream. Default is \code{TRUE}}

\item{decimalSeparator}{Character used as the decimal separator. Default is \code{TRUE}}

\item{tokenSeparator}{Character sequence used as the token separator (i.e. between two successive values). Required}

\item{blockSeparator}{Character sequence used as the block separator (i.e. between two successive blocks in the data set. 
The end of a block is reached once all values from the data tree have been encoded once). Required}
}
\value{
Object of \code{\link{R6Class}} for modelling an SWE text encoding object
}
\description{
Initializes a SWE Text Encoding element
}
\references{
SWE Common Data Model Encoding Standard. https://www.ogc.org/standards/swecommon
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{SWE}
