% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWEXMLEncoding.R
\docType{class}
\name{SWEXMLEncoding}
\alias{SWEXMLEncoding}
\title{SWEXMLEncoding}
\format{
\code{\link{R6Class}} object.
}
\arguments{
\item{xml}{object of class \link{XMLInternalNode-class} from \pkg{XML}}
}
\value{
Object of \code{\link{R6Class}} for modelling an SWE XML encoding object
}
\description{
Initializes a SWE XML Encoding element
}
\references{
SWE Common Data Model Encoding Standard. https://www.ogc.org/standards/swecommon
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{SWE}
