% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineExprs.R
\name{uniqueFeature}
\alias{uniqueFeature}
\title{Convert delimiter to "/" for different system}
\usage{
uniqueFeature(expr, key = "Symbol")
}
\arguments{
\item{expr}{A data.table with Symbol and expression level}

\item{key}{A character for Symbol feature}
}
\value{
A unique data.table
}
\description{
Convert delimiter to "/" for different system, and end of "/"
}
