% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateGPLlist.R
\name{updateAnno}
\alias{updateAnno}
\title{Add new annotation information of GPL to GPLlist}
\usage{
updateAnno(GPL, probeID, symbolID, overwrite = FALSE)
}
\arguments{
\item{GPL}{A character, the number of GSE.}

\item{probeID}{A character, the header of probe from specific GPL platform.}

\item{symbolID}{A character, the header of gene symbol from specific GPL platform.}

\item{overwrite}{By default, use_data() will not overwrite existing files. If you really want to do so, set this to TRUE.}
}
\value{
update the GPLlist.rda in data folder
}
\description{
Given GPL, probeID, and symbolID from GPL data, update the GPLlist data.
}
\details{
nothing
}
\examples{
\dontrun{
GPL="GPL13648"
gplDat = Table(getGEO(GPL,getGPL=TRUE,destdir ="tmp"));colnames(gplDat)
updateAnno(GPL,probeID="ID",symbolID="Gene Symbol")
load("data/GPLlist.rda")
}
}
