% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometa.R
\docType{package}
\name{geometa}
\alias{geometa}
\alias{geometa-package}
\title{geometa: Tools for Reading and Writing ISO/OGC Geographic Metadata}
\description{
Provides facilities to read, write and validate geographic metadata defined with ISO TC211 / OGC ISO geographic information metadata standards, and encoded using the ISO 19139 and ISO 19115-3 (XML) standard technical specifications. This includes ISO 19110 (Feature cataloguing), 19115 (dataset metadata), 19119 (service metadata) and 19136 (GML). Other interoperable schemas from the OGC are progressively supported as well, such as the Sensor Web Enablement (SWE) Common Data Model, the OGC GML Coverage Implementation Schema (GMLCOV), or the OGC GML Referenceable Grid (GMLRGRID).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/eblondel/geometa/wiki}
  \item Report bugs at \url{https://github.com/eblondel/geometa/issues}
}

}
\author{
\strong{Maintainer}: Emmanuel Blondel \email{emmanuel.blondel1@gmail.com} (\href{https://orcid.org/0000-0002-5870-5762}{ORCID})

Other contributors:
\itemize{
  \item R Consortium [funder]
}

}
