% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometa.R
\docType{package}
\name{geometa}
\alias{geometa}
\alias{geometa-package}
\alias{geometa-package}
\title{Tools for Reading and Writing ISO/OGC Geographic Metadata}
\description{
Provides facilities to handle reading and writing of geographic metadata 
defined with OGC/ISO 19115 and 19110 geographic information metadata standards, and 
encoded using the ISO 19139 (XML) standard.
}
\details{
\tabular{ll}{
   Package: \tab geometa\cr
   Type: \tab Package\cr
   Version: \tab 0.3-0\cr
   Date: \tab 2018-02-06\cr
   License: \tab MIT\cr
   LazyLoad: \tab yes\cr
 }
}
\author{
Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}
}
