% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometa_logger.R
\docType{class}
\name{geometaLogger}
\alias{geometaLogger}
\title{geometaLogger}
\format{\code{\link{R6Class}} object.}
\usage{
geometaLogger
}
\value{
Object of \code{\link{R6Class}} for modelling a simple logger
}
\description{
geometaLogger
}
\note{
Logger class used internally by geometa
}
\section{Abstract Methods}{

\describe{
 \item{\code{INFO(text)}}{
   Logger to report information. Used internally
 }
 \item{\code{WARN(text)}}{
   Logger to report warnings. Used internally
 }
 \item{\code{ERROR(text)}}{
   Logger to report errors. Used internally
 }
}
}

\keyword{logger}
