% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOAbstractObject.R
\name{getISOClasses}
\alias{getISOClasses}
\title{getISOClasses}
\usage{
getISOClasses()
}
\description{
get the list of cached ISO classes
}
\examples{
            
  getISOClasses()

}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
