% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOCodelist.R
\name{getISOCodelists}
\alias{getISOCodelists}
\title{getISOCodelists}
\usage{
getISOCodelists()
}
\description{
\code{getISOCodelists} allows to get the list of ISO codelists
registered in \pkg{geometa}, their description and XML definition. The object
returned is of class "data.frame"
}
\examples{
            
  getISOCodelists()

}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
