% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMLAbstractCoordinateOperation.R
\docType{class}
\name{GMLAbstractCoordinateOperation}
\alias{GMLAbstractCoordinateOperation}
\title{GMLAbstractCoordinateOperation}
\format{\code{\link{R6Class}} object.}
\usage{
GMLAbstractCoordinateOperation
}
\value{
Object of \code{\link{R6Class}} for modelling an GMLAbstractCoordinateOperation
}
\description{
GMLAbstractCoordinateOperation
}
\section{Fields}{

\describe{
\item{\code{operationVersion}}{[\code{\link{GMLElement}}]}

\item{\code{coordinateOperationAccuracy}}{[\code{\link{GMLElement}}]}

\item{\code{sourceCRS}}{[\code{\link{GMLElement}}]}

\item{\code{targetCRS}}{[\code{\link{GMLElement}}]}
}}

\section{Methods}{

\describe{
 \item{\code{new(xml, defaults, id)}}{
   This method is used to instantiate a GML Abstract CRS
 }
 \item{\code{setDomainOfValidity(domainOfValidity)}}{
   Sets the domain of validity
 }
 \item{\code{addScope(scope)}}{
   Adds a scope
 }
 \item{\code{delScope(scope)}}{
   Deletes a scope
 }
 \item{\code{setVersion(version)}}{
   Sets version
 }
 \item{\code{addAccuracy(accuracy)}}{
   Adds coordinate operation accuracy, object extending \code{ISOAbstractPositionalAccuracy}
 }
 \item{\code{delAccuracy(accuracy)}}{
   Deletes coordinate operation accuracy, object extending \code{ISOAbstractPositionalAccuracy}
 }
 \item{\code{setSourceCRS(sourceCRS)}}{
   Sets the source CRS, object extending \code{GMLAbstractSingleCRS}
 }
 \item{\code{setTargetCRS(targetCRS)}}{
   Sets the target CRS, object extending \code{GMLAbstractSingleCRS}
 }
}
}

\references{
ISO 19136:2007 Geographic Information -- Geographic Markup Language.
  http://www.iso.org/iso/iso_catalogue/catalogue_tc/catalogue_detail.htm?csnumber=32554 
  
  OGC Geography Markup Language. http://www.opengeospatial.org/standards/gml
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{GML}
\keyword{ISO}
\keyword{Operation}
\keyword{abstract}
\keyword{coordinate}
