% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMLAbstractTimeObject.R
\docType{class}
\name{GMLAbstractTimeObject}
\alias{GMLAbstractTimeObject}
\title{GMLAbstractTimeObject}
\format{\code{\link{R6Class}} object.}
\usage{
GMLAbstractTimeObject
}
\value{
Object of \code{\link{R6Class}} for modelling an GML AbstractTimeObject
}
\description{
GMLAbstractTimeObject
}
\section{Fields}{

\describe{
\item{\code{remarks}}{}
}}

\section{Methods}{

\describe{
 \item{\code{new(xml, defaults)}}{
   This method is used to instantiate a GML AbstractTimeObject
 }
 \item{\code{setId(id)}}{
   Sets the id
 }
 \item{\code{addRemark(remark)}}{
   Adds a remark
 }
 \item{\code{delRemark(remark)}}{
   Deletes a remark
 }
}
}

\references{
ISO 19136:2007 Geographic Information -- Geographic Markup Language.
  http://www.iso.org/iso/iso_catalogue/catalogue_tc/catalogue_detail.htm?csnumber=32554 
  
  OGC Geography Markup Language. http://www.opengeospatial.org/standards/gml
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{GML}
\keyword{ISO}
\keyword{abstract}
\keyword{object}
\keyword{time}
