% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMLConversion.R
\docType{class}
\name{GMLConversion}
\alias{GMLConversion}
\title{GMLConversion}
\format{\code{\link{R6Class}} object.}
\usage{
GMLConversion
}
\value{
Object of \code{\link{R6Class}} for modelling an GMLConversion
}
\description{
GMLConversion
}
\section{Fields}{

\describe{
\item{\code{parameterValue}}{}
}}

\section{Inherited methods}{

\describe{
  from \code{GMLAbstractCoordinateOperation}
}
}

\section{Methods}{

\describe{
 \item{\code{new(xml, defaults, id)}}{
   This method is used to instantiate a GML Conversion
 }
 \item{\code{setMethod(method)}}{
   Sets the method
 }
 \item{\code{addParameterValue(paramValue)}}{
   Adds a parameter value
 }
 \item{\code{delParameterValue(paramValue)}}{
   Deletes a parameter value
 }
}
}

\references{
ISO 19136:2007 Geographic Information -- Geographic Markup Language.
  http://www.iso.org/iso/iso_catalogue/catalogue_tc/catalogue_detail.htm?csnumber=32554 
  
  OGC Geography Markup Language. http://www.opengeospatial.org/standards/gml
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{GML}
\keyword{ISO}
\keyword{conversion}
