% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMLGeneralGridAxis.R
\docType{class}
\name{GMLGeneralGridAxis}
\alias{GMLGeneralGridAxis}
\title{GMLGeneralGridAxis}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an GML GeneralGridAxis
}
\description{
GMLGeneralGridAxis
}
\section{Fields}{

\describe{
\item{\code{offsetVector}}{[\code{\link{GMLElement}}]}

\item{\code{coefficients}}{[\code{\link{GMLElement}}]}

\item{\code{gridAxesSpanned}}{[\code{\link{GMLElement}}]}

\item{\code{sequenceRule}}{[\code{\link{GMLElement}}]}
}}

\note{
Experimental
}
\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to instantiate a GML GeneralGridAxis.
 }
}
}

\references{
OGC Geography Markup Language. http://www.opengeospatial.org/standards/gml
  
  OGC GML 3.3 Schema. http://schemas.opengis.net/gml/3.3/referenceableGrid.xsd
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{GML}
\keyword{GeneralGridAxis}
\keyword{OGC}
