% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMLGeodeticCRS.R
\docType{class}
\name{GMLGeodeticCRS}
\alias{GMLGeodeticCRS}
\title{GMLGeodeticCRS}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an GMLGeodeticCRS
}
\description{
GMLGeodeticCRS
}
\section{Fields}{

\describe{
\item{\code{ellipsoidalCS}}{[\code{\link{GMLEllipsoidalCS}}]}

\item{\code{cartesianCS}}{[\code{\link{GMLCartesianCS}}]}

\item{\code{sphericalCS}}{[\code{\link{GMLSphericalCS}}]}

\item{\code{geodeticDatum}}{[\code{GMLGeodeticDatum}]}
}}

\section{Methods}{

\describe{
 \item{\code{new(xml, defaults, id)}}{
   This method is used to instantiate a GML Abstract single CRS
 }
 \item{\code{setEllipsoidalCS(cs)}}{
   Sets an ellipsoidal CS
 }
 \item{\code{setCartesianCS(cs)}}{
   Sets a cartesian CS
 }
 \item{\code{setSphericalCS(cs)}}{
   Sets a spherical CS
 }
 \item{\code{setGeodeticDatum(datum)}}{
   Sets geodetic datum
 }
}
}

\references{
ISO 19136:2007 Geographic Information -- Geographic Markup Language.
  http://www.iso.org/iso/iso_catalogue/catalogue_tc/catalogue_detail.htm?csnumber=32554 
  
  OGC Geography Markup Language. http://www.opengeospatial.org/standards/gml
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{GML}
\keyword{ISO}
\keyword{crs}
\keyword{geodetic}
