% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMLTimeInstant.R
\docType{class}
\name{GMLTimeInstant}
\alias{GMLTimeInstant}
\title{GMLTimeInstant}
\format{\code{\link{R6Class}} object.}
\usage{
GMLTimeInstant
}
\value{
Object of \code{\link{R6Class}} for modelling an GMLTimeInstant
}
\description{
GMLTimeInstant
}
\section{Fields}{

\describe{
\item{\code{timePosition}}{[\code{\link{numeric}}|\code{\link{character}}|\code{\link{Date}}|\code{\link{POSIXt}}]}
}}

\section{Methods}{

\describe{
 \item{\code{new(xml, timePosition)}}{
   This method is used to instantiate an GMLTimeInstant
 }
 \item{\code{setTimePosition(timePosition)}}{
   Sets the position (date or date and time of the resource contents), 
   as object of class "POSIXct"/"POSIXt" or "Date"
 }
}
}

\examples{
  time <- ISOdate(2000, 1, 12, 12, 59, 45)
  md <- GMLTimeInstant$new(timePosition = time)
  xml <- md$encode()

}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{instant}
\keyword{time}
