% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOAbstractCarrierOfCharacteristics.R
\docType{class}
\name{ISOAbstractCarrierOfCharacteristics}
\alias{ISOAbstractCarrierOfCharacteristics}
\title{ISOAbstractCarrierOfCharacteristics}
\format{\code{\link{R6Class}} object.}
\usage{
ISOAbstractCarrierOfCharacteristics
}
\value{
Object of \code{\link{R6Class}} for modelling an abstract ISOCarrierOfCharacteristics
}
\description{
ISOAbstractCarrierOfCharacteristics
}
\section{Fields}{

\describe{
\item{\code{featureType}}{[\code{\link{ISOFeatureType}}]}

\item{\code{constrainedBy}}{[\code{\link{ISOConstraint}}]}
}}

\note{
abstract class
}
\section{Methods}{

\describe{
 \item{\code{new(xml, defaults)}}{
   This method is used to instantiate an \code{\link{ISOAbstractCarrierOfCharacteristics}}
 }
 \item{\code{setFeatureType(featureType)}}{
   Set a feature type, object of class \code{\link{ISOFeatureType}}
 }
 \item{\code{addConstraint(constraint)}}{
   Add constraint, object of class \code{\link{ISOConstraint}}
 }
 \item{\code{delConstraint(constraint)}}{
   Deletes constraint, object of class \code{\link{ISOConstraint}}
 }
}
}

\references{
ISO 19110:2005 Methodology for Feature cataloguing
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{abstract}
\keyword{carrierOfCharacteristics}
