% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOAbstractGenericName.R
\docType{class}
\name{ISOAbstractGenericName}
\alias{ISOAbstractGenericName}
\title{ISOAbstractGenericName}
\format{\code{\link{R6Class}} object.}
\usage{
ISOAbstractGenericName
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO abstract GenericName
}
\description{
ISOAbstractGenericName
}
\section{Fields}{

\describe{
\item{\code{value}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(xml,value)}}{
   This method is used to instantiate an ISOLocalName
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\references{
ISO/TS 19103:2005 Geographic information -- Conceptual schema language
}
\keyword{ISO}
\keyword{abstract}
\keyword{generic}
\keyword{name}

