% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOAbstractObject.R
\docType{class}
\name{ISOAbstractObject}
\alias{ISOAbstractObject}
\title{ISOAbstractObject}
\format{\code{\link{R6Class}} object.}
\usage{
ISOAbstractObject
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO Metadata Element
}
\description{
ISOAbstractObject
}
\note{
Abstract ISO Metadata class used internally by geometa
}
\section{Static Methods}{

\describe{
 \item{\code{getISOClassByNode(node)}}{
   Inherit the ISO class matching an XML document or node
 }
 \item{\code{compare(metadataElement1, metadataElement2)}}{
   Compares two metadata elements objects. Returns TRUE if they are equal,
   FALSE otherwise. The comparison of object is done by comparing the XML
   representation of the objects (since no R6 object comparison method seems
   to exist)
 }
}
}

\section{Abstract Methods}{

\describe{
 \item{\code{new(xml, element, namespace, defaults, attrs)}}{
   This method is used to instantiate an ISOAbstractObject
 }
 \item{\code{INFO(text)}}{
   Logger to report information. Used internally
 }
 \item{\code{WARN(text)}}{
   Logger to report warnings. Used internally
 }
 \item{\code{ERROR(text)}}{
   Logger to report errors. Used internally
 }
 \item{\code{print()}}{
   Provides a custom print output (as tree) of the current class
 }
 \item{\code{decode(xml)}}{
   Decodes a ISOMetadata* R6 object from XML representation
 }
 \item{\code{encode(addNS, validate, strict)}}{
   Encodes a ISOMetadata* R6 object to XML representation. By default, namespace
   definition will be added to XML root (\code{addNS = TRUE}), and validation
   of object will be performed (\code{validate = TRUE}) prior to its XML encoding.
   The argument \code{strict} allows to stop the encoding in case object is not
   valid, with a default value set to \code{FALSE}.
 }
 \item{\code{validate(xml, strict)}}{
   Validates the encoded XML against ISO 19139 XML schemas. If \code{strict} is
   \code{TRUE}, a error will be raised. Default is \code{FALSE}.
 }
 \item{\code{getNamespaceDefinition(recursive)}}{
   Gets the namespace definition of the current ISO* class. By default, only
   the namespace definition of the current element is retrieved (\code{recursive = FALSE}).
 }
 \item{\code{getClassName()}}{
   Gets the class name
 }
 \item{\code{wrapBaseElement(field, fieldObj)}}{
   Wraps a base element type
 }
 \item{\code{contains(field, metadataElement)}}{
   Indicates of the present class object contains an metadata element object
   for a particular list-based field.
 }
 \item{\code{addListElement(field, metadataElement)}}{
   Adds a metadata element to a list-based field. Returns TRUE if the element
   has been added, FALSE otherwise. In case an element is already added, the
   element will not be added and this method will return FALSE.
 }
 \item{\code{delListElement(field, metadataElement)}}{
   Deletes a metadata element from a list-based field. Returns TRUE if the element
   has been deleted, FALSE otherwise. In case an element is abstent, this method
   will return FALSE.
 }
 \item{\code{setAttr(attrKey, attrValue)}}{
   Set an attribute
 }
 \item{\code{setId(id, addNS)}}{
   Set an id. By default \code{addNS} is \code{FALSE} (no namespace prefix added).
 }
 \item{\code{setHref(href)}}{
   Sets an href reference
 }
 \item{\code{setCodeList(codeList)}}{
   Sets a codeList
 }
 \item{\code{setCodeListValue(codeListValue)}}{
   Sets a codeList value
 }
 \item{\code{setCodeSpace(codeSpace)}}{
   Set a codeSpace
 }
 \item{\code{setValue(value)}}{
   Set a value
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{element}
\keyword{metadata}

