% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOAbstractReferenceSystem.R
\docType{class}
\name{ISOAbstractReferenceSystem}
\alias{ISOAbstractReferenceSystem}
\title{ISOAbstractReferenceSystem}
\format{\code{\link{R6Class}} object.}
\usage{
ISOAbstractReferenceSystem
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO abstract RS Reference system
}
\description{
ISOAbstractReferenceSystem
}
\note{
abstract class
}
\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to instantiate an ISOAbstractReferenceSystem
 }
 \item{\code{setName(name)}}{
   Set the name, object of class \code{ISOReferenceIdentifier}
 }
 \item{\code{addDomainOfValidity(domainOfValidity)}}{
   Adds a domain of validity, object of class \code{ISOExtent}
 }
 \item{\code{delDomainOfValidity(domainOfValidity)}}{
   Deletes a domain of validity, object of class \code{ISOExtent}
 }
}
}

\references{
ISO 19115:2003 - Geographic information -- Metadata
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{abstract}
\keyword{reference}
\keyword{system}
