% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOAssociationRole.R
\docType{class}
\name{ISOAssociationRole}
\alias{ISOAssociationRole}
\title{ISOAssociationRole}
\format{\code{\link{R6Class}} object.}
\usage{
ISOAssociationRole
}
\value{
Object of \code{\link{R6Class}} for modelling an ISOAssociationRole
}
\description{
ISOAssociationRole
}
\section{Fields}{

\describe{
\item{\code{type}}{}

\item{\code{isOrdered}}{}

\item{\code{isNavigable}}{}

\item{\code{relation}}{}

\item{\code{valueType}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to instantiate an ISOAssociationRole
 }
 \item{\code{setType(type)}}{
  Set the type of association
 }
 \item{\code{setIsOrdered(isOrdered)}}{
  Sets TRUE/FALSE if ordered
 }
 \item{\code{setIsNavigable(isNavigable)}}{
  Sets TRUE/FALSE if navigable
 }
 \item{\code{setRelation(relation)}}{
  Sets the relation, objec of class \code{ISOFeatureAssociation}
 }
 \item{\code{setValueType(valueType)}}{
  Sets a featureType, object of class \code{ISOFeatureType}
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\references{
ISO 19110:2005 Methodology for Feature cataloguing
}
\keyword{ISO}
\keyword{association}
\keyword{role}

