% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOAssociationRole.R
\docType{class}
\name{ISOAssociationRole}
\alias{ISOAssociationRole}
\title{ISOAssociationRole}
\format{\code{\link{R6Class}} object.}
\usage{
ISOAssociationRole
}
\value{
Object of \code{\link{R6Class}} for modelling an ISOAssociationRole
}
\description{
ISOAssociationRole
}
\section{Fields}{

\describe{
\item{\code{type}}{}

\item{\code{isOrdered}}{}

\item{\code{isNavigable}}{}

\item{\code{relation}}{}

\item{\code{rolePlayer}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to instantiate an ISOAssociationRole
 }
 \item{\code{setRoleType(roleType)}}{
   Sets the role type, object of class \code{ISORoleType} or any \code{character}
   value among \code{ISORoleType$values()}.
 }
 \item{\code{setIsOrdered(isOrdered)}}{
   Sets \code{TRUE} if ordered, \code{FALSE} otherwise
 }
 \item{\code{setIsNavigable(isNavigable)}}{
   Sets \code{TRUE} if navigable, \code{FALSE} otherwise
 }
 \item{code{setRelation(relation)}}{
   Sets an object of class \code{ISOFeatureAssocation} as relation
 }
 \item{\code{addRolePlayer(rolePlayer)}}{
   Adds a role player, object of class \code{ISOFeatureType}
 }
 \item{\code{delRolePlayer(rolePlayer)}}{
   Deletes a role player, object of class \code{ISOFeatureType}
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\references{
ISO 19110:2005 Methodology for Feature cataloguing
}
\keyword{ISO}
\keyword{association}
\keyword{role}

