% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOAttributes.R
\docType{class}
\name{ISOAttributes}
\alias{ISOAttributes}
\title{ISOAttributes}
\format{\code{\link{R6Class}} object.}
\usage{
ISOAttributes
}
\value{
Spatial object of \code{\link{R6Class}} for modelling a list of ISO xml attributes
}
\description{
ISOAttributes
}
\section{Methods}{

\describe{
 \item{\code{new(...)}}{
   This method is used to instantiate a vector of attributes to be used
   for empty element properties.
 }
}
}

\examples{
  attrs <- ISOAttributes$new(href = "http://somelink", title = "sometitle")

}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{attributes}
