% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOBaseCharacterString.R
\docType{class}
\name{ISOBaseCharacterString}
\alias{ISOBaseCharacterString}
\title{ISOBaseCharacterString}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO BaseCharacterString
}
\description{
ISOBaseCharacterString
}
\section{Fields}{

\describe{
\item{\code{value}}{}
}}

\note{
Class used by geometa internal XML decoder/encoder
}
\section{Methods}{

\describe{
 \item{\code{new(xml,value)}}{
   This method is used to instantiate an ISOBaseCharacterString
 }
}
}

\references{
ISO/TS 19103:2005 Geographic information -- Conceptual schema language
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{character}
\keyword{string}
