% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOBaseDecimal.R
\docType{class}
\name{ISOBaseDecimal}
\alias{ISOBaseDecimal}
\title{ISOBaseDecimal}
\format{\code{\link{R6Class}} object.}
\usage{
ISOBaseDecimal
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO Decimal
}
\description{
ISOBaseDecimal
}
\section{Fields}{

\describe{
\item{\code{value}}{}
}}
\note{
Class used by geometa internal XML decoder/encoder
}
\section{Methods}{

\describe{
 \item{\code{new(xml,value)}}{
   This method is used to instantiate an ISOBaseDecimal
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{decimal}

