% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOBaseInteger.R
\docType{class}
\name{ISOBaseInteger}
\alias{ISOBaseInteger}
\title{ISOBaseInteger}
\format{\code{\link{R6Class}} object.}
\usage{
ISOBaseInteger
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO Integer
}
\description{
ISOBaseInteger
}
\section{Fields}{

\describe{
\item{\code{value}}{}
}}
\note{
Class used by geometa internal XML decoder/encoder
}
\section{Methods}{

\describe{
 \item{\code{new(xml,value)}}{
   This method is used to instantiate an ISOBaseInteger
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{integer}

