% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOBinding.R
\docType{class}
\name{ISOBinding}
\alias{ISOBinding}
\title{ISOBinding}
\format{\code{\link{R6Class}} object.}
\usage{
ISOBinding
}
\value{
Object of \code{\link{R6Class}} for modelling an ISOBinding
}
\description{
ISOBinding
}
\section{Fields}{

\describe{
\item{\code{description}}{[\code{\link{character}}] description}

\item{\code{globalProperty}}{[\code{\link{ISOPropertyType}}] property type}
}}

\section{Methods}{

\describe{
 \item{\code{new(xml, defaults)}}{
   This method is used to instantiate an \code{\link{ISOBinding}}
 }
 \item{\code{setDescription(description, locales)}}{
   Set description of inheritance relation. Locale names can be specified 
   as \code{list} with the \code{locales} argument.
 }
 \item{\code{setPropertyType(propertyType)}}{
   Set global property, object of class \code{\link{ISOPropertyType}}
 }
}
}

\references{
ISO 19110:2005 Methodology for Feature cataloguing
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{Binding}
\keyword{ISO}
