% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOBoundAssociationRole.R
\docType{class}
\name{ISOBoundAssociationRole}
\alias{ISOBoundAssociationRole}
\title{ISOBoundAssociationRole}
\format{\code{\link{R6Class}} object.}
\usage{
ISOBoundAssociationRole
}
\value{
Object of \code{\link{R6Class}} for modelling an ISOBoundAssociationRole
}
\description{
ISOBoundAssociationRole
}
\section{Fields}{

\describe{
\item{\code{rolePlayer}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(xml, defaults)}}{
   This method is used to instantiate an ISOBoundAssociationRole
 }
 \item{\code{setFeatureType(featureType)}}{
   Set feature type, object of class \code{ISOFeatureType}
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\references{
ISO 19110:2005 Methodology for Feature cataloguing
}
\keyword{BoundAssociationRole}
\keyword{ISO}

