% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOCarrierOfCharacteristics.R
\docType{class}
\name{ISOCarrierOfCharacteristics}
\alias{ISOCarrierOfCharacteristics}
\title{ISOCarrierOfCharacteristics}
\format{\code{\link{R6Class}} object.}
\usage{
ISOCarrierOfCharacteristics
}
\value{
Object of \code{\link{R6Class}} for modelling an ISOCarrierOfCharacteristics
}
\description{
ISOCarrierOfCharacteristics
}
\section{Methods inherited from \code{\link{ISOAbstractCarrierOfCharacteristics}}}{

\describe{
 \item{\code{setFeatureType(featureType)}}{
   Set a feature type, object of class \code{\link{ISOFeatureType}}
 }
 \item{\code{addConstraint(constraint)}}{
   Add constraint, object of class \code{\link{ISOConstraint}}
 }
 \item{\code{delConstraint(constraint)}}{
   Deletes constraint, object of class \code{\link{ISOConstraint}}
 }
}
}

\section{Methods}{

\describe{
 \item{\code{new(xml, defaults)}}{
   This method is used to instantiate an \code{\link{ISOCarrierOfCharacteristics}}
 }
}
}

\references{
ISO 19110:2005 Methodology for Feature cataloguing
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{carrierOfCharacteristics}
