% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOCharacterSet.R
\docType{class}
\name{ISOCharacterSet}
\alias{ISOCharacterSet}
\title{ISOCharacterSet}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO CharacterSet
}
\description{
ISOCharacterSet
}
\section{Methods}{

\describe{
 \item{\code{new(xml,value)}}{
   This method is used to instantiate an \code{\link{ISOCharacterSet}}
 }
}
}

\examples{
  #possible values
  values <- ISOCharacterSet$values(labels = TRUE)
  
  #some charset
  charset <- ISOCharacterSet$new(value = "utf8")

}
\references{
ISO 19115:2003 - Geographic information -- Metadata
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{characterSet}
\keyword{charset}
