% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOClassification.R
\docType{class}
\name{ISOClassification}
\alias{ISOClassification}
\title{ISOClassification}
\format{\code{\link{R6Class}} object.}
\usage{
ISOClassification
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO Classification
}
\description{
ISOClassification
}
\section{Fields}{

\describe{
\item{\code{value}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(xml,value)}}{
   This method is used to instantiate an ISOClassification
 }
}
}
\examples{
#possible values
 values <- ISOClassification$values(labels = TRUE)

 #restricted classification
 cl <- ISOClassification$new(value = "restricted")
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{Classification}
\keyword{ISO}

