% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOCodelist.R
\docType{class}
\name{ISOCodelist}
\alias{ISOCodelist}
\title{ISOCodelist}
\format{\code{\link{R6Class}} object.}
\usage{
ISOCodelist
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO codelist
}
\description{
ISOCodelist
}
\section{Fields}{

\describe{
\item{\code{value}}{}
}}

\note{
Class used by geometa internal codelist XML decoder/encoder
}
\section{Methods}{

\describe{
 \item{\code{new(refFile, id)}}{
   This method is used to instantiate an ISOCodelist
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{codelist}
