% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOCodeListValue.R
\docType{class}
\name{ISOCodeListValue}
\alias{ISOCodeListValue}
\title{ISOCodeListValue}
\format{\code{\link{R6Class}} object.}
\usage{
ISOCodeListValue
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO Metadata codelist element
}
\description{
ISOCodeListValue
}
\section{Fields}{

\describe{
\item{\code{codelistId}}{[\code{\link{character}}] codelist identifier}

\item{\code{attrs}}{[\code{\link{list}}] list of codelist attributes}

\item{\code{value}}{[\code{\link{character}}] codelist value}

\item{\code{valueDescription}}{[\code{\link{character}}] codelist value description}
}}

\note{
Abstract ISO codelist class used internally by geometa
}
\section{Methods}{

\describe{
 \item{\code{new(xml, id, value, description, addCodeListAttrs, addCodeSpaceAttr, setValue)}}{
   This method is used to instantiate an ISOCodeListValue. By default,
   \code{addCodeListAttrs = TRUE}, to add codelist atributes to root XML. The 
   parameter \code{addCodeSpaceAttr = TRUE} by default, and ignored if the valueof
   \code{addCodeLisAttrs} is set to \code{FALSE}. The argument \code{setValue}
   sets the value as node text (defaut is \code{TRUE}). The argument \code{setValueDescription}
   allows to force having description set as value, default is \code{FALSE} in which case
   the name will be preferred, and in case no name is provided, code value will be used.
   
 }
 \item{\code{getAcceptedValues()}}{
   This method allows to get the codelist accepted values
 }
}
}

\references{
ISO 19115:2003 - Geographic information -- Metadata
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{code}
\keyword{element}
