% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOConformanceResult.R
\docType{class}
\name{ISOConformanceResult}
\alias{ISOConformanceResult}
\title{ISOConformanceResult}
\format{\code{\link{R6Class}} object.}
\usage{
ISOConformanceResult
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO ConformanceResult
}
\description{
ISOConformanceResult
}
\section{Fields}{

\describe{
\item{\code{result}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(xml,value)}}{
   This method is used to instantiate an ISOConformanceResult
 }
 \item{\code{setSpecification(specification)}}{
   Sets the specification (an ISOCitation object)
 }
 \item{\code{setExplanation(explanation)}}{
   Sets the explanation
 }
 \item{\code{setPass(pass)}}{
   Sets if passing the conformance or not (logical value)
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{conformance}
\keyword{result}

