% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOConstraints.R
\docType{class}
\name{ISOConstraints}
\alias{ISOConstraints}
\title{ISOConstraints}
\format{\code{\link{R6Class}} object.}
\usage{
ISOConstraints
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO abstract Constraints
}
\description{
ISOConstraints
}
\section{Fields}{

\describe{
\item{\code{useLimitation}}{}
}}
\note{
Abstract ISO class
}
\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to instantiate an ISOLegalConstraints
 }
 \item{\code{addUseLimitation(useLimitation)}}{
   Adds a use limitation
 }
 \item{\code{setUseLimitation(useLimitation)}}{
   Sets a use limitation
 }
 \item{\code{delUseLimitation(useLimitation)}}{
   Deletes a use limitation
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{constraints}

