% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOCoupledResource.R
\docType{class}
\name{ISOCoupledResource}
\alias{ISOCoupledResource}
\title{ISOCoupledResource}
\format{\code{\link{R6Class}} object.}
\usage{
ISOCoupledResource
}
\value{
Object of \code{\link{R6Class}} for modelling an ISOCoupledResource
}
\description{
ISOCoupledResource
}
\section{Fields}{

\describe{
\item{\code{identifier}}{}
}}

\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to instantiate an ISOCoupledResource
 }
 \item{\code{setOperationName(operationName)}}{
   Set the operation name
 }
 \item{\code{setIdentifier(identifier)}}{
   Set the identifier
 }
}
}

\examples{
  md <- ISOCoupledResource$new()
  md$setOperationName("name")
  md$setIdentifier("identifier")
  xml <- md$encode()

}
\references{
ISO 19119:2005 - Geographic information -- Services
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{CoupledResource}
\keyword{ISO}
