% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOCouplingType.R
\docType{class}
\name{ISOCouplingType}
\alias{ISOCouplingType}
\title{ISOCouplingType}
\format{\code{\link{R6Class}} object.}
\usage{
ISOCouplingType
}
\value{
Object of \code{\link{R6Class}} for modelling an ISOCouplingType
}
\description{
ISOCouplingType
}
\section{Fields}{

\describe{
\item{\code{value}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(xml,value, description)}}{
   This method is used to instantiate an ISOCouplingType
 }
}
}
\examples{
#possible values
  values <- ISOCouplingType$values(labels = TRUE)

  #couplingType
  couplingType <- ISOCouplingType$new(value = "loose")
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\references{
ISO 19119:2005 - Geographic information -- Services
}
\keyword{ISO}
\keyword{coupling}
\keyword{type}

