% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISODCPList.R
\docType{class}
\name{ISODCPList}
\alias{ISODCPList}
\title{ISODCPList}
\format{\code{\link{R6Class}} object.}
\usage{
ISODCPList
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO DCPList
}
\description{
ISODCPList
}
\section{Fields}{

\describe{
\item{\code{value}}{}
}}

\section{Methods}{

\describe{
 \item{\code{new(xml,value, description)}}{
   This method is used to instantiate an ISODCPList
 }
}
}

\examples{
  #possible values
  values <- ISODCPList$values(labels = TRUE)
  
  #example
  javaDCP <- ISODCPList$new(value = "JAVA")
  
}
\references{
ISO 19119:2005 - Geographic information -- Service
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{DCP}
\keyword{ISO}
