% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISODigitalTransferOptions.R
\docType{class}
\name{ISODigitalTransferOptions}
\alias{ISODigitalTransferOptions}
\title{ISODigitalTransferOptions}
\format{\code{\link{R6Class}} object.}
\usage{
ISODigitalTransferOptions
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO DigitalTransferOptions
}
\description{
ISODigitalTransferOptions
}
\section{Fields}{

\describe{
\item{\code{onLine}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(xml,value)}}{
   This method is used to instantiate an ISODigitalTransferOptions
 }
 \item{\code{addOnlineResource(onlineResource)}}{
   Adds an online resource
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{distribution}

