% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISODimensionNameType.R
\docType{class}
\name{ISODimensionNameType}
\alias{ISODimensionNameType}
\title{ISODimensionNameType}
\format{\code{\link{R6Class}} object.}
\usage{
ISODimensionNameType
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO DimensionNameType
}
\description{
ISODimensionNameType
}
\section{Fields}{

\describe{
\item{\code{value}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(xml,value, description)}}{
   This method is used to instantiate an ISODimensionNameType
 }
}
}
\examples{
#possible values
  values <- ISODimensionNameType$values(labels = TRUE)

  #row DimensionNameType
  rowType <- ISODimensionNameType$new(value = "row")
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\references{
ISO 19115:2003 - Geographic information -- Metadata
}
\keyword{DimensionNameType}
\keyword{ISO}

