% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISODistance.R
\docType{class}
\name{ISODistance}
\alias{ISODistance}
\title{ISODistance}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO Distance measure
}
\description{
ISODistance
}
\section{Methods}{

\describe{
 \item{\code{new(xml,value, uom, useUomURI)}}{
   This method is used to instantiate an ISODistance. The \code{uom} argument represents
   the symbol of unit of measure used. The parameter  \code{useUomURI} can be used to 
   set the uom as URI, its default value is \code{FALSE}.
 }
}
}

\references{
ISO/TS 19103:2005 Geographic information -- Conceptual schema language
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{distance}
\keyword{length}
\keyword{measure}
