% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISODistribution.R
\docType{class}
\name{ISODistribution}
\alias{ISODistribution}
\title{ISODistribution}
\format{\code{\link{R6Class}} object.}
\usage{
ISODistribution
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO Distribution
}
\description{
ISODistribution
}
\section{Fields}{

\describe{
\item{\code{transferOptions}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(xml,value)}}{
   This method is used to instantiate an ISODistribution
 }
 \item{\code{setDigitalTransferOptions(options)}}{
   Sets the digital transfer options
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{distribution}

