% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISODomainConsistency.R
\docType{class}
\name{ISODomainConsistency}
\alias{ISODomainConsistency}
\title{ISODomainConsistency}
\format{\code{\link{R6Class}} object.}
\usage{
ISODomainConsistency
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO DomainConsistency
}
\description{
ISODomainConsistency
}
\section{Fields}{

\describe{
\item{\code{result}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(xml,value)}}{
   This method is used to instantiate an ISODomainConsistency
 }
 \item{\code{addResult(result)}}{
   Adds a result
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{consistency}
\keyword{domain}

