% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOElementSequence.R
\docType{class}
\name{ISOElementSequence}
\alias{ISOElementSequence}
\title{ISOElementSequence}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an ISOElementSequence
}
\description{
ISOElementSequence
}
\note{
This class is used internally by geometa to deal with simple type not
handled by proper class element. e.g. \code{name} property of \code{ISOParameter}
class from ISO 19119:2005
}
\section{Methods}{

\describe{
 \item{\code{new(xml, ...)}}{
   This method is used to instantiate an element sequence
 }
}
}

\references{
ISO/TS 19103:2005 Geographic information -- Conceptual schema language
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{record}
