% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOEvaluationMethodType.R
\docType{class}
\name{ISOEvaluationMethodType}
\alias{ISOEvaluationMethodType}
\title{ISOEvaluationMethodType}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO EvaluationMethodType
}
\description{
ISOEvaluationMethodType
}
\section{Methods}{

\describe{
 \item{\code{new(xml,value, description)}}{
   This method is used to instantiate an \code{\link{ISOEvaluationMethodType}}
 }
}
}

\examples{
  #possible values
  values <- ISOEvaluationMethodType$values(labels = TRUE)
  
  #example of EvaluationMethodType
  indirect <- ISOEvaluationMethodType$new(value = "indirect")
  
}
\references{
ISO 19115:2003 - Geographic information -- Metadata
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{EvaluationMethodType}
\keyword{ISO}
