% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOExtent.R
\docType{class}
\name{ISOExtent}
\alias{ISOExtent}
\title{ISOExtent}
\format{\code{\link{R6Class}} object.}
\usage{
ISOExtent
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO Extent
}
\description{
ISOExtent
}
\section{Fields}{

\describe{
\item{\code{geographicElement}}{}

\item{\code{temporalElement}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to instantiate an ISOExtent
 }
 \item{\code{setGeographicElement(extent)}}{
   Sets the geographic element
 }
 \item{\code{setTemporalElement(extent)}}{
   Sets the temporal element
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{extent}

