% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOFeatureCatalogue.R
\docType{class}
\name{ISOFeatureCatalogue}
\alias{ISOFeatureCatalogue}
\title{ISOFeatureCatalogue}
\format{\code{\link{R6Class}} object.}
\usage{
ISOFeatureCatalogue
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO FeatureCatalogue
}
\description{
ISOFeatureCatalogue
}
\section{Fields}{

\describe{
\item{\code{name}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to instantiate an ISOFeatureCatalogue
 }
 \item{\code{setName(name)}}{
   Sets the name
 }
 \item{\code{addScope(scope)}}{
   Adds scope (object of class \code{character})
 }
 \item{\code{delScope(scope)}}{
   Deletes scope
 }
 \item{\code{addFieldOfApplication(fieldOfApplication)}}{
   Adds a field of application (object of class \code{character})
 }
 \item{\code{delFieldOfApplication(fieldOfApplication)}}{
   Deletes fieldOfApplication
 }
 \item{\code{setVersionNumber(versionNumber)}}{
   Sets version number (object of class \code{character})
 }
 \item{\code{setVersionDate(versionDate)}}{
   Sets version date
 }
 \item{\code{setProducer(producer)}}{
   Sets an object of class \code{ISOResponsibleParty} as producer
 }
 \item{\code{setFunctionalLanguage(functionalLanguage)}}{
   Sets the functional language
 }
 \item{\code{addFeatureType(featureType)}}{
   Adds a feature type, object of class \code{ISOFeatureType}
 }
 \item{\code{delFeatureType(featureType)}}{
   Deletes a feature type, object of class \code{ISOFeatureType}
 }
 \item{\code{addDefinitionSource(source)}}{
   Adds a definition source, object of class \code{ISODefinitionSource} or
   \code{ISOCitation}
 }
 \item{\code{delDefinitionSource(source)}}{
   Deletes a definition source, object of class \code{ISODefinitionSource} or
   \code{ISOCitation}
 }
}
}
\examples{
fc <- ISOFeatureCatalogue$new()
 fc$setName("name")
 fc$addScope("scope1")
 fc$addScope("scope2")
 fc$addFieldOfApplication("field1")
 fc$addFieldOfApplication("field2")
 fc$setVersionNumber("1.0")
 fc$setVersionDate(ISOdate(2015, 1, 1, 1))

 producer <- ISOResponsibleParty$new()
 producer$setIndividualName("someone")
 fc$setProducer(producer)
 fc$setFunctionalLanguage("eng")

 cit <- ISOCitation$new()
 cit$setTitle("some citation title")
 fc$addDefinitionSource(cit)

 xml <- fc$encode()
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\references{
ISO 19110:2005 Methodology for Feature cataloguing
}
\keyword{ISO}
\keyword{catalogue}
\keyword{feature}

