% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOFeatureCatalogue.R
\docType{class}
\name{ISOFeatureCatalogue}
\alias{ISOFeatureCatalogue}
\title{ISOFeatureCatalogue}
\format{\code{\link{R6Class}} object.}
\usage{
ISOFeatureCatalogue
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO FeatureCatalogue
}
\description{
ISOFeatureCatalogue
}
\section{Fields}{

\describe{
\item{\code{name}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to instantiate an ISOFeatureCatalogue
 }
 \item{\code{setName(name)}}{
   Sets the name
 }
}
}
\examples{
fc <- ISOFeatureCatalogue$new()
 xml <- fc$encode()
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{catalogue}
\keyword{feature}

