% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOFeatureType.R
\docType{class}
\name{ISOFeatureType}
\alias{ISOFeatureType}
\title{ISOFeatureType}
\format{\code{\link{R6Class}} object.}
\usage{
ISOFeatureType
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO FeatureType
}
\description{
ISOFeatureType
}
\section{Fields}{

\describe{
\item{\code{typeName}}{}

\item{\code{definition}}{}

\item{\code{code}}{}

\item{\code{isAbstract}}{}

\item{\code{aliases}}{}

\item{\code{inheritsFrom}}{}

\item{\code{inheritsTo}}{}

\item{\code{featureCatalogue}}{}

\item{\code{carrierOfCharacteristics}}{}

\item{\code{constrainedBy}}{}

\item{\code{definitionResource}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to instantiate an ISOFeatureType
 }
 \item{\code{setTypeName(typeName)}}{
   Sets the type name. Object of class \code{ISOLocalName} or \code{"character"}
 }
 \item{\code{setDefinition(definition)}}{
   Sets the definition
 }
 \item{\code{setCode(code)}}{
   Sets the code
 }
 \item{\code{setIsAbstract(isAbstract)}}{
   Sets TRUE/FALSE if the feature type is abstract or not
 }
 \item{\code{addAlias(alias)}}{
   Set alias name. Object of class \code{ISOLocalName} or \code{"character"}
 }
 \item{\code{delAlias(alias)}}{
   Deletes alias name
 }
 \item{\code{setFeatureCatalogue(fc)}}{
   Sets a feature catalogue, object of class \code{ISOFeatureCatalogue}
 }
 \item{\code{addCharacteristic(characteristic)}}{
   Adds a characteristic as object of class \code{ISOPropertyType} or subclass
 }
 \item{\code{delCharacteristic(characteristic)}}{
   Deletes a characteristic as object of class \code{ISOPropertyType} or subclass
 }
 \item{\code{addConstraint(constraint)}}{
   Adds a constraint, object of class \code{ISOConstraint} or \code{character}
 }
 \item{\code{delConstraint(constraint)}}{
   Deletes a constraint, object of class \code{ISOConstraint} or \code{character}
 }
 \item{\code{setDefinitionReference(definitionReference)}}{
   Sets the definition Reference, object of class \code{ISODefinitionReference}
 }
}
}
\examples{
ft <- ISOFeatureType$new()
 ft$setTypeName("typeName")
 ft$setDefinition("definition")
 ft$setCode("code")
 ft$setIsAbstract(FALSE)
 ft$addAlias("alias1")
 ft$addAlias("alias2")
 xml <- ft$encode()
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\references{
ISO 19110:2005 Methodology for Feature cataloguing
}
\keyword{ISO}
\keyword{feature}
\keyword{type}

