% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOFileName.R
\docType{class}
\name{ISOFileName}
\alias{ISOFileName}
\title{ISOFileName}
\format{\code{\link{R6Class}} object.}
\usage{
ISOFileName
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO FileName
}
\description{
ISOFileName
}
\section{Methods}{

\describe{
 \item{\code{new(xml, file, name)}}{
   This method is used to instantiate an ISOFileName
 }
}
}
\examples{
md <- ISOFileName$new(file = "someuri", name = "filename")
  xml <- md$encode()
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\references{
ISO/TS 19139:2007 Geographic information -- XML
}
\keyword{ISO}
\keyword{file}
\keyword{name}

