% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOFreeText.R
\docType{class}
\name{ISOFreeText}
\alias{ISOFreeText}
\title{ISOFreeText}
\format{\code{\link{R6Class}} object.}
\usage{
ISOFreeText
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO FreeText
}
\description{
ISOFreeText
}
\section{Fields}{

\describe{
\item{\code{textGroup}}{[\code{\link{ISOLocalisedCharacterString}}] localised string(s)}
}}

\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to instantiate an ISOFreeText
 }
 \item{\code{addTextGroup(textGroup)}}{
   Add a text group, object of class \code{\link{ISOLocalisedCharacterString}}
 }
 \item{\code{delTextGroup(textGroup)}}{
   Deletes a text group, object of class \code{\link{ISOLocalisedCharacterString}}
 }
}
}

\examples{
  ft <- ISOFreeText$new()
  
}
\references{
ISO 19115:2003 - Geographic information -- Metadata
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{freeText}
