% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOGeographicBoundingBox.R
\docType{class}
\name{ISOGeographicBoundingBox}
\alias{ISOGeographicBoundingBox}
\title{ISOGeographicBoundingBox}
\format{\code{\link{R6Class}} object.}
\usage{
ISOGeographicBoundingBox
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO GeographicBoundingBox
}
\description{
ISOGeographicBoundingBox
}
\section{Fields}{

\describe{
\item{\code{westBoundLongitude}}{}

\item{\code{eastBoundLongitude}}{}

\item{\code{southBoundLatitude}}{}

\item{\code{northBoundLatitude}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(xml, minx, miny, maxx, maxy)}}{
   This method is used to instantiate an ISOGeographicBoundingBox
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{extent}

