% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOGeographicDescription.R
\docType{class}
\name{ISOGeographicDescription}
\alias{ISOGeographicDescription}
\title{ISOGeographicDescription}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO GeographicDescription
}
\description{
ISOGeographicDescription
}
\section{Fields}{

\describe{
\item{\code{geographicIdentifier}}{}
}}

\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to instantiate an ISOGeographicDescription
 }
}
}

\examples{
  md <- ISOGeographicDescription$new()
  md$setGeographicIdentifier(ISOMetaIdentifier$new(code = "identifier"))
  xml <- md$encode()

}
\references{
ISO 19115:2003 - Geographic information -- Metadata
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{description}
\keyword{geographic}
