% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOGeographicExtent.R
\docType{class}
\name{ISOGeographicExtent}
\alias{ISOGeographicExtent}
\title{ISOGeographicExtent}
\format{\code{\link{R6Class}} object.}
\usage{
ISOGeographicExtent
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO abstract geographicExtent
}
\description{
ISOGeographicExtent
}
\section{Fields}{

\describe{
\item{\code{extentTypeCode}}{}
}}
\note{
abstract class
}
\section{Methods}{

\describe{
 \item{\code{new(xml, element, namespace, defaults)}}{
   This method is used to instantiate an ISOGeographicExtent
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\references{
ISO 19115:2003 - Geographic information -- Metadata
}
\keyword{ISO}
\keyword{extent}

