% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOGeometricObjectType.R
\docType{class}
\name{ISOGeometricObjectType}
\alias{ISOGeometricObjectType}
\title{ISOGeometricObjectType}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO GeometricObjectType
}
\description{
ISOGeometricObjectType
}
\section{Methods}{

\describe{
 \item{\code{new(xml,value, description)}}{
   This method is used to instantiate an \code{\link{ISOGeometricObjectType}}
 }
}
}

\examples{
  #possible values
  values <- ISOGeometricObjectType$values(labels = TRUE)
  
  #point type
  pt <- ISOGeometricObjectType$new(value = "point")

}
\references{
ISO 19115:2003 - Geographic information -- Metadata
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{geometric}
\keyword{object}
\keyword{type}
