% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOHierarchyLevel.R
\docType{class}
\name{ISOHierarchyLevel}
\alias{ISOHierarchyLevel}
\title{ISOHierarchyLevel}
\format{\code{\link{R6Class}} object.}
\usage{
ISOHierarchyLevel
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO HierarchyLevel
}
\description{
ISOHierarchyLevel
}
\section{Fields}{

\describe{
\item{\code{value}}{}
}}

\section{Methods}{

\describe{
 \item{\code{new(xml,value, description)}}{
   This method is used to instantiate an ISOHierarchyLevel
 }
}
}

\examples{
  #possible values
  values <- ISOHierarchyLevel$values(labels = TRUE)
  
  #dataset scope
  ds <- ISOHierarchyLevel$new(value = "dataset")

}
\references{
ISO 19115:2003 - Geographic information -- Metadata
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{hierarchyLevel}
