% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOIdentifier.R
\docType{class}
\name{ISOIdentifier}
\alias{ISOIdentifier}
\title{ISOIdentifier}
\format{\code{\link{R6Class}} object.}
\usage{
ISOIdentifier
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO Identifier
}
\description{
ISOIdentifier
}
\section{Fields}{

\describe{
\item{\code{code}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(xml, code)}}{
   This method is used to instantiate an ISOIdentifier
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{identifier}

