% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOIdentifier.R
\docType{class}
\name{ISOIdentifier}
\alias{ISOIdentifier}
\title{ISOIdentifier}
\format{\code{\link{R6Class}} object.}
\usage{
ISOIdentifier
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO Identifier
}
\description{
ISOIdentifier
}
\section{Fields}{

\describe{
\item{\code{authority}}{}

\item{\code{code}}{}
}}
\note{
Abstract ISO class
}
\section{Methods}{

\describe{
 \item{\code{new(xml, prefix, code, codeSpace)}}{
   This method is used to instantiate an ISOIdentifier
 }
 \item{\code{setAuthority(authority)}}{
   Sets an authority object of class \code{ISOCitation}
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\references{
ISO 19115:2003 - Geographic information -- Metadata
}
\keyword{ISO}
\keyword{identifier}

