% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOImageryGeoreferenceable.R
\docType{class}
\name{ISOImageryGeoreferenceable}
\alias{ISOImageryGeoreferenceable}
\title{ISOImageryGeoreferenceable}
\format{\code{\link{R6Class}} object.}
\usage{
ISOImageryGeoreferenceable
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO imagery Georeferenceable
}
\description{
ISOImageryGeoreferenceable
}
\section{Fields}{

\describe{
\item{\code{geolocationInformation}}{[\code{list} of \code{\link{ISOImageryAbstractGeolocationInformation}}]}
}}

\section{Methods}{

\describe{
 \item{\code{new(xml,value)}}{
   This method is used to instantiate an \code{\link{ISOImageryGeoreferenceable}}
 }
 \item{\code{addGeolocationInformation(geolocationInformation)}}{
   Adds a geolocation information, an object of class \code{\link{ISOImageryAbstractGeolocationInformation}}
 }
 \item{\code{delGeolocationInformation(geolocationInformation)}}{
   Deletes geolocation information, an object of class \code{\link{ISOImageryAbstractGeolocationInformation}}
 }
}
}

\section{Methods inherited from \code{\link{ISOGridSpatialRepresentation}}}{

See \code{\link{ISOGridSpatialRepresentation}}
}

\section{Methods inherited from \code{\link{ISOGeoreferenceable}}}{

See \code{\link{ISOGeoreferenceable}}
}

\references{
ISO 19115-2:2009 - Geographic information -- Metadata -- Part 2: Extensions for imagery and gridded data
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{Georeferenceable}
\keyword{ISO}
\keyword{imagery}
