% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOInheritanceRelation.R
\docType{class}
\name{ISOInheritanceRelation}
\alias{ISOInheritanceRelation}
\title{ISOInheritanceRelation}
\format{\code{\link{R6Class}} object.}
\usage{
ISOInheritanceRelation
}
\value{
Object of \code{\link{R6Class}} for modelling an ISOInheritanceRelation
}
\description{
ISOInheritanceRelation
}
\section{Fields}{

\describe{
\item{\code{name}}{}

\item{\code{description}}{}

\item{\code{uniqueInstance}}{}

\item{\code{subtype}}{}

\item{\code{supertype}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(xml, defaults)}}{
   This method is used to instantiate an ISOInheritanceRelation
 }
 \item{\code{setName(name)}}{
   Set name of inheritance relation
 }
 \item{\code{setDescription(description)}}{
   Set description of inheritance relation
 }
 \item{\code{setUniqueInstance(uniqueInstance)}}{
   Set \code{TRUE} if it's a unique instance, \code{FALSE} otherwise
 }
 \item{\code{setSubtype(featureType)}}{
   Set subtype, object of class \code{ISOFeatureType}
 }
 \item{\code{setSupertype(featureType)}}{
   Set supertype, object of class \code{ISOFeatureType}
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\references{
ISO 19110:2005 Methodology for Feature cataloguing
}
\keyword{ISO}
\keyword{InheritanceRelation}

