% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOKeywords.R
\docType{class}
\name{ISOKeywords}
\alias{ISOKeywords}
\title{ISOKeywords}
\format{\code{\link{R6Class}} object.}
\usage{
ISOKeywords
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO Keywords
}
\description{
ISOKeywords
}
\section{Fields}{

\describe{
\item{\code{keyword}}{}

\item{\code{type}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to instantiate an ISOKeywords
 }
 \item{\code{addKeyword(keyword)}}{
   Adds a keyword
 }
 \item{\code{delKeyword(keyword)}}{
   Deletes a keyword
 }
 \item{\code{setKeywordType(keywordType)}}{
   Sets the keyword type
 }
 \item{\code{setThesaurusName(thesaurusName)}}{
   Sets the thesaurus name
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{keywords}

