% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOKeywords.R
\docType{class}
\name{ISOKeywords}
\alias{ISOKeywords}
\title{ISOKeywords}
\format{\code{\link{R6Class}} object.}
\usage{
ISOKeywords
}
\value{
Object of \code{\link{R6Class}} for modelling a ISO set of keywords
}
\description{
ISOKeywords
}
\section{Fields}{

\describe{
\item{\code{keyword}}{}

\item{\code{type}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to instantiate an ISOKeywords
 }
 \item{\code{addKeyword(keyword)}}{
   Adds a keyword
 }
 \item{\code{delKeyword(keyword)}}{
   Deletes a keyword
 }
 \item{\code{setKeywordType(keywordType)}}{
   Sets the keyword type
 }
 \item{\code{setThesaurusName(thesaurusName)}}{
   Sets the thesaurus name
 }
}
}
\examples{
#a basic keyword set
  md <- ISOKeywords$new()
  md$addKeyword("keyword1")
  md$addKeyword("keyword2")
  md$setKeywordType("theme")
  th <- ISOCitation$new()
  th$setTitle("General")
  md$setThesaurusName(th)
  xml <- md$encode()

  #a keyword set with anchors
  md <- ISOKeywords$new()
  kwd1 <- ISOAnchor$new(
    name = "keyword1",
    href = "http://myvocabulary.geometa/keyword1"
  )
  md$addKeyword(kwd1)
  kwd2 <- ISOAnchor$new(
    name = "keyword2",
    href = "http://myvocabulary.geometa/keyword2"
  )
  md$addKeyword(kwd2)
  md$setKeywordType("theme")
  xml <- md$encode()
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\references{
ISO 19115:2003 - Geographic information -- Metadata
}
\keyword{ISO}
\keyword{keywords}

