% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOLanguage.R
\docType{class}
\name{ISOLanguage}
\alias{ISOLanguage}
\title{ISOLanguage}
\format{\code{\link{R6Class}} object.}
\usage{
ISOLanguage
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO Language
}
\description{
ISOLanguage
}
\section{Fields}{

\describe{
\item{\code{value}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(xml,value, description)}}{
   This method is used to instantiate an ISOLanguage
 }
}
}
\examples{
#possible values
  values <- ISOLanguage$values(labels = TRUE)

  #english language
  eng <- ISOLanguage$new(value = "eng")
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\references{
ISO 19115:2003 - Geographic information -- Metadata
}
\keyword{ISO}
\keyword{language}

