% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOLegalConstraints.R
\docType{class}
\name{ISOLegalConstraints}
\alias{ISOLegalConstraints}
\title{ISOLegalConstraints}
\format{\code{\link{R6Class}} object.}
\usage{
ISOLegalConstraints
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO LegalConstraints
}
\description{
ISOLegalConstraints
}
\section{Fields}{

\describe{
\item{\code{accessConstraints}}{}

\item{\code{useConstraints}}{}

\item{\code{otherConstraints}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to instantiate an ISOLegalConstraints
 }
 \item{\code{addAccessConstraint(constraint)}}{
   Adds an access constraint, as object of class "character" or class \code{ISORestriction}.
   If an object of class "character" is specified, it must match the accepted
   values given by \code{ISORestriction$values()}.
 }
 \item{\code{delAccessConstraint(constraint)}}{
   Deletes an access constraint, as object of class "character" or class \code{ISORestriction}.
   If an object of class "character" is specified, it must match the accepted
   values given by \code{ISORestriction$values()}.
 }
 \item{\code{addUseConstraint(constraint)}}{
   Adds a use constraint, as object of class "character" or class \code{ISORestriction}.
   If an object of class "character" is specified, it must match the accepted
   values given by \code{ISORestriction$values()}.
 }
 \item{\code{delUseConstraint(constraint)}}{
   Deletes a use constraint, as object of class "character" or class \code{ISORestriction}.
   If an object of class "character" is specified, it must match the accepted
   values given by \code{ISORestriction$values()}.
 }
 \item{\code{addOtherConstraint(constraint)}}{
   Adds an other constraint as object of class "character
 }
 \item{\code{delOtherConstraint(constraint)}}{
   Deletes an other constraint as object of class "character
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{constraints}
\keyword{legal}

