% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOMaintenanceInformation.R
\docType{class}
\name{ISOMaintenanceInformation}
\alias{ISOMaintenanceInformation}
\title{ISOMaintenanceInformation}
\format{\code{\link{R6Class}} object.}
\usage{
ISOMaintenanceInformation
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO MaintenanceInformation
}
\description{
ISOMaintenanceInformation
}
\section{Fields}{

\describe{
\item{\code{maintenanceAndUpdateFrequency}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to instantiate an ISOMaintenanceInformation
 }
 \item{\code{setMaintenanceFrequency(frequency)}}{
   Sets the maintenance and update frequency
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{information}
\keyword{maintenance}

