% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOMetadata.R
\docType{class}
\name{ISOMetadata}
\alias{ISOMetadata}
\title{ISOMetadataElement}
\format{\code{\link{R6Class}} object.}
\usage{
ISOMetadata
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO Metadata Element
}
\description{
ISOMetadataElement
}
\section{Fields}{

\describe{
\item{\code{fileIdentifier}}{}

\item{\code{language}}{}

\item{\code{characterSet}}{}

\item{\code{parentIdentifier}}{}

\item{\code{hierarchyLevel}}{}

\item{\code{contact}}{}

\item{\code{dateStamp}}{}

\item{\code{metadataStandardName}}{}

\item{\code{metadataStandardVersion}}{}

\item{\code{spatialRepresentationInfo}}{}

\item{\code{referenceSystemInfo}}{}

\item{\code{identificationInfo}}{}

\item{\code{distributionInfo}}{}

\item{\code{dataQualityInfo}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to instantiate an ISOMetadata
 }
 \item{\code{setFileIdentifier(fileIdentifier)}}{
   Sets the file identifier
 }
 \item{\code{setParentIdentifier(parentIdentifier)}}{
   Sets the parentIdentifier
 }
 \item{\code{addLanguage(locale)}}{
   Adds a locale
 }
 \item{\code{setLanguage{locale}}}{
   Sets the locale
 }
 \item{\code{delLanguage(locale)}}{
   Deletes a locale
 }
 \item{\code{addCharacterSet(charset)}}{
   Adds a character set
 }
 \item{\code{setCharacterSet(charset)}}{
   Sets the character set
 }
 \item{\code{delCharacterSet(charset)}}{
   Deletes a character set
 }
 \item{\code{setHierarchyLevel(level)}}{
   Sets the hierarchy level
 }
 \item{\code{setDateStamp(date)}}{
   Sets the date stamp
 }
 \item{\code{setMetadataStandardName(name)}}{
   Sets the metadata standard name
 }
 \item{\code{setMetadataStandardVersion(version)}}{
   Sets the metadata standard version
 }
 \item{\code{addContact(contact)}}{
   Adds a contact (responsible party)
 }
 \item{\code{setSpatialRepresentationInfo(spatialRepresentationInfo)}}{
   Sets the spatial representation
 }
 \item{\code{setReferenceSystemInfo(referenceSystemInfo)}}{
   Sets the reference system
 }
 \item{\code{setIdentificationInfo(identificationInfo)}}{
   Sets the data identification
 }
 \item{\code{setDistributionInfo(distributionInfo)}}{
   Sets the distribution
 }
 \item{\code{setDataQualityInfo(dataQualityInfo)}}{
   Sets the data quality
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{element}
\keyword{metadata}

