% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOMetadata.R
\docType{class}
\name{ISOMetadata}
\alias{ISOMetadata}
\title{ISOMetadataElement}
\format{\code{\link{R6Class}} object.}
\usage{
ISOMetadata
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO Metadata Element
}
\description{
ISOMetadataElement
}
\section{Fields}{

\describe{
\item{\code{fileIdentifier}}{}

\item{\code{language}}{}

\item{\code{characterSet}}{}

\item{\code{parentIdentifier}}{}

\item{\code{hierarchyLevel}}{}

\item{\code{contact}}{}

\item{\code{dateStamp}}{}

\item{\code{metadataStandardName}}{}

\item{\code{metadataStandardVersion}}{}

\item{\code{dataSetURI}}{}

\item{\code{spatialRepresentationInfo}}{}

\item{\code{referenceSystemInfo}}{}

\item{\code{identificationInfo}}{}

\item{\code{distributionInfo}}{}

\item{\code{dataQualityInfo}}{}

\item{\code{metadataMaintenance}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to instantiate an ISOMetadata
 }
 \item{\code{setFileIdentifier(fileIdentifier)}}{
   Sets the file identifier
 }
 \item{\code{setLanguage{locale}}}{
   Sets the locale
 }
 \item{\code{setCharacterSet(charset)}}{
   Sets the character set
 }
 \item{\code{setParentIdentifier(parentIdentifier)}}{
   Sets the parentIdentifier
 }
 \item{\code{addHierarchyLevel(level)}}{
   Adds the hierarchy level
 }
 \item{\code{setHierarchyLevel(level)}}{
   Sets the hierarchy level
 }
 \item{\code{delHierarchyLevel(level)}}{
   Deletes the hierarchy level
 }
 \item{\code{addContact(contact)}}{
   Adds a contact as object of class \code{ISOResponsibleParty}
 }
 \item{\code{delContact(contact)}}{
   Deletes a contact as object of class \code{ISOResponsibleParty}
 }
 \item{\code{setDateStamp(date)}}{
   Sets the date stamp
 }
 \item{\code{setMetadataStandardName(name)}}{
   Sets the metadata standard name
 }
 \item{\code{setMetadataStandardVersion(version)}}{
   Sets the metadata standard version
 }
 \item{\code{setDataSetURI(dataSetURI)}}{
   Sets the metadata dataSet URI
 }
 \item{\code{addSpatialRepresentationInfo(spatialRepresentationInfo)}}{
   Adds a spatial representation
 }
 \item{\code{setSpatialRepresentationInfo(spatialRepresentationInfo)}}{
   Sets a spatial representation
 }
 \item{\code{delSpatialRepresentationInfo(spatialRepresentationInfo)}}{
   Deletes a spatial representation
 }
 \item{\code{addReferenceSystemInfo(referenceSystemInfo)}}{
   Adds a reference system
 }
 \item{\code{setReferenceSystemInfo(referenceSystemInfo)}}{
   Sets the reference system
 }
 \item{\code{delReferenceSystemInfo(referenceSystemInfo)}}{
   Deletes a reference system
 }
 \item{\code{addIdentificationInfo(identificationInfo)}}{
   Adds a data identification
 }
 \item{\code{setIdentificationInfo(identificationInfo)}}{
   Sets the data identification
 }
 \item{\code{delIdentificationInfo(identificationInfo)}}{
   Deletes a data identification
 }
 \item{\code{setDistributionInfo(distributionInfo)}}{
   Sets the distribution
 }
 \item{\code{addDataQualityInfo(dataQualityInfo)}}{
   Adds a data quality
 }
 \item{\code{setDataQualityInfo(dataQualityInfo)}}{
   Sets the data quality
 }
 \item{\code{delDataQualityInfo(dataQualityInfo)}}{
   Deletes a data quality
 }
 \item{\code{setMetadataMaintenance(metadataMaintenance)}}{
   Sets a metadata maintenance as object of class \code{ISOMaintenanceInformation}
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{element}
\keyword{metadata}

