% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOMetadataCodelistElement.R
\docType{class}
\name{ISOMetadataCodelistElement}
\alias{ISOMetadataCodelistElement}
\title{ISOMetadataCodelistElement}
\format{\code{\link{R6Class}} object.}
\usage{
ISOMetadataCodelistElement
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO Metadata codelist element
}
\description{
ISOMetadataCodelistElement
}
\section{Fields}{

\describe{
\item{\code{value}}{}
}}
\note{
Abstract ISO codelist class used internally by geometa
}
\section{Methods}{

\describe{
 \item{\code{new(xml, id, value, addCodeListAttrs, addCodeSpaceAttr, setValue)}}{
   This method is used to instantiate an ISOMetadataCodelistElement. By default,
   \code{addCodeListAttrs = TRUE}, to add codelist atributes to root XML. The
   parameter \code{addCodeSpaceAttr = TRUE} by default, and ignored if the valueof
   \code{addCodeLisAttrs} is set to \code{FALSE}. The argument \code{setValue}
   sets the value as node text (defaut is \code{TRUE}).

 }
 \item{\code{getAcceptedValues()}}{
   This method allows to get the codelist accepted values
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\references{
ISO 19115:2003 - Geographic information -- Metadata
}
\keyword{ISO}
\keyword{code}
\keyword{element}

