% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOMetadataNamespace.R
\docType{class}
\name{ISOMetadataNamespace}
\alias{ISOMetadataNamespace}
\title{ISOMetadataNamespace}
\format{\code{\link{R6Class}} object.}
\usage{
ISOMetadataNamespace
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO Metadata Namespace
}
\description{
ISOMetadataNamespace
}
\note{
ISO class used internally by geometa for specifying XML namespaces
}
\section{Methods}{

\describe{
 \item{\code{new(id, uri)}}{
   This method is used to instantiate an ISOMetadata
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{metadata}
\keyword{namespace}

