% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOMimeFileType.R
\docType{class}
\name{ISOMimeFileType}
\alias{ISOMimeFileType}
\title{ISOMimeFileType}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO MimeFileType
}
\description{
ISOMimeFileType
}
\section{Methods}{

\describe{
 \item{\code{new(xml, type, name)}}{
   This method is used to instantiate an \code{\link{ISOMimeFileType}}
 }
}
}

\examples{
  md <- ISOMimeFileType$new(type = "somemimetype", name = "Mime type name")
  xml <- md$encode()

}
\references{
ISO/TS 19139:2007 Geographic information -- XML
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{file}
\keyword{mime}
\keyword{type}
