% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOObligation.R
\docType{class}
\name{ISOObligation}
\alias{ISOObligation}
\title{ISOObligation}
\format{\code{\link{R6Class}} object.}
\usage{
ISOObligation
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO Obligation
}
\description{
ISOObligation
}
\section{Fields}{

\describe{
\item{\code{value}}{}
}}

\section{Methods}{

\describe{
 \item{\code{new(xml,value, description)}}{
   This method is used to instantiate an ISOObligation
 }
}
}

\examples{
  #possible values
  values <- ISOObligation$values(labels = TRUE)
  
  #mandatory value
  mandatory <- ISOObligation$new(value = "mandatory")

}
\references{
ISO 19115:2003 - Geographic information -- Metadata
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{Obligation}
