% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOOnlineResource.R
\docType{class}
\name{ISOOnlineResource}
\alias{ISOOnlineResource}
\title{ISOOnlineResource}
\format{\code{\link{R6Class}} object.}
\usage{
ISOOnlineResource
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO Online Resource
}
\description{
ISOOnlineResource
}
\section{Fields}{

\describe{
\item{\code{linkage}}{}

\item{\code{protocol}}{}

\item{\code{name}}{}

\item{\code{description}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(xml, linkage, name)}}{
   This method is used to instantiate an ISOOnlineResource
 }
 \item{\code{setLinkage(linkage)}}{
   Sets the linkage (URL)
 }
 \item{\code{setProtocol(protocol)}}{
   Sets the protocol
 }
 \item{\code{setName(name)}}{
   Sets the name
 }
 \item{\code{setDescription(description)}}{
   Sets the description
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{online}
\keyword{resource}

