% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOPresentationForm.R
\docType{class}
\name{ISOPresentationForm}
\alias{ISOPresentationForm}
\title{ISOPresentationForm}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO PresentationForm
}
\description{
ISOPresentationForm
}
\section{Methods}{

\describe{
 \item{\code{new(xml,value, description)}}{
   This method is used to instantiate an \code{\link{ISOPresentationForm}}
 }
}
}

\examples{
  #possible values
  values <- ISOPresentationForm$values(labels = TRUE)
  
  #mapDigital type
  map <- ISOPresentationForm$new(value = "mapDigital")
  
}
\references{
ISO 19115:2003 - Geographic information -- Metadata
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{form}
\keyword{presentation}
