% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISORecordType.R
\docType{class}
\name{ISORecordType}
\alias{ISORecordType}
\title{ISORecordType}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an ISORecordType
}
\description{
ISORecordType
}
\section{Fields}{

\describe{
\item{\code{value}}{[\code{\link{character}}] record type}
}}

\section{Methods}{

\describe{
 \item{\code{new(xml,value)}}{
   This method is used to instantiate an ISORecordType
 }
}
}

\references{
ISO/TS 19103:2005 Geographic information -- Conceptual schema language
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{record}
