% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOReferenceIdentifier.R
\docType{class}
\name{ISOReferenceIdentifier}
\alias{ISOReferenceIdentifier}
\title{ISOReferenceIdentifier}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO ReferenceIdentifier
}
\description{
ISOReferenceIdentifier
}
\section{Fields}{

\describe{
\item{\code{codeSpace}}{[\code{\link{character}}] codespace}

\item{\code{version}}{[\code{\link{character}}] version}
}}

\section{Methods}{

\describe{
 \item{\code{new(xml, code, codeSpace)}}{
   This method is used to instantiate an \code{\link{ISOReferenceIdentifier}}
 }
 \item{\code{setCodeSpace(codeSpace)}}{
   Sets a codeSpace
 }
 \item{\code{setVersion(version)}}{
   Sets a version
 }
}
}

\examples{
  md <- ISOReferenceIdentifier$new(code = "4326", codeSpace = "EPSG")
  xml <- md$encode()
  
}
\references{
ISO 19115:2003 - Geographic information -- Metadata
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{identifier}
\keyword{reference}
