% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOReferenceSystem.R
\docType{class}
\name{ISOReferenceSystem}
\alias{ISOReferenceSystem}
\title{ISOReferenceSystem}
\format{\code{\link{R6Class}} object.}
\usage{
ISOReferenceSystem
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO ReferenceSystem
}
\description{
ISOReferenceSystem
}
\section{Fields}{

\describe{
\item{\code{referenceSystemIdentifier}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(xml, value)}}{
   This method is used to instantiate an ISOReferenceSystem
 }
 \item{\code{setReferenceSystemIdentifier(code, codeSpace)}}{
   Sets the reference system identifier
 }
}
}
\examples{
md <- ISOReferenceSystem$new()
  rsId <- ISOReferenceIdentifier$new(code = "4326", codeSpace = "EPSG")
  md$setReferenceSystemIdentifier(rsId)
  xml <- md$encode()
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\references{
ISO 19115:2003 - Geographic information -- Metadata
}
\keyword{ISO}
\keyword{reference}
\keyword{system}

