% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOResponsibleParty.R
\docType{class}
\name{ISOResponsibleParty}
\alias{ISOResponsibleParty}
\title{ISOResponsibleParty}
\format{\code{\link{R6Class}} object.}
\usage{
ISOResponsibleParty
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO ResponsibleParty
}
\description{
ISOResponsibleParty
}
\section{Fields}{

\describe{
\item{\code{value}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(xml,value)}}{
   This method is used to instantiate an ISOResponsibleParty
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{file}
\keyword{identifier}

