% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISORestriction.R
\docType{class}
\name{ISORestriction}
\alias{ISORestriction}
\title{ISOHierarchyLevel}
\format{\code{\link{R6Class}} object.}
\usage{
ISORestriction
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO Restriction
}
\description{
ISOHierarchyLevel
}
\section{Fields}{

\describe{
\item{\code{value}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(xml,value)}}{
   This method is used to instantiate an ISORestriction
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{Restriction}

