% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOScale.R
\docType{class}
\name{ISOScale}
\alias{ISOScale}
\title{ISOScale}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an ISOScale measure
}
\description{
ISOScale
}
\section{Fields}{

\describe{
\item{\code{value}}{}
}}

\section{Methods}{

\describe{
 \item{\code{new(xml,value, uom, useUomURI)}}{
   This method is used to instantiate an \code{\link{ISOScale}}. The \code{uom} argument represents
   the symbol of unit of measure used. The parameter  \code{useUomURI} can be used to 
   set the uom as URI, its default value is \code{FALSE}.
 }
}
}

\references{
ISO/TS 19103:2005 Geographic information -- Conceptual schema language
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{measure}
\keyword{scale}
