% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOScope.R
\docType{class}
\name{ISOScope}
\alias{ISOScope}
\title{ISOScope}
\format{\code{\link{R6Class}} object.}
\usage{
ISOScope
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO Scope
}
\description{
ISOScope
}
\section{Fields}{

\describe{
\item{\code{level}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to instantiate an ISOScope
 }
 \item{\code{setLevel(level)}}{
   Sets the scope level
 }
}
}
\examples{
md <- ISOScope$new()
  md$setLevel("dataset")
  xml <- md$encode()
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{extent}

