% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOScopedName.R
\docType{class}
\name{ISOScopedName}
\alias{ISOScopedName}
\title{ISOScopedName}
\format{\code{\link{R6Class}} object.}
\usage{
ISOScopedName
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO ScopedName
}
\description{
ISOScopedName
}
\section{Fields}{

\describe{
\item{\code{value}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(xml,value)}}{
   This method is used to instantiate an ISOScopedName
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\references{
ISO/TS 19103:2005 Geographic information -- Conceptual schema language
}
\keyword{ISO}
\keyword{name}
\keyword{scoped}

