% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ISOServiceIdentification.R
\docType{class}
\name{ISOServiceIdentification}
\alias{ISOServiceIdentification}
\title{ISOServiceIdentification}
\format{\code{\link{R6Class}} object.}
\usage{
ISOServiceIdentification
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO ServiceIdentification
}
\description{
ISOServiceIdentification
}
\section{Fields}{

\describe{
\item{\code{serviceType}}{}

\item{\code{serviceTypeVersion}}{}

\item{\code{accessProperties}}{}

\item{\code{restrictions}}{}

\item{\code{keywords}}{}

\item{\code{extent}}{}

\item{\code{coupledResource}}{}

\item{\code{couplingType}}{}

\item{\code{containsOperations}}{}

\item{\code{operatesOn}}{}
}}
\section{Inherited methods}{

\describe{
 \item{\code{setCitation(citation)}}{
   Sets an object of class \code{ISOCitation}
 }
 \item{\code{setAbstract(abstract)}}{
   Sets an abstract (object of class "character")
 }
 \item{\code{setPurpose(purpose)}}{
   Sets a purpose (object of class "character")
 }
 \item{\code{addCredit(credit)}}{
   Adds a credit (object of class "character")
 }
 \item{\code{delCredit(credit)}}{
   Deletes a credit (object of class "character")
 }
 \item{\code{addStatus(status)}}{
   Adds a status, as object of class "character" or class \code{ISOStatus}. If
   an object of class "character" is specified, it must match the accepted
   progress status values \code{ISOStatus$values()}.
 }
 \item{\code{delStatus(status)}}{
   Deletes a status, as object of class "character" or class \code{ISOStatus}. If
   an object of class "character" is specified, it must match the accepted
   progress status values \code{ISOStatus$values()}.
 }
 \item{\code{addPointOfContact(pointOfContact)}}{
   Adds an object of class \code{ISOResponsibleParty}
 }
 \item{\code{delPointOfContact(pointOfContact)}}{
   Deletes an object of class \code{ISOResponsibleParty}
 }
 \item{\code{addResourceMaintenance(resourceMaintenance)}}{
   Adds a resource maintenance information as object of class
   \code{ISOMaintenanceInformation}.
 }
 \item{\code{setResourceMaintenance(resourceMaintenance)}}{
   Sets a resource maintenance information as object of class
   \code{ISOMaintenanceInformation}.
 }
 \item{\code{delResourceMaintenance(resourceMaintenance)}}{
   Deletes a resource maintenance information as object of class
   \code{ISOMaintenanceInformation}.
 }
 \item{\code{addGraphicOverview(graphicOverview)}}{
   Adds an object of class \code{ISOBrowseGraphic}
 }
 \item{\code{setGraphicOverview(graphicOverview)}}{
   Sets an object of class \code{ISOBrowseGraphic}
 }
 \item{\code{delGraphicOverview(graphicOverview)}}{
   Deletes an object of class \code{ISOBrowseGraphic}
 }
 \item{\code{addKeywords(keywords)}}{
   Adds a set of keywords as object of class \code{ISOKeywords}
 }
 \item{\code{setKeywords(keywords)}}{
   Sets a set of keywords as object of class \code{ISOKeywords}
 }
 \item{\code{delKeywords(keywords)}}{
   Deletes a set of keywords as object of class \code{ISOKeywords}
 }
 \item{\code{addResourceConstraints(resourceConstraints)}}{
   Adds an object of class \code{ISOLegalConstraints}
 }
 \item{\code{setResourceConstraints(resourceConstraints)}}{
   Sets an object of class \code{ISOLegalConstraints}
 }
 \item{\code{addResourceConstraints(resourceConstraints)}}{
   Deletes an object of class \code{ISOLegalConstraints}
 }
}
}

\section{Methods}{

\describe{
 \item{\code{new(xml,value)}}{
   This method is used to instantiate an ISOServiceIdentification
 }
}
}
\examples{
#create ServiceIdentification
   md <- ISOServiceIdentification$new()
   md$setAbstract("abstract")
   md$setPurpose("purpose")

   #adding a point of contact
   rp <- ISOResponsibleParty$new()
   rp$setIndividualName("someone")
   rp$setOrganisationName("somewhere")
   rp$setPositionName("someposition")
   rp$setRole("pointOfContact")
   contact <- ISOContact$new()
   phone <- ISOTelephone$new()
   phone$setVoice("myphonenumber")
   phone$setFacsimile("myfacsimile")
   contact$setPhone(phone)
   address <- ISOAddress$new()
   address$setDeliveryPoint("theaddress")
   address$setCity("thecity")
   address$setPostalCode("111")
   address$setCountry("France")
   address$setEmail("someone@theorg.org")
   contact$setAddress(address)
   res <- ISOOnlineResource$new()
   res$setLinkage("http://www.somewhereovertheweb.org")
   res$setName("somename")
   contact$setOnlineResource(res)
   rp$setContactInfo(contact)
   md$addPointOfContact(rp)

   #citation
   ct <- ISOCitation$new()
   ct$setTitle("sometitle")
   d <- ISODate$new()
   d$setDate(ISOdate(2015, 1, 1, 1))
   d$setDateType("publication")
   ct$addDate(d)
   ct$setEdition("1.0")
   ct$setEditionDate(ISOdate(2015, 1, 1, 1))
   ct$setIdentifier(ISOMetaIdentifier$new(code = "identifier"))
   ct$setPresentationForm("mapDigital")
   ct$setCitedResponsibleParty(rp)
   md$setCitation(ct)

   #graphic overview
   go <- ISOBrowseGraphic$new(
     fileName = "http://wwww.somefile.org/png",
     fileDescription = "Service Overview",
     fileType = "image/png"
   )
   md$setGraphicOverview(go)

   #maintenance information
   mi <- ISOMaintenanceInformation$new()
   mi$setMaintenanceFrequency("daily")
   md$setResourceMaintenance(mi)

   #adding legal constraints
   lc <- ISOLegalConstraints$new()
   lc$addUseLimitation("limitation1")
   lc$addUseLimitation("limitation2")
   lc$addUseLimitation("limitation3")
   lc$addAccessConstraint("copyright")
   lc$addAccessConstraint("license")
   lc$addUseConstraint("copyright")
   lc$addUseConstraint("license")
   md$setResourceConstraints(lc)

   #specific serviceIdentification elements


   xml <- md$encode()
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\references{
ISO 19115:2003 - Geographic information -- Metadata
}
\keyword{ISO}
\keyword{identification}
\keyword{service}

