% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOSpatialRepresentation.R
\docType{class}
\name{ISOSpatialRepresentation}
\alias{ISOSpatialRepresentation}
\title{ISOSpatialRepresentation}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO abstract SpatialRepresentation
}
\description{
ISOSpatialRepresentation
}
\note{
abstract class
}
\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to instantiate an \code{\link{ISOSpatialRepresentation}}
 }
}
}

\references{
ISO 19115:2003 - Geographic information -- Metadata
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{address}
