% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOStatus.R
\docType{class}
\name{ISOStatus}
\alias{ISOStatus}
\title{ISOStatus}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO progress status
}
\description{
ISOStatus
}
\section{Methods}{

\describe{
 \item{\code{new(xml,value, description)}}{
   This method is used to instantiate an \code{\link{ISOStatus}}
 }
}
}

\examples{
  #possible values
  values <- ISOStatus$values(labels = TRUE)
  
  #pending status
  pending <- ISOStatus$new(value = "pending")
  
}
\references{
ISO 19115:2003 - Geographic information -- Metadata
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{status}
