% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOTopologyLevel.R
\docType{class}
\name{ISOTopologyLevel}
\alias{ISOTopologyLevel}
\title{ISOTopologyLevel}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an ISO TopologyLevel
}
\description{
ISOTopologyLevel
}
\section{Methods}{

\describe{
 \item{\code{new(xml,value, description)}}{
   This method is used to instantiate an \code{\link{ISOTopologyLevel}}
 }
}
}

\examples{
  #possible values
  values <- ISOTopologyLevel$values(labels = TRUE)
  
  #geomOnly
  geomOnly <- ISOTopologyLevel$new(value = "geometryOnly")
  
}
\references{
ISO 19115:2003 - Geographic information -- Metadata
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{charset}
\keyword{level}
\keyword{topology}
