% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOTypeName.R
\docType{class}
\name{ISOTypeName}
\alias{ISOTypeName}
\title{ISOTypeName}
\format{\code{\link{R6Class}} object.}
\usage{
ISOTypeName
}
\value{
Object of \code{\link{R6Class}} for modelling an ISOTypeName
}
\description{
ISOTypeName
}
\section{Fields}{

\describe{
\item{\code{aName}}{}
}}

\section{Methods}{

\describe{
 \item{\code{new(xml, aName)}}{
   This method is used to instantiate an ISOTypeName
 }
 \item{\code{setName(aName)}}{
   Sets the aName
 }
}
}

\examples{
  typeName <- ISOTypeName$new(aName = "name")
  xml <- typeName$encode()

}
\references{
ISO/TS 19103:2005 Geographic information -- Conceptual schema language
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{typename}
