% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOURL.R
\docType{class}
\name{ISOURL}
\alias{ISOURL}
\title{ISOURL}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an ISOURL
}
\description{
ISOURL
}
\section{Fields}{

\describe{
\item{\code{value}}{[\code{\link{character}}] the url}
}}

\note{
Class used by geometa internal XML decoder/encoder
}
\section{Methods}{

\describe{
 \item{\code{new(xml,value)}}{
   This method is used to instantiate an \code{\link{ISOURL}}
 }
 \item{\code{setUrl(url)}}{
   Set the url
 }
}
}

\references{
ISO 19115:2003 - Geographic information -- Metadata
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{URL}
