% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMatrixGZ.R
\name{readMatrixGZ}
\alias{readMatrixGZ}
\title{Read expression matrix in compressed format}
\usage{
readMatrixGZ(GSE)
}
\arguments{
\item{GSE}{a character of GSE number}
}
\value{
a GSE expression dataframe
}
\description{
nothing
}
\details{
nothing
}
