% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMatrixGZ.R
\name{readMatrixGZ}
\alias{readMatrixGZ}
\title{Regional Plot Association Results of Pig}
\usage{
readMatrixGZ(GSE)
}
\arguments{
\item{GSE}{a character of GSE number}

\item{GSE_GPL}{a character vactor comparised of GSE and GPL number.Note if GSE was sequenced on only one platform, please just give GSE number.}
}
\value{
download GSE expression matrix and GPL annotation,
        and return a matrix with colanmes of "GSE" and "GPL" and row names of GSE_GPL

a GSE expression dataframe
}
\description{
This is a plotting tool to simultaneously display the association p-value,
             the LD, recombination rate and annotated gene in region of interesting in pig.
             The interpretation and visualization tool of association results will
             contribute to further identification of causal mutations.
}
\details{
nothing
}
