% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOMetadataSchemas.R
\name{registerISOMetadataSchema}
\alias{registerISOMetadataSchema}
\title{registerISOMetadataSchema}
\usage{
registerISOMetadataSchema(version, xsdFile)
}
\arguments{
\item{version}{the schema version}

\item{xsdFile}{the schema XSD file}
}
\description{
\code{registerISOMetadataSchema} allows to register a new schema
in \pkg{geometa}
}
\examples{
\donttest{             
  registerISOMetadataSchema(version = "19139", xsdFile = "http://www.isotc211.org/2005/gmd/gmd.xsd")
}

}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
