% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOMetadataSchemas.R
\name{setMetadataStandard}
\alias{setMetadataStandard}
\title{setMetadataStandard}
\usage{
setMetadataStandard(version)
}
\arguments{
\item{version}{the standard version}
}
\description{
\code{setMetadataStandard} allows to set the standard to use for encoding/decoding in \pkg{geometa}.
 By default the standard "19139" will be used. Possible alternative value "19115-3"
}
\examples{
\donttest{          
  setMetadataStandard(version = "19139")
}
}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
