% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWEAbstractDataComponent.R
\docType{class}
\name{SWEAbstractDataComponent}
\alias{SWEAbstractDataComponent}
\title{SWEAbstractDataComponent}
\format{
\code{\link{R6Class}} object.
}
\arguments{
\item{xml}{object of class \link{XMLInternalNode-class} from \pkg{XML}}

\item{element}{element}

\item{attrs}{attrs}

\item{defaults}{defaults}

\item{wrap}{wrap}

\item{definition}{definition}

\item{description}{description}

\item{name}{name}

\item{codeSpace}{codespace}

\item{identifier}{identifier}
}
\value{
Object of \code{\link{R6Class}} for modelling an SWE Abstract data component
}
\description{
Initializes an object of class \link{SWEAbstractDataComponent}

Set definition

Set description

Adds name

Deletes name

Set identifier
}
\section{Fields}{

\describe{
\item{\code{definition}}{definition}

\item{\code{description}}{description}

\item{\code{label}}{label}

\item{\code{name}}{name}

\item{\code{identifier}}{identifier}
}}

\note{
Class used internally by geometa
}
\references{
OGC Geography Markup Language. https://www.ogc.org/standards/swecommon
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{SWE}
