% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWEAbstractSimpleComponent.R
\docType{class}
\name{SWEAbstractSimpleComponent}
\alias{SWEAbstractSimpleComponent}
\title{SWEAbstractSimpleComponent}
\format{
\code{\link{R6Class}} object.
}
\arguments{
\item{xml}{object of class \link{XMLInternalNode-class} from \pkg{XML}}

\item{element}{element}

\item{updatable}{updatable}

\item{optional}{optional}

\item{definition}{definition}

\item{nilValue}{value to set as nil Value. object of class \code{numeric}}
}
\value{
Object of \code{\link{R6Class}} for modelling an SWE Abstract simple component
}
\description{
Initializes an object of class \link{SWEAbstractSimpleComponent}

Set nil value and its reason (optional)
}
\section{Fields}{

\describe{
\item{\code{nilValues}}{nil values}
}}

\references{
SWE Common Data Model Encoding Standard. https://www.ogc.org/standards/swecommon
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{SWE}
