% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWENilValues.R
\docType{class}
\name{SWENilValues}
\alias{SWENilValues}
\title{SWENilValues}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an SWE nil values object
}
\description{
SWENilValues
}
\section{Fields}{

\describe{
\item{\code{nilValue}}{nil value}
}}

\references{
SWE Common Data Model Encoding Standard. https://www.ogc.org/standards/swecommon
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{SWE}
